JAMAlabels {meta}	R Documentation

Create study labels in JAMA layout (for forest plot)

Description

Create study labels in JAMA layout (for forest plot).

Usage

JAMAlabels(author, year, citation, data = NULL)

Arguments

author
A vector providing study authors.

year
A vector providing year of publication.

citation
A vector proving citation numbers.

data
An optional data frame containing the study information.

Details

This auxiliary function can be used to create study labels in JAMA layout which can be added to a forest plot using argument 'studlab'.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

See Also

forest.meta

Examples

data(Fleiss1993bin)

refs <- 20 + 1:7

Fleiss1993bin$mylabs <-
  JAMAlabels(study, year, refs, data = Fleiss1993bin)

m <- metabin(d.asp, n.asp, d.plac, n.plac, data = Fleiss1993bin,
             studlab = paste(study, year),
             sm = "OR", comb.random = FALSE)

forest(m, studlab = mylabs, layout = "JAMA",
       fontfamily = "Times", fontsize = 10)

[Package meta version 4.18-1 Index]