amlodipine {meta}	R Documentation

Amlodipine for Work Capacity

Description

Meta-analysis on the effect of amlodipine on work capacity.

This meta-analysis is used as a data example in Hartung and Knapp (2001).

Usage

data(amlodipine)

Format

A data frame with the following columns:

study

Study label

n.amlo

Number of observations in amlodipine group

mean.amlo

Estimated mean in amlodipine group

var.amlo

Variance in amlodipine group

n.plac

Number of observations in placebo group

mean.plac

Estimated mean in placebo group

var.plac

Variance in placebo group

Source

Hartung J & Knapp G (2001), On tests of the overall treatment effect in meta-analysis with normally distributed responses. Statistics in Medicine, 20, 1771–82. doi: 10.1002/sim.791 .

See Also

metacont

Examples

data(amlodipine)
#
m <- metacont(n.amlo, mean.amlo, sqrt(var.amlo),
              n.plac, mean.plac, sqrt(var.plac),
              data=amlodipine, studlab=study)

m.hakn <- metacont(n.amlo, mean.amlo, sqrt(var.amlo),
                   n.plac, mean.plac, sqrt(var.plac),
                   data=amlodipine, studlab=study,
                   hakn=TRUE)

# Results for mean difference - see Table III in Hartung and Knapp (2001) 
#
res.md <- rbind(data.frame(summary(m)$fixed)[c("TE", "lower", "upper")],
                data.frame(summary(m)$random)[c("TE", "lower", "upper")],
                data.frame(summary(m.hakn)$random)[c("TE", "lower", "upper")])
#
res.md <- round(res.md, 5)
#
row.names(res.md) <- c("FE", "RE", "RE (HaKn)")
names(res.md) <- c("Absolute difference", "CI lower", "CI upper")
#
res.md

[Package meta version 4.9-4 Index]