baujat.meta {meta}	R Documentation

Baujat plot to explore heterogeneity in meta-analysis

Description

Draw a Baujat plot to explore heterogeneity in meta-analysis.

Usage

## S3 method for class 'meta'
baujat(
  x,
  yscale = 1,
  xlim,
  ylim,
  xlab = "Contribution to overall heterogeneity",
  ylab = "Influence on overall result",
  pch = 21,
  cex = 1,
  col = "black",
  bg = "darkgray",
  studlab = TRUE,
  cex.studlab = 0.8,
  pos.studlab = 2,
  offset = 0.5,
  xmin = 0,
  ymin = 0,
  grid = TRUE,
  col.grid = "lightgray",
  lty.grid = "dotted",
  lwd.grid = par("lwd"),
  pty = "s",
  ...
)

Arguments

x
An object of class meta.

yscale
Scaling factor for values on y-axis.

xlim
The x limits (min,max) of the plot.

ylim
The y limits (min,max) of the plot.

xlab
A label for the x-axis.

ylab
A label for the y-axis.

pch
The plotting symbol used for individual studies.

cex
The magnification to be used for plotting symbol.

col
A vector with colour of plotting symbols.

bg
A vector with background colour of plotting symbols (only used if pch in 21:25).

studlab
A logical indicating whether study labels should be printed in the graph. A vector with study labels can also be provided (must be of same length as x$TE then).

cex.studlab
The magnification for study labels.

pos.studlab
Position of study labels, see argument pos in text.

offset
Offset for study labels (see text).

xmin
A numeric specifying minimal value to print study labels (on x-axis).

ymin
A numeric specifying minimal value to print study labels (on y-axis).

grid
A logical indicating whether a grid is printed in the plot.

col.grid
Colour for grid lines.

lty.grid
The line type for grid lines.

lwd.grid
The line width for grid lines.

pty
A character specifying type of plot region (see par).

...
Graphical arguments as in par may also be passed as arguments.

Details

Baujat et al. (2002) introduced a scatter plot to explore heterogeneity in meta-analysis. On the x-axis the contribution of each study to the overall heterogeneity statistic (see list object Q of the meta-analysis object x) is plotted. On the y-axis the standardised difference of the overall treatment effect with and without each study is plotted; this quantity describes the influence of each study on the overal treatment effect.

Internally, the metainf function is used to calculate the values on the y-axis.

Value

A data.frame with the following variables:

x
Coordinate on x-axis (contribution to heterogeneity statistic)

y
Coordinate on y-axis (influence on overall treatment effect)

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Baujat B, Mahé C, Pignon JP, Hill C (2002): A graphical method for exploring heterogeneity in meta-analyses: Application to a meta-analysis of 65 trials. Statistics in Medicine, 30, 2641–52

See Also

metagen, metainf

Examples

data(Olkin1995)

m1 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
              data = Olkin1995, sm = "OR", method = "I",
              studlab = paste(author, year))

# Generate Baujat plot
baujat(m1)

## Not run: 
# Do not print study labels if the x-value is smaller than 4 and
# the y-value is smaller than 1
baujat(m1, yscale = 10, xmin = 4, ymin = 1)

# Change position of study labels
baujat(m1, yscale = 10, xmin = 4, ymin = 1,
       pos = 1, xlim = c(0, 6.5))

# Generate Baujat plot and assign x- and y- coordinates to R object
# b1
b1 <- baujat(m1)

# Calculate overall heterogeneity statistic
sum(b1$x)
m1$Q

## End(Not run)

[Package meta version 4.18-1 Index]