ci {meta}	R Documentation

Calculation of confidence intervals (based on normal approximation or t-distribution)

Description

Calculation of confidence intervals; based on normal approximation or t-distribution.

Usage

ci(TE, seTE, level = 0.95, df = NULL, null.effect = 0)

Arguments

TE
Estimated treatment effect.

seTE
Standard error of treatment estimate.

level
The confidence level required.

df
Degrees of freedom (for confidence intervals based on t-distribution).

null.effect
A numeric value specifying the effect under the null hypothesis.

Value

List with components

TE
Estimated treatment effect

seTE
Standard error of treatment estimate

lower
Lower confidence limits

upper
Upper confidence limits

statistic
Test statistic (either z-score or t-score)

p
P-value of text.with null hypothesis TE=0

level
The confidence level required

df
Degrees of freedom (t-distribution)

Note

This function is primarily called from other functions of the library meta, e.g. forest.meta, summary.meta.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

Examples

data.frame(ci(170, 10))
data.frame(ci(170, 10, 0.99))
data.frame(ci(1.959964, 1))
data.frame(ci(2.2621571628, 1, df = 9))

[Package meta version 4.18-1 Index]