forest.metabind {meta}	R Documentation

Forest plot to display the result of a meta-analysis

Description

Draws a forest plot in the active graphics window (using grid graphics system).

Usage

## S3 method for class 'metabind'
forest(
  x,
  leftcols,
  leftlabs,
  rightcols = c("effect", "ci"),
  rightlabs,
  overall = FALSE,
  subgroup = FALSE,
  hetstat = if (any(x$is.subgroup)) FALSE else "study",
  overall.hetstat = FALSE,
  lab.NA = "",
  digits = gs("digits.forest"),
  digits.se = gs("digits.se"),
  digits.stat = gs("digits.stat"),
  digits.pval = max(gs("digits.pval") - 2, 2),
  digits.pval.Q = max(gs("digits.pval.Q") - 2, 2),
  digits.Q = gs("digits.Q"),
  digits.tau2 = gs("digits.tau2"),
  digits.tau = gs("digits.tau"),
  digits.I2 = max(gs("digits.I2") - 1, 0),
  scientific.pval = gs("scientific.pval"),
  big.mark = gs("big.mark"),
  smlab,
  calcwidth.pooled = overall,
  ...
)

Arguments

x
An object of class metabind.

leftcols
A character vector specifying (additional) columns to be plotted on the left side of the forest plot or a logical value (see Details).

leftlabs
A character vector specifying labels for (additional) columns on left side of the forest plot (see Details).

rightcols
A character vector specifying (additional) columns to be plotted on the right side of the forest plot or a logical value (see Details).

rightlabs
A character vector specifying labels for (additional) columns on right side of the forest plot (see Details).

overall
A logical indicating whether overall summaries should be plotted. This argument is useful in a meta-analysis with subgroups if summaries should only be plotted on group level.

subgroup
A logical indicating whether subgroup results should be shown in forest plot. This argument is useful in a meta-analysis with subgroups if summaries should not be plotted on group level.

hetstat
Either a logical value indicating whether to print results for heterogeneity measures at all or a character string (see Details).

overall.hetstat
A logical value indicating whether to print heterogeneity measures for overall treatment comparisons. This argument is useful in a meta-analysis with subgroups if heterogeneity statistics should only be printed on subgroup level.

lab.NA
A character string to label missing values.

digits
Minimal number of significant digits for treatment effects, see print.default.

digits.se
Minimal number of significant digits for standard errors, see print.default.

digits.stat
Minimal number of significant digits for z- or t-statistic for test of overall effect, see print.default.

digits.pval
Minimal number of significant digits for p-value of overall treatment effect, see print.default.

digits.pval.Q
Minimal number of significant digits for p-value of heterogeneity test, see print.default.

digits.Q
Minimal number of significant digits for heterogeneity statistic Q, see print.default.

digits.tau2
Minimal number of significant digits for between-study variance, see print.default.

digits.tau
Minimal number of significant digits for square root of between-study variance, see print.default.

digits.I2
Minimal number of significant digits for I-squared statistic, see print.default.

scientific.pval
A logical specifying whether p-values should be printed in scientific notation, e.g., 1.2345e-01 instead of 0.12345.

big.mark
A character used as thousands separator.

smlab
A label for the summary measurex (printed at top of figure).

calcwidth.pooled
A logical indicating whether text for fixed effect and random effects model should be considered to calculate width of the column with study labels.

...
Additional graphical arguments (passed on to forest.meta).

Details

A forest plot, also called confidence interval plot, is drawn in the active graphics window. The forest functions in R package meta are based on the grid graphics system. In order to print the forest plot, resize the graphics window and either use dev.copy2eps or dev.copy2pdf. Another possibility is to create a file using pdf, png, or svg and to specify the width and height of the graphic (see forest.meta examples).

The arguments leftcols and rightcols can be used to specify columns which are plotted on the left and right side of the forest plot, respectively.

The arguments leftlabs and rightlabs can be used to specify column headings which are plotted on left and right side of the forest plot, respectively. For certain columns predefined labels exist. For other columns, the column name will be used as a label. It is possible to only provide labels for new columns (see forest.meta examples). Otherwise the length of leftlabs and rightlabs must be the same as the number of printed columns, respectively. The value NA can be used to specify columns which should use default labels.

Argument hetstat can be a character string to specify where to print heterogeneity information:

row with results for fixed effect model (hetstat = "fixed"),

row with results for random effects model (hetstat = "random"),

rows with 'study' information (hetstat = "study").

Otherwise, information on heterogeneity is printed in dedicated rows.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

See Also

forest.meta, metabin, metacont, metagen, metabind, settings.meta

Examples

data(Fleiss1993cont)

# Add some (fictitious) grouping variables:
#
Fleiss1993cont$age <- c(55, 65, 55, 65, 55)
Fleiss1993cont$region <- c("Europe", "Europe", "Asia", "Asia", "Europe")

m1 <- metacont(n.psyc, mean.psyc, sd.psyc, n.cont, mean.cont, sd.cont,
               data = Fleiss1993cont, sm = "MD")

# Conduct two subgroup analyses
#
mu1 <- update(m1, byvar = age, bylab = "Age group")
mu2 <- update(m1, byvar = region, bylab = "Region")

# Combine subgroup meta-analyses and show forest plot with subgroup
# results
#
mb1 <- metabind(mu1, mu2)
mb1
forest(mb1)

[Package meta version 4.18-1 Index]