gs {meta}	R Documentation

Get default for a meta-analysis setting.

Description

Get default for a meta-analysis setting in R package meta.

Usage

gs(x)

Arguments

x
A character string holding a settings name.

Details

This function can be used to get the default for a meta-analysis setting defined using settings.meta.

This function is primarily used to define default settings in meta-analysis functions, e.g., metabin or metacont. A list of all arguments with current settings is printed using the command settings.meta("print").

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

See Also

settings.meta

Examples

# Get default setting for Hartung-Knapp method
#
gs("hakn")

# Get default setting for summary measure in metabin()
#
gs("smbin")

[Package meta version 4.18-1 Index]