labbe.metabin {meta}	R Documentation

L'Abbé plot for meta-analysis with binary outcomes

Description

Draw a L'Abbé plot for meta-analysis with binary outcomes.

Usage

## S3 method for class 'metabin'
labbe(
  x,
  xlim,
  ylim,
  xlab = NULL,
  ylab = NULL,
  TE.fixed = x$TE.fixed,
  TE.random = x$TE.random,
  comb.fixed = x$comb.fixed,
  comb.random = x$comb.random,
  backtransf = x$backtransf,
  axes = TRUE,
  pch = 21,
  text = NULL,
  cex = 1,
  col = "black",
  bg = "lightgray",
  lwd = 1,
  lwd.fixed = lwd,
  lwd.random = lwd,
  lty.fixed = 2,
  lty.random = 9,
  col.fixed = col,
  col.random = col,
  nulleffect = TRUE,
  lwd.nulleffect = lwd,
  col.nulleffect = "lightgray",
  sm = x$sm,
  weight,
  studlab = FALSE,
  cex.studlab = 0.8,
  pos.studlab = 2,
  label.e = x$label.e,
  label.c = x$label.c,
  ...
)

## Default S3 method:
labbe(
  x,
  y,
  xlim,
  ylim,
  xlab = NULL,
  ylab = NULL,
  TE.fixed = NULL,
  TE.random = NULL,
  comb.fixed = !is.null(TE.fixed),
  comb.random = !is.null(TE.random),
  backtransf = TRUE,
  axes = TRUE,
  pch = 21,
  text = NULL,
  cex = 1,
  col = "black",
  bg = "lightgray",
  lwd = 1,
  lwd.fixed = lwd,
  lwd.random = lwd,
  lty.fixed = 2,
  lty.random = 9,
  col.fixed = col,
  col.random = col,
  nulleffect = TRUE,
  lwd.nulleffect = lwd,
  col.nulleffect = "lightgray",
  sm = "",
  weight,
  studlab = FALSE,
  cex.studlab = 0.8,
  pos.studlab = 2,
  label.e = NULL,
  label.c = NULL,
  ...
)

Arguments

x
An object of class metabin. Alternatively, the x coordinates of points of the L'Abbé plot.

xlim
The x limits (min, max) of the plot.

ylim
The y limits (min, max) of the plot.

xlab
A label for the x-axis.

ylab
A label for the y-axis.

TE.fixed
A numeric or vector specifying combined fixed effect estimate(s).

TE.random
A numeric or vector specifying combined random effects estimate(s).

comb.fixed
A logical indicating whether the pooled fixed effect estimate should be plotted.

comb.random
A logical indicating whether the pooled random effects estimate should be plotted.

backtransf
A logical indicating which values should be printed on x- and y-axis (see Details).

axes
A logical indicating whether axes should be drawn on the plot.

pch
The plotting symbol used for individual studies.

text
A character vector specifying the text to be used instead of plotting symbol.

cex
The magnification to be used for plotting symbol.

col
A vector with colour of plotting symbols.

bg
A vector with background colour of plotting symbols (only used if pch in 21:25).

lwd
The line width.

lwd.fixed
The line width(s) for fixed effect estimate(s) (if comb.fixed is not NULL or FALSE).

lwd.random
The line width(s) for random effects estimate(s) (if comb.random is not NULL or FALSE).

lty.fixed
Line type(s) for fixed effect estimate(s).

lty.random
Line type(s) for random effects estimate(s).

col.fixed
Colour of line(s) for fixed effect estimate(s).

col.random
Colour of line(s) for random effects estimate(s).

nulleffect
A logical indicating whether line for null effect should be added to the plot..

lwd.nulleffect
Width of line for null effect.

col.nulleffect
Colour of line for null effect.

sm
A character string indicating underlying summary measure, i.e., "RD", "RR", "OR", or "ASD".

weight
Either a numeric vector specifying relative sizes of plotting symbols or a character string indicating which type of plotting symbols is to be used for individual treatment estimates. One of missing (see Details), "same", "fixed", or "random", can be abbreviated. Plot symbols have the same size for all studies or represent study weights from fixed effect or random effects model.

studlab
A logical indicating whether study labels should be printed in the graph. A vector with study labels can also be provided (must be of same length as x$event.e then).

cex.studlab
Size of study labels.

pos.studlab
Position of study labels, see argument pos in text.

label.e
Label for experimental group.

label.c
Label for control group.

...
Graphical arguments as in par may also be passed as arguments.

y
The y coordinates of the L'Abbé plot, if argument x is not an object of class metabin.

Details

A L'Abbé plot is a scatter plot with the risk in the control group on the x-axis and the risk in the experimental group on the y-axis (L'Abbé et al., 1987). It can be used to evaluate heterogeneity in meta-analysis. Furthermore, this plot can aid to choose a summary measure (odds ratio, risk ratio, risk difference) that will result in more consistent results (Jiménez et al., 1997; Deeks, 2002).

If argument backtransf is TRUE (default), event probabilities will be printed on x- and y-axis. Otherwise, transformed event probabilities will be printed as defined by the summary measure, i.e., log odds of probabilities for odds ratio as summary measure (sm = "OR"), log probabilities for sm = "RR", and arcsine-transformed probabilities for sm = "ASD".

If comb.fixed is TRUE, the pooled estimate of the fixed effect model is plotted as a line. If comb.random is TRUE, the pooled estimate of the random effects model is plotted as a line.

Information from object x is utilised if argument weight is missing. Weights from the fixed effect model are used (weight = "fixed") if argument x$comb.fixed is TRUE; weights from the random effects model are used (weight = "random") if argument x$comb.random is TRUE and x$comb.fixed is FALSE.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Deeks JJ (2002): Issues in the selection of a summary statistic for meta-analysis of clinical trials with binary outcomes. Statistics in Medicine, 21, 1575–600

Jiménez FJ, Guallar E, Martín-Moreno JM (1997): A graphical display useful for meta-analysis. European Journal of Public Health, 1, 101–5

L'Abbé KA, Detsky AS, O'Rourke K (1987): Meta-analysis in clinical research. Annals of Internal Medicine, 107, 224–33

See Also

metabin

Examples

data(Olkin1995)
m1 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
              data = Olkin1995,
              studlab = paste(author, year),
              sm = "RR", method = "I")

# L'Abbe plot for risk ratio
#
labbe(m1)

# L'Abbe plot for odds ratio
#
labbe(m1, sm = "OR")
# same plot
labbe(update(m1, sm = "OR"))

# L'Abbe plot for risk difference
#
labbe(m1, sm = "RD")

# L'Abbe plot on log odds scale
#
labbe(m1, sm = "OR", backtransf = FALSE)

# L'Abbe plot for odds ratio with coloured lines for various
# treatment effects (defined as log odds ratios)
#
mycols <- c("blue", "yellow", "green", "red",
            "green", "yellow", "blue")
labbe(m1, sm = "OR",
      comb.random = FALSE,
      TE.fixed = log(c(1 / 10, 1 / 5, 1 / 2, 1, 2, 5, 10)),
      col.fixed = mycols, lwd.fixed = 2)

# L'Abbe plot on log odds scale with coloured lines for various
# treatment effects (defined as log odds ratios)
#
labbe(m1, sm = "OR",
      comb.random = FALSE,
      TE.fixed = log(c(1 / 10, 1 / 5, 1 / 2, 1, 2, 5, 10)),
      col.fixed = mycols, lwd.fixed = 2,
      backtransf = FALSE)

[Package meta version 4.18-1 Index]