labbe.metabin {meta}	R Documentation

L'Abbe plot

Description

Draw a L'Abbe plot.

Usage

## S3 method for class 'metabin':
labbe(x, y,
                        xlim, ylim,
                        xlab=NULL, ylab=NULL,
                        TE.fixed=x$TE.fixed,
                        TE.random=x$TE.random,
                        comb.fixed=x$comb.fixed,
                        comb.random=x$comb.random,
                        axes=TRUE,
                        pch=21, text=NULL, cex=1,
                        col="black", bg="lightgray",
                        lwd=1, lwd.fixed=lwd, lwd.random=lwd,
                        lty.fixed=2, lty.random=9,
                        sm=x$sm, weight,
                        studlab=FALSE, cex.studlab=0.8,
                        ...)

## Default S3 method:
labbe(x, y,
                        xlim, ylim,
                        xlab=NULL, ylab=NULL,
                        TE.fixed, TE.random,
                        comb.fixed=FALSE, comb.random=FALSE,
                        axes=TRUE,
                        pch=21, text=NULL, cex=1,
                        col="black", bg="lightgray",
                        lwd=1, lwd.fixed=lwd, lwd.random=lwd,
                        lty.fixed=2, lty.random=9,
                        sm=NULL, weight,
                        studlab=FALSE, cex.studlab=0.8,
                        ...)

Arguments

x 	 The x coordinates of points of the L'Abbe plot. Alternatively, an object of class metabin.

y 	 The y coordinates of the L'Abbe plot, optional if x is an appropriate structure.

xlim 	 The x limits (min,max) of the plot.

ylim 	 The y limits (min,max) of the plot.

xlab 	 A label for the x axis.

ylab 	 A label for the y axis.

TE.fixed 	 A numeric or vector specifying combined fixed effect estimate(s).

TE.random 	 A numeric or vector specifying combined random effects estimate(s).

comb.fixed 	 A logical indicating whether the pooled fixed effect estimate should be plotted.

comb.random 	 A logical indicating whether the pooled random effects estimate should be plotted.

axes 	 A logical indicating whether axes should be drawn on the plot.

pch 	 The plotting symbol used for individual studies.

text 	 A character vector specifying the text to be used instead of plotting symbol.

cex 	 The magnification to be used for plotting symbol.

col 	 A vector with colour of plotting symbols.

bg 	 A vector with background colour of plotting symbols (only used if pch in 21:25).

lwd 	 The line width.

lwd.fixed 	 The line width for fixed effect estimate (if comb.fixed is not NULL or FALSE).

lwd.random 	 The line width for random effects estimate (if comb.random is not NULL or FALSE).

lty.fixed 	 Line type (pooled fixed effect estimate).

lty.random 	 Line type (pooled random effects estimate).

sm 	 A character string indicating underlying summary measure, i.e., "RD", "RR", "OR".

weight 	 Either a numeric vector specifying relative sizes of plotting symbols or a character string indicating which type of plotting symbols is to be used for individual treatment estimates. One of missing (see Details), "same", "fixed", or "random", can be abbreviated. Plot symbols have the same size for all studies or represent study weights from fixed effect or random effects model.

studlab 	 A logical indicating whether study labels should be printed in the graph. A vector with study labels can also be provided (must be of same length as x$event.e then).

cex.studlab 	 Size of study labels.

... 	 Graphical parameters as in par may also be passed as arguments.

Details

A L'Abbe plot is drawn in the active graphics window.

If comb.fixed is TRUE, the pooled estimate of the fixed effect model is plotted as a line. If comb.random is TRUE, the pooled estimate of the random effects model is plotted as a line.

Information from object x is utilised if argument weight is missing. Weights from the fixed effect model are used (weight="fixed") if parameter x$comb.fixed is TRUE; weights from the random effects model are used (weight="random") if parameter x$comb.random is TRUE and x$comb.fixed is FALSE.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

L'Abbe KA, Detsky AS, O'Rourke K (1987), Meta-analysis in clinical research. Annals of Internal Medicine, 107, 224–233.

See Also

metabin

Examples

data(Olkin95)
meta1 <- metabin(event.e, n.e, event.c, n.c,
                 data=Olkin95,
                 studlab=paste(author, year),
                 sm="RR", meth="I")

##
## L'Abbe plot
##
labbe(meta1)

[Package meta version 1.1-8 Index]