meta-package {meta}	R Documentation

meta: Brief overview of methods and general hints

Description

R package meta is a user-friendly general package providing standard methods for meta-analysis and supporting Schwarzer et al. (2015), http://meta-analysis-with-r.org/.

Details

R package meta (Schwarzer, 2007) provides the following statistical methods for meta-analysis.

Fixed effect and random effects model:

Meta-analysis of continuous outcome data (metacont)

Meta-analysis of binary outcome data (metabin)

Meta-analysis of incidence rates (metainc)

Generic inverse variance meta-analysis (metagen)

Meta-analysis of single correlations (metacor)

Meta-analysis of single means (metamean)

Meta-analysis of single proportions (metaprop)

Meta-analysis of single incidence rates (metarate)

Several plots for meta-analysis:

Forest plot (forest)

Funnel plot (funnel)

Galbraith plot / radial plot (radial)

L'Abbe plot for meta-analysis with binary outcome data (labbe)

Baujat plot to explore heterogeneity in meta-analysis (baujat)

Bubble plot to display the result of a meta-regression (bubble)

Statistical tests for funnel plot asymmetry (metabias) and trim-and-fill method (trimfill) to evaluate bias in meta-analysis

Import data from 'RevMan 5' (read.rm5); see also metacr to conduct meta-analysis for a single comparison and outcome from a Cochrane review

Prediction interval, Hartung-Knapp and Paule-Mandel method for random effects model (see arguments prediction, hakn, and method.tau, respectively, in meta-analysis functions listed under 1. Fixed effect and random effects model)

Cumulative meta-analysis (metacum) and leave-one-out meta-analysis (metainf)

Meta-regression (metareg); if R package metafor is installed

Generalised linear mixed models (metabin, metainc, metaprop, and metarate); if R packages metafor, lme4, numDeriv, and BiasedUrn are installed

The following more advanced statistical methods are provided by add-on R packages:

Frequentist methods for network meta-analysis (R package netmeta)

Advanced methods to model and adjust for bias in meta-analysis (R package metasens)

Results of several meta-analyses can be combined with metabind. This is, for example, useful to generate a forest plot with results of subgroup analyses.

See settings.meta to learn how to print and specify default meta-analysis methods used during your R session. For example, the function can be used to specify general settings:

settings.meta("revman5")

settings.meta("jama")

The first command can be used to reproduce meta-analyses from Cochrane reviews conducted with Review Manager 5 (RevMan 5, http://community.cochrane.org/tools/review-production-tools/revman-5) and specifies to use a RevMan 5 layout in forest plots. The second command can be used to generate forest plots following instructions for authors of the Journal of the American Medical Association (http://jamanetwork.com/journals/jama/pages/instructions-for-authors).

In addition, settings.meta can be used to change individual settings. For example, the following R command specifies the use of the Hartung-Knapp and Paule-Mandel methods, and the printing of prediction intervals in the current R session for any meta-analysis generated after execution of this command:

settings.meta(hakn=TRUE, method.tau="PM", prediction=TRUE)

Type help(package = "meta") for a listing of R functions and datasets available in meta.

Schwarzer (2007) is the preferred citation in publications for meta. Type citation("meta") for a BibTeX entry of this publication.

To report problems and bugs

type bug.report(package = "meta") if you do not use RStudio,

send an email to Guido Schwarzer sc@imbi.uni-freiburg.de if you use RStudio.

The development version of meta is available on GitHub https://github.com/guido-s/meta.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Schwarzer G (2007), meta: An R package for meta-analysis. R News, 7(3), 40–5. https://cran.r-project.org/doc/Rnews/Rnews_2007-3.pdf

Schwarzer G, Carpenter JR and Rücker G (2015), Meta-Analysis with R (Use-R!). Springer International Publishing, Switzerland. http://www.springer.com/gp/book/9783319214153

[Package meta version 4.9-1 Index]