metabias {meta}	R Documentation

Test for funnel plot asymmetry

Description

Test for funnel plot asymmetry, based on rank correlation or linear regression method.

Usage

metabias(x, ...)

## S3 method for class 'meta'
metabias(x, method.bias=x$method.bias,
         plotit=FALSE, correct=FALSE, k.min=10, ...)

## Default S3 method:
metabias(x, seTE, method.bias="linreg",
         plotit=FALSE, correct=FALSE, k.min=10, ...)

Arguments

x
An object of class meta or estimated treatment effect in individual studies.

seTE
Standard error of estimated treatment effect (mandatory if x not of class meta).

method.bias
A character string indicating which test is to be used. Either "rank", "linreg", "mm", "count", "score", or "peters", can be abbreviated.

plotit
A logical indicating whether a plot should be produced for method.bias "rank", "linreg", "mm", or "score".

correct
A logical indicating whether a continuity corrected statistic is used for rank correlation methods "rank" and "count".

k.min
Minimum number of studies to perform test for funnel plot asymmetry.

...
Additional arguments (ignored at the moment).

Details

Following recommendations by Sterne et al. (2011), by default, a test for funnel plot asymmetry is only conducted if the number of studies is ten or larger (argument k.min=10). This behaviour can be changed by setting a smaller value for argument k.min. Note, the minimum number of studies is three.

If argument method.bias is "rank", the test statistic is based on the rank correlation between standardised treatment estimates and variance estimates of estimated treatment effects; Kendall's tau is used as correlation measure (Begg & Mazumdar, 1994). The test statistic follows a standard normal distribution. By default (if correct is FALSE), no continuity correction is utilised (Kendall & Gibbons, 1990).

If argument method.bias is "linreg", the test statistic is based on a weighted linear regression of the treatment effect on its standard error (Egger et al., 1997). The test statistic follows a t distribution with number of studies - 2 degrees of freedom.

If argument method.bias is "mm", the test statistic is based on a weighted linear regression of the treatment effect on its standard error using the method of moments estimator for the additive between-study variance component (method 3a in Thompson, Sharp, 1999). The test statistic follows a t distribution with number of studies - 2 degrees of freedom.

If argument method.bias is "peters", the test statistic is based on a weighted linear regression of the treatment effect on the inverse of the total sample size using the variance of the average event rate as weights (Peters et al., 2006). The test statistic follows a t distribution with number of studies - 2 degrees of freedom. This test is available for meta-analyses comparing two binary outcomes or combining single proportions, i.e. generated with functions metabin and metaprop.

The following tests for funnel plot asymmetry are only available for meta-analyses comparing two binary outcomes, i.e. meta-analyses generated with the metabin function.

If argument method.bias is "count", the test statistic is based on the rank correlation between a standardised cell frequency and the inverse of the variance of the cell frequency; Kendall's tau is used as correlation measure (Schwarzer et al., 2007). The test statistic follows a standard normal distribution. By default (if correct is FALSE), no continuity correction is utilised (Kendall & Gibbons, 1990).

If argument method.bias is "score", the test statistic is based on a weighted linear regression utilising efficient score and score variance (Harbord et al., 2006). The test statistic follows a t distribution with number of studies - 2 degrees of freedom.

In order to calculate an arcsine test for funnel plot asymmetry (Rücker et al., 2008), one has to use the metabin function with argument sm="ASD" as input to the metabias command. The three arcsine tests described in Rücker et al. (2008) can be calculated by setting method.bias to "rank", "linreg" and "mm", respectively.

If argument method.bias is missing, the Harbord test (method.bias="score") is used for the odds ratio as effect measure and the Egger test (method.bias="linreg") for other effect measures (Sterne et al., 2011).

No test for funnel plot asymmetry is conducted in meta-analyses with subgroups.

Value

A list with class "htest" containing the following components if a test for funnel plot asymmetry is conducted:

estimate
The estimated degree of funnel plot asymmetry, with name "ks" or "bias" corresponding to the method employed, i.e., rank correlation or regression method.

statistic
The value of the test statistic.

parameters
The degrees of freedom of the test statistic in the case that it follows a t distribution.

p.value
The p-value for the test.

alternative
A character string describing the alternative hypothesis.

method
A character string indicating what type of test was used.

data.name
A character string giving the names of the data.

title
Title of Cochrane review.

complab
Comparison label.

outclab
Outcome label.

version
Version of R package meta used to create object.

Or a list with the following elements if test is not conducted due to the number of studies:

k
Number of studies in meta-analysis.

k.min
Minimum number of studies to perform test for funnel plot asymmetry.

version
Version of R package meta used to create object.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Begg CB & Mazumdar M (1994), Operating characteristics of a rank correlation test for publication bias. Biometrics, 50, 1088–1101.

Egger M, Smith GD, Schneider M & Minder C (1997), Bias in meta-analysis detected by a simple, graphical test. British Medical Journal, 315, 629–634.

Harbord RM, Egger M & Sterne J (2006), A modified test for small-study effects in meta-analyses of controlled trials with binary endpoints. Statistics in Medicine, 25, 3443–3457.

Kendall M & Gibbons JD (1990), Rank Correlation Methods. London: Edward Arnold.

Peters JL, Sutton AJ, Jones DR, Abrams KR & Rushton L (2006), Comparison of two methods to detect publication bias in meta-analysis. Journal of the American Medical Association, 295, 676–680.

Rücker G, Schwarzer G, Carpenter JR (2008) Arcsine test for publication bias in meta-analyses with binary outcomes. Statistics in Medicine, 27,746–763.

Schwarzer G, Antes G & Schumacher M (2007), A test for publication bias in meta-analysis with sparse binary data. Statistics in Medicine, 26, 721–733.

Sterne, JAC et al. (2011), Recommendations for Examining and Interpreting Funnel Plot Asymmetry in Meta-Analyses of Randomised Controlled Trials. BMJ (Clinical research ed.), 343, 1, doi: 10.1136/bmj.d4002 .

Thompson SG & Sharp, SJ (1999), Explaining heterogeneity in meta-analysis: A comparison of methods, Statistics in Medicine, 18, 2693–2708.

See Also

funnel, funnel.meta, metabin, metacont, metagen

Examples

data(Olkin95)
meta1 <- metabin(event.e, n.e, event.c, n.c,
                 data=Olkin95, subset=1:10,
                 sm="RR", method="I")

metabias(meta1)
metabias(meta1, plotit=TRUE)

metabias(meta1, method.bias="rank")
metabias(meta1, method.bias="rank", correct=TRUE)

metabias(meta1, method.bias="count")

metabias(meta1, method.bias="linreg")$p.value

#
# Arcsine test (based on linear regression):
#
meta1.as <- metabin(event.e, n.e, event.c, n.c,
                    data=Olkin95, subset=1:10,
                    sm="ASD", method="I")
metabias(meta1.as)
#
# Same result (using function metabias.default):
#
metabias(meta1.as$TE, meta1.as$seTE)

#
# No test for funnel plot asymmetry calculated:
#
meta2 <- metabin(event.e, n.e, event.c, n.c,
                 data=Olkin95, subset=1:5,
                 sm="RR", method="I")
metabias(meta2)

meta3 <- metabin(event.e, n.e, event.c, n.c,
                 data=Olkin95, subset=1:2,
                 sm="RR", method="I")
metabias(meta3)

# Test for funnel plot asymmetry calculated
# (use of argument k.min):
#
metabias(meta2, k.min=5)

[Package meta version 4.9-1 Index]