metabias.rm5 {meta}	R Documentation

Cochrane review: Test for funnel plot asymmetry

Description

Conduct a test for funnel plot asymmetry for all outcomes in a Cochrane review

Usage

## S3 method for class 'rm5'
metabias(
  x,
  comp.no,
  outcome.no,
  method.bias = "linreg",
  method.bias.binary = method.bias,
  method.bias.or = "score",
  k.min = 10,
  ...
)

Arguments

x
An object of class rm5.

comp.no
Comparison number.

outcome.no
Outcome number.

method.bias
A character string indicating which test for small-study effects is to be used for all outcomes. Either "rank", "linreg", or "mm", can be abbreviated. See function metabias

method.bias.binary
A character string indicating which test is to be used for binary outcomes. Either "rank", "linreg", "mm", "count", "score", or "peters", can be abbreviated. See function metabias

method.bias.or
A character string indicating which test is to be used for binary outcomes with odds ratio as summary measure. Either "rank", "linreg", "mm", "count", "score", or "peters", can be abbreviated. See function metabias

k.min
Minimum number of studies to perform test for small-study effects.

...
Additional arguments (ignored at the moment)

Details

This function can be used to conduct a test for funnel plot asymmetry for all or selected meta-analyses in a Cochrane Review.

Review Manager 5 (RevMan 5) is the current software used for preparing and maintaining Cochrane Reviews (https://training.cochrane.org/online-learning/core-software-cochrane-reviews/revman). In RevMan 5, subgroup analyses can be defined and data from a Cochrane review can be imported to R using the function read.rm5.

The R function metacr is called internally.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Higgins, J.P.T and S. Green (2011): Cochrane Handbook for Systematic Reviews of Interventions Version 5.1.0 [Updated March 2011]. The Cochrane Library: http://www.cochrane-handbook.org/

See Also

metabias, metacr, read.rm5, summary.rm5

Examples

# Locate export data file "Fleiss1993_CR.csv" in sub-directory of
# package "meta"
#
filename <- system.file("extdata", "Fleiss1993_CR.csv", package = "meta")
Fleiss1993_CR <- read.rm5(filename)

# Print results for all tests of small-study effects
#
metabias(Fleiss1993_CR, k.min = 5)

# Print result of test of small-study effects for second outcome in
# first comparison
#
metabias(Fleiss1993_CR, comp.no = 1, outcome.no = 2, k.min = 5)

[Package meta version 4.18-1 Index]