metacont {meta}	R Documentation

Meta-analysis of continuous outcome data

Description

Calculation of fixed and random effects estimates for meta-analyses with continuous outcome data; inverse variance weighting is used for pooling.

Usage

metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, studlab,
         data=NULL, subset=NULL, exclude=NULL,
         sm=gs("smcont"), pooledvar=gs("pooledvar"),
         method.smd=gs("method.smd"), sd.glass=gs("sd.glass"),
         exact.smd=gs("exact.smd"),
         level=gs("level"), level.comb=gs("level.comb"),
         comb.fixed=gs("comb.fixed"), comb.random=gs("comb.random"),
         hakn=gs("hakn"),
         method.tau=gs("method.tau"), tau.preset=NULL, TE.tau=NULL,
         tau.common=gs("tau.common"),
         prediction=gs("prediction"), level.predict=gs("level.predict"),
         method.bias=gs("method.bias"),
         backtransf=gs("backtransf"),
         title=gs("title"), complab=gs("complab"), outclab="",
         label.e=gs("label.e"), label.c=gs("label.c"),
         label.left=gs("label.left"), label.right=gs("label.right"),
         byvar, bylab, print.byvar=gs("print.byvar"),
         byseparator=gs("byseparator"),
         keepdata=gs("keepdata"),
         warn=gs("warn"))

Arguments

n.e
Number of observations in experimental group.

mean.e
Estimated mean in experimental group.

sd.e
Standard deviation in experimental group.

n.c
Number of observations in control group.

mean.c
Estimated mean in control group.

sd.c
Standard deviation in control group.

studlab
An optional vector with study labels.

data
An optional data frame containing the study information.

subset
An optional vector specifying a subset of studies to be used.

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots.

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

hakn
A logical indicating whether the method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2. Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.

tau.preset
Prespecified value for the square-root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance tau-squared.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

method.bias
A character string indicating which test is to be used. Either "rank", "linreg", or "mm", can be abbreviated. See function metabias

backtransf
A logical indicating whether results for ratio of means (sm="ROM") should be back transformed in printouts and plots. If TRUE (default), results will be presented as ratio of means; otherwise log ratio of means will be shown.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

label.e
Label for experimental group.

label.c
Label for control group.

label.left
Graph label on left side of forest plot.

label.right
Graph label on right side of forest plot.

sm
A character string indicating which summary measure ("MD", "SMD", or "ROM") is to be used for pooling of studies.

pooledvar
A logical indicating if a pooled variance should be used for the mean difference (sm="MD").

method.smd
A character string indicating which method is used to estimate the standardised mean difference (sm="SMD"). Either "Hedges" for Hedges' g (default), "Cohen" for Cohen's d, or "Glass" for Glass' delta, can be abbreviated.

sd.glass
A character string indicating which standard deviation is used in the denominator for Glass' method to estimate the standardised mean difference. Either "control" using the standard deviation in the control group (sd.c) or "experimental" using the standard deviation in the experimental group (sd.e), can be abbreviated.

exact.smd
A logical indicating whether exact formulae should be used in estimation of the standardised mean difference and its standard error (see Details).

byvar
An optional vector containing grouping information (must be of same length as n.e).

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

warn
A logical indicating whether warnings should be printed (e.g., if studies are excluded from meta-analysis due to zero standard deviations).

Details

Calculation of fixed and random effects estimates for meta-analyses with continuous outcome data; inverse variance weighting is used for pooling.

Three different types of summary measures are available for continuous outcomes:

mean difference (argument sm="MD")

standardised mean difference (sm="SMD")

ratio of means (sm="ROM")

Meta-analysis of ratio of means – also called response ratios – is described in Hedges et al. (1999) and Friedrich et al. (2008).

For the standardised mean difference three methods are implemented:

Hedges' g (default, method.smd="Hedges") - see Hedges (1981)

Cohen's d (method.smd="Cohen") - see Cohen (1988)

Glass' delta (method.smd="Glass") - see Glass (1976)

Hedges (1981) calculated the exact bias in Cohen's d which is a ratio of gamma distributions with the degrees of freedom, i.e. total sample size minus two, as argument. By default (argument exact.smd=FALSE), an accurate approximation of this bias provided in Hedges (1981) is utilised for Hedges' g as well as its standard error; these approximations are also used in RevMan 5. Following Borenstein et al. (2009) these approximations are not used in the estimation of Cohen's d. White and Thomas (2005) argued that approximations are unnecessary with modern software and accordingly promote to use the exact formulae; this is possible using argument exact.smd=TRUE. For Hedges' g the exact formulae are used to calculate the standardised mean difference as well as the standard error; for Cohen's d the exact formula is only used to calculate the standard error. In typical applications (with sample sizes above 10), the differences between using the exact formulae and the approximation will be minimal.

For Glass' delta, by default (argument sd.glass="control"), the standard deviation in the control group (sd.c) is used in the denominator of the standard mean difference. The standard deviation in the experimental group (sd.e) can be used by specifying sd.glass="experimental".

Calculations are conducted on the log scale for ratio of means (sm="ROM"). Accordingly, list elements TE, TE.fixed, and TE.random contain the logarithm of ratio of means. In printouts and plots these values are back transformed if argument backtransf=TRUE.

For several arguments defaults settings are utilised (assignments using gs function). These defaults can be changed using the settings.meta function.

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if argument comb.random=FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random=FALSE.

The function metagen is called internally to calculate individual and overall treatment estimates and standard errors.

A prediction interval for treatment effect of a new study is calculated (Higgins et al., 2009) if arguments prediction and comb.random are TRUE.

R function update.meta can be used to redo the meta-analysis of an existing metacont object by only specifying arguments which should be changed.

For the random effects, the method by Hartung and Knapp (2003) is used to adjust test statistics and confidence intervals if argument hakn=TRUE.

The DerSimonian-Laird estimate (1986) is used in the random effects model if method.tau="DL". The iterative Paule-Mandel method (1982) to estimate the between-study variance is used if argument method.tau="PM". Internally, R function paulemandel is called which is based on R function mpaule.default from R package metRology from S.L.R. Ellison <s.ellison at lgc.co.uk>.

If R package metafor (Viechtbauer 2010) is installed, the following methods to estimate the between-study variance τ^2 (argument method.tau) are also available:

Restricted maximum-likelihood estimator (method.tau="REML")

Maximum-likelihood estimator (method.tau="ML")

Hunter-Schmidt estimator (method.tau="HS")

Sidik-Jonkman estimator (method.tau="SJ")

Hedges estimator (method.tau="HE")

Empirical Bayes estimator (method.tau="EB").

For these methods the R function rma.uni of R package metafor is called internally. See help page of R function rma.uni for more details on these methods to estimate between-study variance.

Value

An object of class c("metacont", "meta") with corresponding print, summary, and forest functions. The object is a list containing the following components:

n.e, mean.e, sd.e,

n.c, mean.c, sd.c,

studlab, exclude, sm, level, level.comb,

comb.fixed, comb.random,

pooledvar, method.smd, sd.glass,

hakn, method.tau, tau.preset, TE.tau, method.bias,

tau.common, title, complab, outclab,

label.e, label.c, label.left, label.right,

byvar, bylab, print.byvar, byseparator, warn
As defined above.

TE, seTE
Estimated treatment effect and standard error of individual studies.

lower, upper
Lower and upper confidence interval limits for individual studies.

zval, pval
z-value and p-value for test of treatment effect for individual studies.

w.fixed, w.random
Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed
Estimated overall treatment effect and standard error (fixed effect model).

lower.fixed, upper.fixed
Lower and upper confidence interval limits (fixed effect model).

zval.fixed, pval.fixed
z-value and p-value for test of overall treatment effect (fixed effect model).

TE.random, seTE.random
Estimated overall treatment effect and standard error (random effects model).

lower.random, upper.random
Lower and upper confidence interval limits (random effects model).

zval.random, pval.random
z-value or t-value and corresponding p-value for test of overall treatment effect (random effects model).

prediction, level.predict
As defined above.

seTE.predict
Standard error utilised for prediction interval.

lower.predict, upper.predict
Lower and upper limits of prediction interval.

k
Number of studies combined in meta-analysis.

Q
Heterogeneity statistic.

tau
Square-root of between-study variance.

se.tau
Standard error of square-root of between-study variance.

C
Scaling factor utilised internally to calculate common tau-squared across subgroups.

method
Pooling method: "Inverse".

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn=TRUE).

bylevs
Levels of grouping variable - if byvar is not missing.

TE.fixed.w, seTE.fixed.w
Estimated treatment effect and standard error in subgroups (fixed effect model) - if byvar is not missing.

lower.fixed.w, upper.fixed.w
Lower and upper confidence interval limits in subgroups (fixed effect model) - if byvar is not missing.

zval.fixed.w, pval.fixed.w
z-value and p-value for test of treatment effect in subgroups (fixed effect model) - if byvar is not missing.

TE.random.w, seTE.random.w
Estimated treatment effect and standard error in subgroups (random effects model) - if byvar is not missing.

lower.random.w, upper.random.w
Lower and upper confidence interval limits in subgroups (random effects model) - if byvar is not missing.

zval.random.w, pval.random.w
z-value or t-value and corresponding p-value for test of treatment effect in subgroups (random effects model) - if byvar is not missing.

w.fixed.w, w.random.w
Weight of subgroups (in fixed and random effects model) - if byvar is not missing.

df.hakn.w
Degrees of freedom for test of treatment effect for Hartung-Knapp method in subgroups - if byvar is not missing and hakn=TRUE.

n.harmonic.mean.w
Harmonic mean of number of observations in subgroups (for back transformation of Freeman-Tukey Double arcsine transformation) - if byvar is not missing.

n.e.w
Number of observations in experimental group in subgroups - if byvar is not missing.

n.c.w
Number of observations in control group in subgroups - if byvar is not missing.

k.w
Number of studies combined within subgroups - if byvar is not missing.

k.all.w
Number of all studies in subgroups - if byvar is not missing.

Q.w
Heterogeneity statistics within subgroups - if byvar is not missing.

Q.w.fixed
Overall within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.w.random
Overall within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing (only calculated if argument tau.common is TRUE).

df.Q.w
Degrees of freedom for test of overall within subgroups heterogeneity - if byvar is not missing.

Q.b.fixed
Overall between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.b.random
Overall between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

df.Q.b
Degrees of freedom for test of overall between subgroups heterogeneity - if byvar is not missing.

tau.w
Square-root of between-study variance within subgroups - if byvar is not missing.

C.w
Scaling factor utilised internally to calculate common tau-squared across subgroups - if byvar is not missing.

H.w
Heterogeneity statistic H within subgroups - if byvar is not missing.

lower.H.w, upper.H.w
Lower and upper confidence limti for heterogeneity statistic H within subgroups - if byvar is not missing.

I2.w
Heterogeneity statistic I2 within subgroups - if byvar is not missing.

lower.I2.w, upper.I2.w
Lower and upper confidence limti for heterogeneity statistic I2 within subgroups - if byvar is not missing.

keepdata
As defined above.

data
Original data (set) used in function call (if keepdata=TRUE).

subset
Information on subset of original data used in meta-analysis (if keepdata=TRUE).

call
Function call.

version
Version of R package meta used to create object.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Borenstein et al. (2009), Introduction to Meta-Analysis, Chichester: Wiley.

Cohen J (1988), Statistical Power Analysis for the Behavioral Sciences (second ed.), Lawrence Erlbaum Associates.

Cooper H & Hedges LV (1994), The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation.

DerSimonian R & Laird N (1986), Meta-analysis in clinical trials. Controlled Clinical Trials, 7, 177–88.

Friedrich JO, Adhikari NK, Beyene J (2008), The ratio of means method as an alternative to mean differences for analyzing continuous outcome variables in meta-analysis: A simulation study. BMC Med Res Methodol, 8, 32.

Glass G (1976), Primary, secondary, and meta-analysis of research. Educational Researcher, 5, 3–8.

Hartung J & Knapp G (2001), On tests of the overall treatment effect in meta-analysis with normally distributed responses. Statistics in Medicine, 20, 1771–82. doi: 10.1002/sim.791 .

Hedges LV (1981), Distribution theory for Glass's estimator of effect size and related estimators. Journal of Educational and Behavioral Statistics, 6, 107–28.

Hedges LV, Gurevitch J, Curtis PS (1999), The meta-analysis of response ratios in experimental ecology. Ecology, 80, 1150–6.

Higgins JPT, Thompson SG, Spiegelhalter DJ (2009), A re-evaluation of random-effects meta-analysis. Journal of the Royal Statistical Society: Series A, 172, 137–59.

Knapp G & Hartung J (2003), Improved Tests for a Random Effects Meta-regression with a Single Covariate. Statistics in Medicine, 22, 2693–710, doi: 10.1002/sim.1482 .

Paule RC & Mandel J (1982), Consensus values and weighting factors. Journal of Research of the National Bureau of Standards, 87, 377–85.

Review Manager (RevMan) [Computer program]. Version 5.3. Copenhagen: The Nordic Cochrane Centre, The Cochrane Collaboration, 2014.

Viechtbauer W (2010), Conducting Meta-Analyses in R with the Metafor Package. Journal of Statistical Software, 36, 1–48.

White IR, Thomas J (2005), Standardized mean differences in individually-randomized and cluster-randomized trials, with applications to meta-analysis. Clinical Trials, 2, 141–51.

See Also

update.meta, metabin, metagen

Examples

data(Fleiss93cont)
meta1 <- metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, data=Fleiss93cont, sm="SMD")
meta1
forest(meta1)

meta2 <- metacont(Fleiss93cont$n.e, Fleiss93cont$mean.e,
                  Fleiss93cont$sd.e,
                  Fleiss93cont$n.c, Fleiss93cont$mean.c,
                  Fleiss93cont$sd.c,
                  sm="SMD")
meta2

data(amlodipine)
meta3 <- metacont(n.amlo, mean.amlo, sqrt(var.amlo),
                  n.plac, mean.plac, sqrt(var.plac),
                  data=amlodipine, studlab=study)
summary(meta3)

# Use pooled variance
#
meta4 <- metacont(n.amlo, mean.amlo, sqrt(var.amlo),
                  n.plac, mean.plac, sqrt(var.plac),
                  data=amlodipine, studlab=study,
                  pooledvar=TRUE)
summary(meta4)

# Use Cohen's d instead of Hedges' g as effect measure
#
meta5 <- metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, data=Fleiss93cont,
                  sm="SMD", method.smd="Cohen")
meta5

# Use Glass' delta instead of Hedges' g as effect measure
#
meta6 <- metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, data=Fleiss93cont,
                  sm="SMD", method.smd="Glass")
meta6

# Use Glass' delta based on the standard deviation in the experimental group
#
meta7 <- metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, data=Fleiss93cont,
                  sm="SMD", method.smd="Glass", sd.glass="experimental")
meta7

# Calculate Hedges' g based on exact formulae
#
update(meta1, exact.smd=TRUE)

#
# Meta-analysis of response ratios (Hedges et al., 1999)
#
data(woodyplants)
meta8 <- metacont(n.elev, mean.elev, sd.elev,
		  n.amb, mean.amb, sd.amb,
                  data=woodyplants, sm="ROM")
summary(meta8)
summary(meta8, backtransf=FALSE)

[Package meta version 4.9-1 Index]