metacor {meta}	R Documentation

Meta-analysis of correlations

Description

Calculation of fixed and random effects estimates for meta-analyses with correlations; inverse variance weighting is used for pooling.

Usage

metacor(cor, n, studlab,
        data=NULL, subset=NULL, exclude=NULL,
        sm=gs("smcor"),
        level=gs("level"), level.comb=gs("level.comb"),
        comb.fixed=gs("comb.fixed"), comb.random=gs("comb.random"),
        hakn=gs("hakn"),
        method.tau=gs("method.tau"), tau.preset=NULL, TE.tau=NULL,
        tau.common=gs("tau.common"),
        prediction=gs("prediction"), level.predict=gs("level.predict"),
	null.effect=0,
        method.bias=gs("method.bias"),
        backtransf=gs("backtransf"),
        title=gs("title"), complab=gs("complab"), outclab="",
        byvar, bylab, print.byvar=gs("print.byvar"),
        byseparator = gs("byseparator"),
        keepdata=gs("keepdata"),
        control=NULL
        )

Arguments

cor
Correlation.

n
Number of observations.

studlab
An optional vector with study labels.

data
An optional data frame containing the study information, i.e., cor and n.

subset
An optional vector specifying a subset of studies to be used.

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots.

sm
A character string indicating which summary measure ("ZCOR" or "COR") is to be used for pooling of studies.

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

hakn
A logical indicating whether the method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2. Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.

tau.preset
Prespecified value for the square-root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance tau-squared.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

null.effect
A numeric value specifying the effect under the null hypothesis.

method.bias
A character string indicating which test is to be used. Either "rank", "linreg", or "mm", can be abbreviated. See function metabias

backtransf
A logical indicating whether results for Fisher's z transformed correlations (sm="ZCOR") should be back transformed in printouts and plots. If TRUE (default), results will be presented as correlations; otherwise Fisher's z transformed correlations will be shown.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

byvar
An optional vector containing grouping information (must be of same length as event.e).

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

control
An optional list to control the iterative process to estimate the between-study variance tau^2. This argument is passed on to rma.uni.

Details

Fixed effect and random effects meta-analysis of correlations based either on Fisher's z transformation of correlations (sm="ZCOR") or direct combination of correlations (sm="COR") (see Cooper et al., p264-5 and p273-4).

Only few statisticians would advocate the use of untransformed correlations unless sample sizes are very large (see Cooper et al., p265). The artificial example given below shows that the smallest study gets the largest weight if correlations are combined directly because the correlation is closest to 1.

For several arguments defaults settings are utilised (assignments using gs function). These defaults can be changed using the settings.meta function.

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if argument comb.random=FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random=FALSE.

A prediction interval for treatment effect of a new study is calculated (Higgins et al., 2009) if arguments prediction and comb.random are TRUE.

R function update.meta can be used to redo the meta-analysis of an existing metacor object by only specifying arguments which should be changed.

For the random effects, the method by Hartung and Knapp (2003) is used to adjust test statistics and confidence intervals if argument hakn=TRUE.

The DerSimonian-Laird estimate (1986) is used in the random effects model if method.tau="DL". The iterative Paule-Mandel method (1982) to estimate the between-study variance is used if argument method.tau="PM". Internally, R function paulemandel is called which is based on R function mpaule.default from R package metRology from S.L.R. Ellison <s.ellison at lgc.co.uk>.

If R package metafor (Viechtbauer 2010) is installed, the following methods to estimate the between-study variance τ^2 (argument method.tau) are also available:

Restricted maximum-likelihood estimator (method.tau="REML")

Maximum-likelihood estimator (method.tau="ML")

Hunter-Schmidt estimator (method.tau="HS")

Sidik-Jonkman estimator (method.tau="SJ")

Hedges estimator (method.tau="HE")

Empirical Bayes estimator (method.tau="EB").

For these methods the R function rma.uni of R package metafor is called internally. See help page of R function rma.uni for more details on these methods to estimate between-study variance.

Value

An object of class c("metacor", "meta") with corresponding print, summary, and forest functions. The object is a list containing the following components:

cor, n, studlab, exclude,

sm, level, level.comb,

comb.fixed, comb.random,

hakn, method.tau, tau.preset, TE.tau, null.effect,

method.bias, tau.common, title, complab, outclab,

byvar, bylab, print.byvar, byseparator
As defined above.

TE, seTE
Either Fisher's z transformation of correlations (sm="ZCOR") or correlations (sm="COR") for individual studies.

lower, upper
Lower and upper confidence interval limits for individual studies.

zval, pval
z-value and p-value for test of treatment effect for individual studies.

w.fixed, w.random
Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed
Estimated overall effect (Fisher's z transformation of correlation or correlation) and its standard error (fixed effect model).

lower.fixed, upper.fixed
Lower and upper confidence interval limits (fixed effect model).

zval.fixed, pval.fixed
z-value and p-value for test of overall effect (fixed effect model).

TE.random, seTE.random
Estimated overall effect (Fisher's z transformation of correlation or correlation) and its standard error (random effects model).

lower.random, upper.random
Lower and upper confidence interval limits (random effects model).

zval.random, pval.random
z-value or t-value and corresponding p-value for test of overall effect (random effects model).

prediction, level.predict
As defined above.

seTE.predict
Standard error utilised for prediction interval.

lower.predict, upper.predict
Lower and upper limits of prediction interval.

k
Number of studies combined in meta-analysis.

Q
Heterogeneity statistic Q.

tau
Square-root of between-study variance.

se.tau
Standard error of square-root of between-study variance.

C
Scaling factor utilised internally to calculate common tau-squared across subgroups.

method
A character string indicating method used for pooling: "Inverse"

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn=TRUE).

bylevs
Levels of grouping variable - if byvar is not missing.

TE.fixed.w, seTE.fixed.w
Estimated treatment effect and standard error in subgroups (fixed effect model) - if byvar is not missing.

lower.fixed.w, upper.fixed.w
Lower and upper confidence interval limits in subgroups (fixed effect model) - if byvar is not missing.

zval.fixed.w, pval.fixed.w
z-value and p-value for test of treatment effect in subgroups (fixed effect model) - if byvar is not missing.

TE.random.w, seTE.random.w
Estimated treatment effect and standard error in subgroups (random effects model) - if byvar is not missing.

lower.random.w, upper.random.w
Lower and upper confidence interval limits in subgroups (random effects model) - if byvar is not missing.

zval.random.w, pval.random.w
z-value or t-value and corresponding p-value for test of treatment effect in subgroups (random effects model) - if byvar is not missing.

w.fixed.w, w.random.w
Weight of subgroups (in fixed and random effects model) - if byvar is not missing.

df.hakn.w
Degrees of freedom for test of treatment effect for Hartung-Knapp method in subgroups - if byvar is not missing and hakn=TRUE.

n.harmonic.mean.w
Harmonic mean of number of observations in subgroups (for back transformation of Freeman-Tukey Double arcsine transformation) - if byvar is not missing.

n.w
Number of observations in subgroups - if byvar is not missing.

k.w
Number of studies combined within subgroups - if byvar is not missing.

k.all.w
Number of all studies in subgroups - if byvar is not missing.

Q.w.fixed
Overall within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.w.random
Overall within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing (only calculated if argument tau.common is TRUE).

df.Q.w
Degrees of freedom for test of overall within subgroups heterogeneity - if byvar is not missing.

pval.Q.w.fixed
P-value of within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.w.random
P-value of within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

Q.b.fixed
Overall between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.b.random
Overall between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

df.Q.b
Degrees of freedom for test of overall between subgroups heterogeneity - if byvar is not missing.

pval.Q.b.fixed
P-value of between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.b.random
P-value of between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

tau.w
Square-root of between-study variance within subgroups - if byvar is not missing.

C.w
Scaling factor utilised internally to calculate common tau-squared across subgroups - if byvar is not missing.

H.w
Heterogeneity statistic H within subgroups - if byvar is not missing.

lower.H.w, upper.H.w
Lower and upper confidence limti for heterogeneity statistic H within subgroups - if byvar is not missing.

I2.w
Heterogeneity statistic I2 within subgroups - if byvar is not missing.

lower.I2.w, upper.I2.w
Lower and upper confidence limti for heterogeneity statistic I2 within subgroups - if byvar is not missing.

keepdata
As defined above.

data
Original data (set) used in function call (if keepdata=TRUE).

subset
Information on subset of original data used in meta-analysis (if keepdata=TRUE).

call
Function call.

version
Version of R package meta used to create object.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H, Hedges LV, Valentine JC (2009), The Handbook of Research Synthesis and Meta-Analysis, 2nd Edition. New York: Russell Sage Foundation.

DerSimonian R & Laird N (1986), Meta-analysis in clinical trials. Controlled Clinical Trials, 7, 177–188.

Higgins JPT, Thompson SG, Spiegelhalter DJ (2009), A re-evaluation of random-effects meta-analysis. Journal of the Royal Statistical Society: Series A, 172, 137–159.

Knapp G & Hartung J (2003), Improved Tests for a Random Effects Meta-regression with a Single Covariate. Statistics in Medicine, 22, 2693–710, doi: 10.1002/sim.1482 .

Paule RC & Mandel J (1982), Consensus values and weighting factors. Journal of Research of the National Bureau of Standards, 87, 377–385.

Viechtbauer W (2010), Conducting Meta-Analyses in R with the Metafor Package. Journal of Statistical Software, 36, 1–48.

See Also

update.meta, metacont, metagen, print.meta

Examples

m1 <- metacor(c(0.85, 0.7, 0.95), c(20, 40, 10))

#
# Print correlations (back transformed from Fisher's z transformation)
#
m1

#
# Print Fisher's z transformed correlations 
#
print(m1, backtransf=FALSE)

#
# Forest plot with back transformed correlations
#
forest(m1)

#
# Forest plot with Fisher's z transformed correlations
#
forest(m1, backtransf=FALSE)

m2 <- update(m1, sm="cor")
m2
# Identical forest plots (as back transformation is the identity transformation)
# forest(m2)
# forest(m2, backtransf=FALSE)

[Package meta version 4.9-4 Index]