metacr {meta}	R Documentation

Meta-analysis of outcome data from Cochrane review

Description

Wrapper function to perform meta-analysis for a single outcome of a Cochrane Intervention review.

Usage

metacr(
  x,
  comp.no = 1,
  outcome.no = 1,
  method,
  sm,
  level = gs("level"),
  level.comb = gs("level.comb"),
  comb.fixed,
  comb.random,
  hakn = FALSE,
  method.tau = "DL",
  method.tau.ci = gs("method.tau.ci"),
  tau.common = FALSE,
  prediction = gs("prediction"),
  level.predict = gs("level.predict"),
  swap.events,
  logscale,
  backtransf = gs("backtransf"),
  text.fixed = gs("text.fixed"),
  text.random = gs("text.random"),
  text.predict = gs("text.predict"),
  text.w.fixed = gs("text.w.fixed"),
  text.w.random = gs("text.w.random"),
  title,
  complab,
  outclab,
  keepdata = gs("keepdata"),
  warn = FALSE
)

Arguments

x
An object of class rm5 created by R function read.rm5.

comp.no
Comparison number.

outcome.no
Outcome number.

method
A character string indicating which method is to be used for pooling of studies. One of "Inverse", "MH", or "Peto", can be abbreviated.

sm
A character string indicating which summary measure ("RR", "OR", "RD", "ASD", "HR", "MD", or "SMD", or "ROM") is to be used for pooling of studies.

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

hakn
A logical indicating whether the method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2 and its square root τ. Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.

method.tau.ci
A character string indicating which method is used to estimate the confidence interval of τ^2 and τ. Either "QP", "BJ", or "J", or "", can be abbreviated.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

swap.events
A logical indicating whether events and non-events should be interchanged.

logscale
A logical indicating whether effect estimates are entered on log-scale.

backtransf
A logical indicating whether results should be back transformed in printouts and plots. If backtransf=TRUE (default), results for sm="OR" are printed as odds ratios rather than log odds ratios and results for sm="ZCOR" are printed as correlations rather than Fisher's z transformed correlations, for example.

text.fixed
A character string used in printouts and forest plot to label the pooled fixed effect estimate.

text.random
A character string used in printouts and forest plot to label the pooled random effects estimate.

text.predict
A character string used in printouts and forest plot to label the prediction interval.

text.w.fixed
A character string used to label weights of fixed effect model.

text.w.random
A character string used to label weights of random effects model.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

warn
A logical indicating whether warnings should be printed (e.g., if incr is added to studies with zero cell frequencies).

Details

Cochrane Intervention reviews are based on the comparison of two interventions. Each Cochrane Intervention review can have a variable number of comparisons. For each comparison, a variable number of outcomes can be define. For each outcome, a seperate meta-analysis is conducted. Review Manager 5 (RevMan 5) is the current software used for preparing and maintaining Cochrane Reviews (https://training.cochrane.org/online-learning/core-software-cochrane-reviews/revman).

This wrapper function can be used to perform meta-analysis for a single outcome of a Cochrane Intervention review. Internally, R functions metabin, metacont, and metagen are called - depending on the definition of the outcome in RevMan 5.

Note, it is recommended to choose the RevMan 5 settings before executing metacr, i.e., settings.meta("revman5").

Value

An object of class "meta" and "metabin", "metacont", or "metagen" depending on outcome type utilised in Cochrane Intervention review for selected outcome.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Review Manager (RevMan) [Computer program]. Version 5.4. The Cochrane Collaboration, 2020

See Also

metabin, metacont, metagen, read.rm5, settings.meta

Examples

# Locate export data file "Fleiss1993_CR.csv"
# in sub-directory of package "meta"
#
filename <- system.file("extdata", "Fleiss1993_CR.csv", package = "meta")
#
Fleiss1993_CR <- read.rm5(filename)

# Choose RevMan 5 settings and store old settings
#
oldset <- settings.meta("revman5")

# Same result as R command example(Fleiss1993bin)
#
metacr(Fleiss1993_CR)

# Same result as R command example(Fleiss1993cont)
#
metacr(Fleiss1993_CR, 1, 2)
forest(metacr(Fleiss1993_CR, 1, 2))

# Change summary measure to RR
#
m1 <- metacr(Fleiss1993_CR)
update(m1, sm="RR")

# Use old settings
#
settings.meta(oldset)

[Package meta version 4.18-1 Index]