metagen {meta}	R Documentation

Generic inverse variance meta-analysis

Description

Fixed and random effects meta-analysis based on estimates (e.g. log hazard ratios) and their standard errors; inverse variance weighting is used for pooling.

Usage

metagen(TE, seTE, studlab,
        data=NULL, subset=NULL, exclude=NULL, sm="",
        level=gs("level"), level.comb=gs("level.comb"),
        comb.fixed=gs("comb.fixed"), comb.random=gs("comb.random"),
        hakn=gs("hakn"),
        method.tau=gs("method.tau"), tau.preset=NULL, TE.tau=NULL,
        tau.common=gs("tau.common"),
        prediction=gs("prediction"), level.predict=gs("level.predict"),
        null.effect=0,
        method.bias=gs("method.bias"),
        n.e=NULL, n.c=NULL,
        backtransf=gs("backtransf"),
        pscale=1, irscale = 1, irunit = "person-years",
        title=gs("title"), complab=gs("complab"), outclab="",
        label.e=gs("label.e"), label.c=gs("label.c"),
        label.left=gs("label.left"), label.right=gs("label.right"),
        byvar, bylab, print.byvar=gs("print.byvar"),
        byseparator = gs("byseparator"),
        keepdata=gs("keepdata"),
        warn=gs("warn"), control=NULL)

Arguments

TE
Estimate of treatment effect, e.g., log hazard ratio or risk difference.

seTE
Standard error of treatment estimate.

studlab
An optional vector with study labels.

data
An optional data frame containing the study information.

subset
An optional vector specifying a subset of studies to be used.

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots.

sm
A character string indicating underlying summary measure, e.g., "RD", "RR", "OR", "ASD", "HR", "MD", "SMD", or "ROM".

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

null.effect
A numeric value specifying the effect under the null hypothesis.

n.e
Number of observations in experimental group.

n.c
Number of observations in control group.

hakn
A logical indicating whether method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2. Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.

tau.preset
Prespecified value for the square-root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance tau-squared.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

method.bias
A character string indicating which test is to be used. Either "rank", "linreg", or "mm", can be abbreviated. See function metabias

backtransf
A logical indicating whether results should be back transformed in printouts and plots. If backtransf=TRUE (default), results for sm="OR" are printed as odds ratios rather than log odds ratios and results for sm="ZCOR" are printed as correlations rather than Fisher's z transformed correlations, for example.

pscale
A numeric giving scaling factor for printing of single event probabilities or risk differences, i.e. if argument sm is equal to "PLOGIT", "PLN", "PRAW", "PAS", "PFT", or "RD".

irscale
A numeric defining a scaling factor for printing of single incidence rates or incidence rate differences, i.e. if argument sm is equal to "IR", "IRLN", "IRS", "IRFT", or "IRD".

irunit
A character specifying the time unit used to calculate rates, e.g. person-years.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

label.e
Label for experimental group.

label.c
Label for control group.

label.left
Graph label on left side of forest plot.

label.right
Graph label on right side of forest plot.

byvar
An optional vector containing grouping information (must be of same length as TE).

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

warn
A logical indicating whether warnings should be printed (e.g., if studies are excluded from meta-analysis due to zero standard errors).

control
An optional list to control the iterative process to estimate the between-study variance tau^2. This argument is passed on to rma.uni.

Details

Generic method for meta-analysis, only treatment estimates and their standard error are needed. The method is useful, e.g., for pooling of survival data (using log hazard ratio and standard errors as input). The inverse variance method is used for pooling.

For several arguments defaults settings are utilised (assignments using gs function). These defaults can be changed using the settings.meta function.

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if argument comb.random=FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random=FALSE.

A prediction interval for treatment effect of a new study is calculated (Higgins et al., 2009) if arguments prediction and comb.random are TRUE.

R function update.meta can be used to redo the meta-analysis of an existing metagen object by only specifying arguments which should be changed.

For the random effects, the method by Hartung and Knapp (2003) is used to adjust test statistics and confidence intervals if argument hakn=TRUE.

The DerSimonian-Laird estimate (1986) is used in the random effects model if method.tau="DL". The iterative Paule-Mandel method (1982) to estimate the between-study variance is used if argument method.tau="PM". Internally, R function paulemandel is called which is based on R function mpaule.default from R package metRology from S.L.R. Ellison <s.ellison at lgc.co.uk>.

If R package metafor (Viechtbauer 2010) is installed, the following methods to estimate the between-study variance τ^2 (argument method.tau) are also available:

Restricted maximum-likelihood estimator (method.tau="REML")

Maximum-likelihood estimator (method.tau="ML")

Hunter-Schmidt estimator (method.tau="HS")

Sidik-Jonkman estimator (method.tau="SJ")

Hedges estimator (method.tau="HE")

Empirical Bayes estimator (method.tau="EB").

For these methods the R function rma.uni of R package metafor is called internally. See help page of R function rma.uni for more details on these methods to estimate between-study variance.

Argument pscale can be used to rescale single proportions or risk differences, e.g. pscale=1000 means that proportions are expressed as events per 1000 observations. This is useful in situations with (very) low event probabilities.

Argument irscale can be used to rescale single rates or rate differences, e.g. irscale=1000 means that rates are expressed as events per 1000 time units, e.g. person-years. This is useful in situations with (very) low rates. Argument irunit can be used to specify the time unit used in individual studies (default: "person-years"). This information is printed in summaries and forest plots if argument irscale is not equal to 1.

Value

An object of class c("metagen", "meta") with corresponding print, summary, and forest functions. The object is a list containing the following components:

TE, seTE, studlab, exclude, n.e, n.c

sm, level, level.comb,

comb.fixed, comb.random,

hakn, method.tau, tau.preset, TE.tau, method.bias,

tau.common, title, complab, outclab,

label.e, label.c, label.left, label.right,

byvar, bylab, print.byvar, byseparator, warn
As defined above.

lower, upper
Lower and upper confidence interval limits for individual studies.

zval, pval
z-value and p-value for test of treatment effect for individual studies.

w.fixed, w.random
Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed
Estimated overall treatment effect and standard error (fixed effect model).

lower.fixed, upper.fixed
Lower and upper confidence interval limits (fixed effect model).

zval.fixed, pval.fixed
z-value and p-value for test of overall treatment effect (fixed effect model).

TE.random, seTE.random
Estimated overall treatment effect and standard error (random effects model).

lower.random, upper.random
Lower and upper confidence interval limits (random effects model).

zval.random, pval.random
z-value or t-value and corresponding p-value for test of overall treatment effect (random effects model).

prediction, level.predict
As defined above.

seTE.predict
Standard error utilised for prediction interval.

lower.predict, upper.predict
Lower and upper limits of prediction interval.

null.effect
As defined above.

k
Number of studies combined in meta-analysis.

Q
Heterogeneity statistic.

df.Q
Degrees of freedom for heterogeneity statistic.

pval.Q
P-value of heterogeneity test.

tau
Square-root of between-study variance.

se.tau
Standard error of square-root of between-study variance.

C
Scaling factor utilised internally to calculate common tau-squared across subgroups.

method
Pooling method: "Inverse".

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn=TRUE).

bylevs
Levels of grouping variable - if byvar is not missing.

TE.fixed.w, seTE.fixed.w
Estimated treatment effect and standard error in subgroups (fixed effect model) - if byvar is not missing.

lower.fixed.w, upper.fixed.w
Lower and upper confidence interval limits in subgroups (fixed effect model) - if byvar is not missing.

zval.fixed.w, pval.fixed.w
z-value and p-value for test of treatment effect in subgroups (fixed effect model) - if byvar is not missing.

TE.random.w, seTE.random.w
Estimated treatment effect and standard error in subgroups (random effects model) - if byvar is not missing.

lower.random.w, upper.random.w
Lower and upper confidence interval limits in subgroups (random effects model) - if byvar is not missing.

zval.random.w, pval.random.w
z-value or t-value and corresponding p-value for test of treatment effect in subgroups (random effects model) - if byvar is not missing.

w.fixed.w, w.random.w
Weight of subgroups (in fixed and random effects model) - if byvar is not missing.

df.hakn.w
Degrees of freedom for test of treatment effect for Hartung-Knapp method in subgroups - if byvar is not missing and hakn=TRUE.

n.harmonic.mean.w
Harmonic mean of number of observations in subgroups (for back transformation of Freeman-Tukey Double arcsine transformation) - if byvar is not missing.

n.e.w
Number of observations in experimental group in subgroups - if byvar is not missing.

n.c.w
Number of observations in control group in subgroups - if byvar is not missing.

k.w
Number of studies combined within subgroups - if byvar is not missing.

k.all.w
Number of all studies in subgroups - if byvar is not missing.

Q.w.fixed
Overall within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.w.random
Overall within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing (only calculated if argument tau.common is TRUE).

df.Q.w
Degrees of freedom for test of overall within subgroups heterogeneity - if byvar is not missing.

pval.Q.w.fixed
P-value of within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.w.random
P-value of within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

Q.b.fixed
Overall between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.b.random
Overall between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

df.Q.b
Degrees of freedom for test of overall between subgroups heterogeneity - if byvar is not missing.

pval.Q.b.fixed
P-value of between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.b.random
P-value of between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

tau.w
Square-root of between-study variance within subgroups - if byvar is not missing.

C.w
Scaling factor utilised internally to calculate common tau-squared across subgroups - if byvar is not missing.

H.w
Heterogeneity statistic H within subgroups - if byvar is not missing.

lower.H.w, upper.H.w
Lower and upper confidence limti for heterogeneity statistic H within subgroups - if byvar is not missing.

I2.w
Heterogeneity statistic I2 within subgroups - if byvar is not missing.

lower.I2.w, upper.I2.w
Lower and upper confidence limti for heterogeneity statistic I2 within subgroups - if byvar is not missing.

keepdata
As defined above.

data
Original data (set) used in function call (if keepdata=TRUE).

subset
Information on subset of original data used in meta-analysis (if keepdata=TRUE).

call
Function call.

version
Version of R package meta used to create object.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994), The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation.

DerSimonian R & Laird N (1986), Meta-analysis in clinical trials. Controlled Clinical Trials, 7, 177–188.

Higgins JPT, Thompson SG, Spiegelhalter DJ (2009), A re-evaluation of random-effects meta-analysis. Journal of the Royal Statistical Society: Series A, 172, 137–159.

Knapp G & Hartung J (2003), Improved Tests for a Random Effects Meta-regression with a Single Covariate. Statistics in Medicine, 22, 2693–2710, doi: 10.1002/sim.1482 .

Paule RC & Mandel J (1982), Consensus values and weighting factors. Journal of Research of the National Bureau of Standards, 87, 377–385.

Viechtbauer W (2010), Conducting Meta-Analyses in R with the Metafor Package. Journal of Statistical Software, 36, 1–48.

See Also

update.meta, metabin, metacont, print.meta

Examples

data(Fleiss93)
meta1 <- metabin(event.e, n.e, event.c, n.c, data=Fleiss93, sm="RR", method="I")
meta1

#
# Identical results by using the following commands:
#
meta1
metagen(meta1$TE, meta1$seTE, sm="RR")

forest(metagen(meta1$TE, meta1$seTE, sm="RR"))

#
# Meta-analysis with prespecified between-study variance
#
summary(metagen(meta1$TE, meta1$seTE, sm="RR", tau.preset=sqrt(0.1)))

#
# Meta-analysis of survival data:
#
logHR <- log(c(0.95, 1.5))
selogHR <- c(0.25, 0.35)

metagen(logHR, selogHR, sm="HR")

#
# Paule-Mandel method to estimate between-study variance
# Data from Paule & Mandel (1982)
#
average <- c(27.044, 26.022, 26.340, 26.787, 26.796)
variance <- c(0.003, 0.076, 0.464, 0.003, 0.014)
#
summary(metagen(average, sqrt(variance), sm="MD", method.tau="PM"))

[Package meta version 4.9-4 Index]