metainf {meta}	R Documentation

Influence analysis in meta-analysis using leave-one-out method

Description

Performs an influence analysis. Pooled estimates are calculated omitting one study at a time.

Usage

metainf(x, pooled, sortvar)

Arguments

x
An object of class meta.

pooled
A character string indicating whether a fixed effect or random effects model is used for pooling. Either missing (see Details), "fixed" or "random", can be abbreviated.

sortvar
An optional vector used to sort the individual studies (must be of same length as x$TE).

Details

Performs a influence analysis; pooled estimates are calculated omitting one study at a time. Studies are sorted according to sortvar.

Information from object x is utilised if argument pooled is missing. A fixed effect model is assumed (pooled="fixed") if argument x$comb.fixed is TRUE; a random effects model is assumed (pooled="random") if argument x$comb.random is TRUE and x$comb.fixed is FALSE.

Value

An object of class c("metainf", "meta") with corresponding print, and forest functions. The object is a list containing the following components:

TE, seTE
Estimated treatment effect and standard error of pooled estimate in influence analysis.

lower, upper
Lower and upper confidence interval limits.

studlab
Study label describing omission of studies.

p.value
P-value for test of overall effect.

w
Sum of weights from fixed effect or random effects model.

I2
Heterogeneity statistic I2.

Rb
Heterogeneity statistic Rb.

tau
Square-root of between-study variance.

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn=TRUE).

sm
Summary measure.

method
Method used for pooling.

k
Number of studies combined in meta-analysis.

pooled
As defined above.

comb.fixed
A logical indicating whether analysis is based on fixed effect model.

comb.random
A logical indicating whether analysis is based on random effects model.

TE.fixed, seTE.fixed
Value is NA.

TE.random, seTE.random
Value is NA.

Q
Value is NA.

level.comb
The level used to calculate confidence intervals for pooled estimates.

hakn
A logical indicating whether the method by Hartung and Knapp is used to adjust test statistics and confidence intervals.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2.

tau.preset
Prespecified value for the square-root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance τ^2.

n.harmonic.mean
Harmonic mean of number of observations (for back transformation of Freeman-Tukey Double arcsine transformation).

version
Version of R package meta used to create object.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994), The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation.

See Also

metabin, metacont, print.meta

Examples

data(Fleiss93)
meta1 <- metabin(event.e, n.e, event.c, n.c,
                 data=Fleiss93, studlab=study,
                 sm="RR", method="I")
meta1

metainf(meta1)
metainf(meta1, pooled="random")

forest(metainf(meta1))
forest(metainf(meta1), layout="revman5")
forest(metainf(meta1, pooled="random"))

metainf(meta1, sortvar=study)
metainf(meta1, sortvar=7:1)

meta2 <- update(meta1, title="Fleiss93 meta-analysis",
                backtransf=FALSE)
metainf(meta2)

data(Fleiss93cont)
meta3 <- metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c,
                  data = Fleiss93cont, sm = "SMD")
metainf(meta3)

[Package meta version 4.9-4 Index]