metaprop {meta}	R Documentation

Meta-analysis of single proportions

Description

Calculation of an overall proportion from studies reporting a single proportion. Inverse variance method and generalised linear mixed model (GLMM) are available for pooling. For GLMMs, the rma.glmm function from R package metafor (Viechtbauer 2010) is called internally.

Usage

metaprop(event, n, studlab,
         data=NULL, subset=NULL, exclude=NULL,
         method = "Inverse",
         sm=gs("smprop"),
         incr=gs("incr"), allincr=gs("allincr"),
         addincr=gs("addincr"),
         method.ci=gs("method.ci"),
         level=gs("level"), level.comb=gs("level.comb"),
         comb.fixed=gs("comb.fixed"), comb.random=gs("comb.random"),
         hakn=gs("hakn"),
         method.tau=
         ifelse(!is.na(charmatch(tolower(method), "glmm", nomatch = NA)),
                "ML", gs("method.tau")),
         tau.preset=NULL, TE.tau=NULL,
         tau.common=gs("tau.common"),
         prediction=gs("prediction"), level.predict=gs("level.predict"),
         null.effect=NA,
         method.bias=gs("method.bias"),
         backtransf=gs("backtransf"),
         pscale=1,
         title=gs("title"), complab=gs("complab"), outclab="",
         byvar, bylab, print.byvar=gs("print.byvar"),
         byseparator = gs("byseparator"),
         keepdata=gs("keepdata"),
         warn=gs("warn"),
         control=NULL,
         ...)

Arguments

event
Number of events.

n
Number of observations.

studlab
An optional vector with study labels.

data
An optional data frame containing the study information, i.e., event and n.

subset
An optional vector specifying a subset of studies to be used.

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots.

method
A character string indicating which method is to be used for pooling of studies. One of "Inverse" and "GLMM", can be abbreviated.

sm
A character string indicating which summary measure ("PFT", "PAS", "PRAW", "PLN", or "PLOGIT") is to be used for pooling of studies, see Details.

incr
A numeric which is added to event number and sample size of studies with zero or all events, i.e., studies with an event probability of either 0 or 1.

allincr
A logical indicating if incr is considered for all studies if at least one study has either zero or all events. If FALSE (default), incr is considered only in studies with zero or all events.

addincr
A logical indicating if incr is used for all studies irrespective of number of events.

method.ci
A character string indicating which method is used to calculate confidence intervals for individual studies, see Details.

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

hakn
A logical indicating whether the method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2, see Details.

tau.preset
Prespecified value for the square-root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance tau-squared.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

null.effect
A numeric value specifying the effect under the null hypothesis.

method.bias
A character string indicating which test is to be used. Either "rank", "linreg", or "mm", can be abbreviated. See function metabias.

backtransf
A logical indicating whether results for transformed proportions (argument sm!="PRAW") should be back transformed in printouts and plots. If TRUE (default), results will be presented as proportions; otherwise transformed proportions will be shown. See Details for presentation of confidence intervals.

pscale
A numeric defining a scaling factor for printing of single event probabilities.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

byvar
An optional vector containing grouping information (must be of same length as event).

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

warn
A logical indicating whether the addition of incr to studies with zero or all events should result in a warning.

control
An optional list to control the iterative process to estimate the between-study variance tau^2. This argument is passed on to rma.uni or rma.glmm, respectively.

...
Additional arguments passed on to rma.glmm function.

Details

Fixed effect and random effects meta-analysis of single proportions to calculate an overall proportion. The following transformations of proportions are implemented to calculate an overall proportion:

Logit transformation (sm="PLOGIT", default)

Log transformation (sm="PLN")

Freeman-Tukey Double arcsine transformation (sm="PFT")

Arcsine transformation (sm="PAS")

Raw, i.e. untransformed, proportions (sm="PRAW")

Note, you should use R function metabin to compare proportions of pairwise comparisons instead of using metaprop for each treatment arm separately which will break randomisation in randomised controlled trials.

Various methods are available to calculate confidence intervals for individual study results (see Agresti & Coull 1998; Newcombe 1988):

Clopper-Pearson interval also called 'exact' binomial interval (method.ci="CP", default)

Wilson Score interval (method.ci="WS")

Wilson Score interval with continuity correction (method.ci="WSCC")

Agresti-Coull interval (method.ci="AC")

Simple approximation interval (method.ci="SA")

Simple approximation interval with continuity correction (method.ci="SACC")

Normal approximation interval based on summary measure, i.e. defined by argument sm (method.ci="NAsm")

Note, with exception of the normal approximation based on the summary measure, i.e. method.ci="NAsm", the same confidence interval is calculated for any summary measure (argument sm) as only number of events and observations are used in the calculation disregarding the chosen summary measure. Results will be presented for transformed proportions if argument backtransf=FALSE in the print.meta, print.summary.meta, or forest.meta function. In this case, argument method.ci="NAsm" is used, i.e. confidence intervals based on the normal approximation based on the summary measure.

Argument pscale can be used to rescale proportions, e.g. pscale=1000 means that proportions are expressed as events per 1000 observations. This is useful in situations with (very) low event probabilities.

For several arguments defaults settings are utilised (assignments using gs function). These defaults can be changed using the settings.meta function.

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if argument comb.random=FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random=FALSE.

A distinctive and frequently overlooked advantage of binary data is that individual patient data (IPD) can be extracted. Accordingly, a random intercept logistic regression model can be utilised for the meta-analysis of proportions (Stijnen et al., 2010). This method is available (argument method = "GLMM") by calling the rma.glmm function from R package metafor internally.

If the summary measure is equal to "PRAW", "PLN", or "PLOGIT", a continuity correction is applied if any study has either zero or all events, i.e., an event probability of either 0 or 1. By default, 0.5 is used as continuity correction (argument incr). This continuity correction is used both to calculate individual study results with confidence limits and to conduct meta-analysis based on the inverse variance method. For GLMMs no continuity correction is used.

Argument byvar can be used to conduct subgroup analysis for all methods but GLMMs. Instead use the metareg function for GLMMs which can also be used for continuous covariates.

A prediction interval for treatment effect of a new study is calculated (Higgins et al., 2009) if arguments prediction and comb.random are TRUE.

R function update.meta can be used to redo the meta-analysis of an existing metaprop object by only specifying arguments which should be changed.

For the random effects, the method by Hartung and Knapp (2003) is used to adjust test statistics and confidence intervals if argument hakn=TRUE.

The DerSimonian-Laird estimate (1986) is used in the random effects model if method.tau="DL". The iterative Paule-Mandel method (1982) to estimate the between-study variance is used if argument method.tau="PM". Internally, R function paulemandel is called which is based on R function mpaule.default from R package metRology from S.L.R. Ellison <s.ellison at lgc.co.uk>.

If R package metafor (Viechtbauer 2010) is installed, the following methods to estimate the between-study variance τ^2 (argument method.tau) are also available:

Restricted maximum-likelihood estimator (method.tau="REML")

Maximum-likelihood estimator (method.tau="ML")

Hunter-Schmidt estimator (method.tau="HS")

Sidik-Jonkman estimator (method.tau="SJ")

Hedges estimator (method.tau="HE")

Empirical Bayes estimator (method.tau="EB").

For these methods the R function rma.uni of R package metafor is called internally. See help page of R function rma.uni for more details on these methods to estimate between-study variance.

Value

An object of class c("metaprop", "meta") with corresponding print, summary, and forest functions. The object is a list containing the following components:

event, n, studlab, exclude,

sm, incr, allincr, addincr, method.ci,

level, level.comb,
As defined above.

comb.fixed, comb.random,

hakn, method.tau, tau.preset, TE.tau, null.hypothesis,

method.bias, tau.common, title, complab, outclab,

byvar, bylab, print.byvar, byseparator, warn

TE, seTE
Estimated (un)transformed proportion and its standard error for individual studies.

lower, upper
Lower and upper confidence interval limits for individual studies.

zval, pval
z-value and p-value for test of treatment effect for individual studies.

w.fixed, w.random
Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed
Estimated overall (un)transformed proportion and standard error (fixed effect model).

lower.fixed, upper.fixed
Lower and upper confidence interval limits (fixed effect model).

zval.fixed, pval.fixed
z-value and p-value for test of overall effect (fixed effect model).

TE.random, seTE.random
Estimated overall (un)transformed proportion and standard error (random effects model).

lower.random, upper.random
Lower and upper confidence interval limits (random effects model).

zval.random, pval.random
z-value or t-value and corresponding p-value for test of overall effect (random effects model).

prediction, level.predict
As defined above.

seTE.predict
Standard error utilised for prediction interval.

lower.predict, upper.predict
Lower and upper limits of prediction interval.

k
Number of studies combined in meta-analysis.

Q
Heterogeneity statistic Q.

df.Q
Degrees of freedom for heterogeneity statistic.

pval.Q
P-value of heterogeneity test.

Q.LRT
Heterogeneity statistic for likelihood-ratio test (only if method = "GLMM").

df.Q.LRT
Degrees of freedom for likelihood-ratio test

pval.Q.LRT
P-value of likelihood-ratio test.

tau
Square-root of between-study variance.

se.tau
Standard error of square-root of between-study variance.

C
Scaling factor utilised internally to calculate common tau-squared across subgroups.

method
A character string indicating method used for pooling: "Inverse"

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn=TRUE).

bylevs
Levels of grouping variable - if byvar is not missing.

TE.fixed.w, seTE.fixed.w
Estimated treatment effect and standard error in subgroups (fixed effect model) - if byvar is not missing.

lower.fixed.w, upper.fixed.w
Lower and upper confidence interval limits in subgroups (fixed effect model) - if byvar is not missing.

zval.fixed.w, pval.fixed.w
z-value and p-value for test of treatment effect in subgroups (fixed effect model) - if byvar is not missing.

TE.random.w, seTE.random.w
Estimated treatment effect and standard error in subgroups (random effects model) - if byvar is not missing.

lower.random.w, upper.random.w
Lower and upper confidence interval limits in subgroups (random effects model) - if byvar is not missing.

zval.random.w, pval.random.w
z-value or t-value and corresponding p-value for test of treatment effect in subgroups (random effects model) - if byvar is not missing.

w.fixed.w, w.random.w
Weight of subgroups (in fixed and random effects model) - if byvar is not missing.

df.hakn.w
Degrees of freedom for test of treatment effect for Hartung-Knapp method in subgroups - if byvar is not missing and hakn=TRUE.

n.harmonic.mean.w
Harmonic mean of number of observations in subgroups (for back transformation of Freeman-Tukey Double arcsine transformation) - if byvar is not missing.

event.w
Number of events in subgroups - if byvar is not missing.

n.w
Number of observations in subgroups - if byvar is not missing.

k.w
Number of studies combined within subgroups - if byvar is not missing.

k.all.w
Number of all studies in subgroups - if byvar is not missing.

Q.w.fixed
Overall within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.w.random
Overall within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing (only calculated if argument tau.common is TRUE).

df.Q.w
Degrees of freedom for test of overall within subgroups heterogeneity - if byvar is not missing.

pval.Q.w.fixed
P-value of within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.w.random
P-value of within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

Q.b.fixed
Overall between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.b.random
Overall between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

df.Q.b
Degrees of freedom for test of overall between subgroups heterogeneity - if byvar is not missing.

pval.Q.b.fixed
P-value of between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.b.random
P-value of between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

tau.w
Square-root of between-study variance within subgroups - if byvar is not missing.

C.w
Scaling factor utilised internally to calculate common tau-squared across subgroups - if byvar is not missing.

H.w
Heterogeneity statistic H within subgroups - if byvar is not missing.

lower.H.w, upper.H.w
Lower and upper confidence limti for heterogeneity statistic H within subgroups - if byvar is not missing.

I2.w
Heterogeneity statistic I2 within subgroups - if byvar is not missing.

lower.I2.w, upper.I2.w
Lower and upper confidence limti for heterogeneity statistic I2 within subgroups - if byvar is not missing.

incr.event
Increment added to number of events.

keepdata
As defined above.

data
Original data (set) used in function call (if keepdata=TRUE).

subset
Information on subset of original data used in meta-analysis (if keepdata=TRUE).

.glmm.fixed
GLMM object generated by call of rma.glmm function (fixed effect model).

.glmm.random
GLMM object generated by call of rma.glmm function (random effects model).

call
Function call.

version
Version of R package meta used to create object.

version.metafor
Version of R package metafor used for GLMMs.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Agresti A & Coull BA (1998), Approximate is better than “exact” for interval estimation of binomial proportions. The American Statistician, 52, 119–126.

DerSimonian R & Laird N (1986), Meta-analysis in clinical trials. Controlled Clinical Trials, 7, 177–188.

Edward JM et al. (2006), Adherence to antiretroviral therapy in sub-saharan Africa and North America - a meta-analysis. Journal of the American Medical Association, 296, 679–690.

Freeman MF & Tukey JW (1950), Transformations related to the angular and the square root. Annals of Mathematical Statistics, 21, 607–611.

Higgins JPT, Thompson SG, Spiegelhalter DJ (2009), A re-evaluation of random-effects meta-analysis. Journal of the Royal Statistical Society: Series A, 172, 137–159.

Knapp G & Hartung J (2003), Improved Tests for a Random Effects Meta-regression with a Single Covariate. Statistics in Medicine, 22, 2693–2710, doi: 10.1002/sim.1482 .

Miller JJ (1978), The inverse of the Freeman-Tukey double arcsine transformation. The American Statistician, 32, 138.

Newcombe RG (1998), Two-sided confidence intervals for the single proportion: Comparison of seven methods. Statistics in Medicine, 17, 857–872.

Paule RC & Mandel J (1982), Consensus values and weighting factors. Journal of Research of the National Bureau of Standards, 87, 377–385.

Pettigrew HM, Gart JJ, Thomas DG (1986), The bias and higher cumulants of the logarithm of a binomial variate. Biometrika, 73, 425–435.

Stijnen T, Hamza TH, Ozdemir P (2010), Random effects meta-analysis of event outcome in the framework of the generalized linear mixed model with applications in sparse data. Statistics in Medicine, 29, 3046–67.

Viechtbauer W (2010), Conducting Meta-Analyses in R with the Metafor Package. Journal of Statistical Software, 36, 1–48.

See Also

update.meta, metacont, metagen, print.meta

Examples

# Apply various meta-analysis methods to estimate proportions
#
m1 <- metaprop(4:1, 10 * 1:4)
m2 <- update(m1, sm="PAS")
m3 <- update(m1, sm="PRAW")
m4 <- update(m1, sm="PLN")
m5 <- update(m1, sm="PFT")
#
m1
m2
m3
m4
m5
#
forest(m1)
# forest(m2)
# forest(m3)
# forest(m3, pscale=100)
# forest(m4)
# forest(m5)

# Do not back transform results, e.g. print logit transformed
# proportions if sm="PLOGIT" and store old settings
#
oldset <- settings.meta(backtransf=FALSE)
#
m6  <- metaprop(4:1, c(10, 20, 30, 40))
m7  <- update(m6, sm="PAS")
m8  <- update(m6, sm="PRAW")
m9  <- update(m6, sm="PLN")
m10 <- update(m6, sm="PFT")
#
forest(m6)
# forest(m7)
# forest(m8)
# forest(m8, pscale=100)
# forest(m9)
# forest(m10)
#
# Use old settings
#
settings.meta(oldset)

# Examples with zero events
#
m1 <- metaprop(c(0, 0, 10, 10), rep(100, 4))
m2 <- metaprop(c(0, 0, 10, 10), rep(100, 4), incr=0.1)
#
summary(m1)
summary(m2)
#
# forest(m1)
# forest(m2)

# Example from Miller (1978):
#
death <- c(3, 6, 10, 1)
animals <- c(11, 17, 21, 6)
#
m3 <- metaprop(death, animals, sm="PFT")
forest(m3)

# Data examples from Newcombe (1998)
# - apply various methods to estimate confidence intervals for
#   individual studies
#
event <- c(81, 15, 0, 1)
n <- c(263, 148, 20, 29)
#
m1 <- metaprop(event, n, sm="PLOGIT", method.ci="SA")
m2 <- update(m1, method.ci="SACC")
m3 <- update(m1, method.ci="WS")
m4 <- update(m1, method.ci="WSCC")
m5 <- update(m1, method.ci="CP")
#
lower <- round(rbind(NA, m1$lower, m2$lower, NA, m3$lower, m4$lower, NA, m5$lower), 4)
upper <- round(rbind(NA, m1$upper, m2$upper, NA, m3$upper, m4$upper, NA, m5$upper), 4)
#
tab1 <- data.frame(
  scen1=meta:::formatCI(lower[,1], upper[,1]),
  scen2=meta:::formatCI(lower[,2], upper[,2]),
  scen3=meta:::formatCI(lower[,3], upper[,3]),
  scen4=meta:::formatCI(lower[,4], upper[,4]),
  stringsAsFactors=FALSE
  )
names(tab1) <- c("r=81, n=263", "r=15, n=148", "r=0, n=20", "r=1, n=29")
row.names(tab1) <- c("Simple", "- SA", "- SACC",
                     "Score", "- WS", "- WSCC",
                     "Binomial", "- CP")
tab1[is.na(tab1)] <- ""
#
# Newcombe (1998), Table I, methods 1-5:
#
tab1
#
# Same confidence interval, i.e. unaffected by choice of summary measure
#
print(metaprop(event, n, sm="PLOGIT", method.ci="WS"), ma=FALSE)
print(metaprop(event, n, sm="PLN", method.ci="WS"), ma=FALSE)
print(metaprop(event, n, sm="PFT", method.ci="WS"), ma=FALSE)
print(metaprop(event, n, sm="PAS", method.ci="WS"), ma=FALSE)
print(metaprop(event, n, sm="PRAW", method.ci="WS"), ma=FALSE)
#
# Different confidence intervals as argument sm="NAsm"
#
print(metaprop(event, n, sm="PLOGIT", method.ci="NAsm"), ma=FALSE)
print(metaprop(event, n, sm="PLN", method.ci="NAsm"), ma=FALSE)
print(metaprop(event, n, sm="PFT", method.ci="NAsm"), ma=FALSE)
print(metaprop(event, n, sm="PAS", method.ci="NAsm"), ma=FALSE)
print(metaprop(event, n, sm="PRAW", method.ci="NAsm"), ma=FALSE)
#
# Different confidence intervals as argument backtransf=FALSE.
# Accordingly, method.ci="NAsm" used internally.
#
print(metaprop(event, n, sm="PLOGIT", method.ci="WS"), ma=FALSE, backtransf=FALSE)
print(metaprop(event, n, sm="PLN", method.ci="WS"), ma=FALSE, backtransf=FALSE)
print(metaprop(event, n, sm="PFT", method.ci="WS"), ma=FALSE, backtransf=FALSE)
print(metaprop(event, n, sm="PAS", method.ci="WS"), ma=FALSE, backtransf=FALSE)
print(metaprop(event, n, sm="PRAW", method.ci="WS"), ma=FALSE, backtransf=FALSE)
#
# Same results (printed on original and log scale, respectively)
#
print(metaprop(event, n, sm="PLN", method.ci="NAsm"), ma=FALSE)
print(metaprop(event, n, sm="PLN"), ma=FALSE, backtransf=FALSE)
# Results for first study (on log scale)
round(log(c(0.3079848, 0.2569522, 0.3691529)), 4)

# Meta-analysis using generalised linear mixed models
# (R packages 'metafor' and 'lme4' must be available)
#
# metaprop(event, n, method = "GLMM")

#
# Print results as events per 1000 observations
#
print(metaprop(6:8, c(100, 1200, 1000)), pscale = 1000, digits = 1)

[Package meta version 4.9-4 Index]