or2smd {meta}	R Documentation

Conversion from log odds ratio to standardised mean difference

Description

Conversion from log odds ratio to standardised mean difference using method by Hasselblad & Hedges (1995) or Cox (1970).

Usage

or2smd(
  lnOR,
  selnOR,
  studlab,
  data = NULL,
  subset = NULL,
  exclude = NULL,
  method = "HH",
  ...
)

Arguments

lnOR
Log odds ratio(s) or meta-analysis object.

selnOR
Standard error(s) of log odds ratio(s) (ignored if argument lnOR is a meta-analysis object).

studlab
An optional vector with study labels (ignored if argument lnOR is a meta-analysis object).

data
An optional data frame containing the study information (ignored if argument lnOR is a meta-analysis object).

subset
An optional vector specifying a subset of studies to be used (ignored if argument lnOR is a meta-analysis object).

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots (ignored if argument lnOR is a meta-analysis object).

method
A character string indicating which method is used to convert log odds ratios to standardised mean differences. Either "HH" or "CS", can be abbreviated.

...
Additional arguments passed on to metagen (ignored if argument lnOR is a meta-analysis object).

Details

This function implements the following methods for the conversion from log odds ratios to standardised mean difference:

Hasselblad & Hedges (1995) assuming logistic distributions (method == "HH")

Cox (1970) and Cox & Snell (1989) assuming normal distributions (method == "CS")

Internally, metagen is used to conduct a meta-analysis with the standardised mean difference as summary measure.

Argument lnOR can be either a vector of log odds ratios or a meta-analysis object created with metabin or metagen and the odds ratio as summary measure.

Argument selnOR is mandatory if argument lnOR is a vector and ignored otherwise. Additional arguments in ... are only passed on to metagen if argument lnOR is a vector.

Value

An object of class "meta" and "metagen"; see metagen.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Borenstein M, Hedges LV, Higgins JPT, Rothstein HR (2009): Introduction to Meta-Analysis. Chichester: Wiley

Cox DR (1970): Analysis of Binary Data. London: Chapman and Hall / CRC

Cox DR, Snell EJ (1989): Analysis of Binary Data (2nd edition). London: Chapman and Hall / CRC

Hasselblad V, Hedges LV (1995): Meta-analysis of screening and diagnostic tests. Psychological Bulletin, 117, 167–78

See Also

smd2or, metabin, metagen, metacont

Examples

# Example from Borenstein et al. (2009), Chapter 7
#
mb <- or2smd(0.9069, sqrt(0.0676))
# TE = standardised mean difference (SMD); seTE = standard error of SMD
data.frame(SMD = round(mb$TE, 4), varSMD = round(mb$seTE^2, 4))

# Use dataset from Fleiss (1993)
#
data(Fleiss1993bin)
m1 <- metabin(d.asp, n.asp, d.plac, n.plac,
              data = Fleiss1993bin,
              studlab = paste(study, year),
              sm = "OR", comb.random = FALSE)
or2smd(m1)

[Package meta version 4.18-1 Index]