plot.meta {meta}	R Documentation

Plot function for objects of class meta

Description

Draws a forest plot in the active graphics window.

Usage

## S3 method for class 'meta':
plot(x, byvar=x$byvar, bylab=x$bylab,
     print.byvar=x$print.byvar, 
     sortvar, studlab=TRUE, level=x$level, level.comb=x$level.comb,
     comb.fixed=x$comb.fixed, comb.random=x$comb.random, overall=TRUE,
     text.fixed="Fixed effect model", text.random="Random effects model",
     lty.fixed=2, lty.random=3, xlab=NULL, xlim, ylim, lwd=1, cex=1,
     cex.comb=1.2 * cex, cex.axis=cex, cex.lab=cex,
     log=ifelse(x$sm %in% c("RR", "OR", "HR"), "x", ""),
     axes=TRUE, allstudies=TRUE,
     weight=ifelse(comb.random, "random", "fixed"), scale.diamond=1,
     scale.square= 1, col.i="black",
     clim=xlim, arrow.length=0.1,
     ref=ifelse(x$sm %in% c("RR", "OR", "HR"), 1, 0),
     ...)

Arguments

x 	 An object of class meta.

byvar 	 An optional vector containing grouping information (must be of same length as x$TE).

bylab 	 A character string with a label for the grouping variable.

print.byvar 	 A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

sortvar 	 An optional vector used to sort the individual studies (must be of same length as x$TE).

studlab 	 A logical indicating whether study labels should be printed in the graph. A vector with study labels can also be provided (must be of same length as x$TE then).

level 	 The level used to calculate confidence intervals for individual studies.

level.comb 	 The level used to calculate confidence intervals for pooled estimates.

comb.fixed 	 A logical indicating whether fixed effect estimate should be plotted.

comb.random 	 A logical indicating whether random effects estimate should be plotted.

overall 	 A logical indicating whether overall summaries should be plotted. This parameter is useful in combination with the parameter byvar if summaries should only be plotted on group level.

text.fixed 	 A character string used in the plot to label the pooled fixed effects estimate.

text.random 	 A character string used in the plot to label the pooled random effects estimate.

lty.fixed 	 Line type (pooled fixed effect estimate).

lty.random 	 Line type (pooled random effects estimate).

xlab 	 A label for the x axis.

xlim 	 The x limits (min,max) of the plot.

ylim 	 The y limits (min,max) of the plot.

lwd 	 The line width.

cex 	 A numerical value giving the amount by which plotting text and symbols should be scaled relative to the default.

cex.comb 	 A numerical value giving the amount by which plotting text and symbols for pooled fixed and random effects estimates should be scaled.

cex.axis 	 The magnification to be used for axis annotation relative to the current setting of cex.

cex.lab 	 The magnification to be used for x and y labels relative to the current setting of cex.

log 	 A character string which contains "x" if the x axis is to be logarithmic (other values for log are not reasonable).

axes 	 A logical indicating whether the x axis should be drawn on the plot.

allstudies 	 A logical indicating whether studies with inestimable treatment effects should be plotted.

weight 	 A character string indicating which type of plotting symbols is to be used for individual treatment estimates. One of "same", "fixed", or "random", can be abbreviated. Plot symbols have the same size for all studies or represent study weights from fixed effect or random effects model.

scale.diamond 	 A numerical value giving the amount by which the diamond representing pooled treatment effects should be scaled relative to the default.

scale.square 	 A numerical value giving the amount by which the square representing treatment effects in individual studies should be scaled relative to the default.

... 	 Graphical parameters as in par may also be passed as arguments.

col.i 	 The colour for individual study results and confidence limits.

clim 	 Limits (min,max) where to cut confidence limits; arrows are plotted if confidence limits are outside the range of clim.

arrow.length 	 Length of the edges of the arrow head (in inches) which is plotted if confidence limits are outside the range of clim. See also function arrows.

ref 	 A numerical value defining a reference value which is plotted as a vertical line.

Details

A forest plot, also called confidence interval plot, is drawn in the active graphics window. Sub-group analyses are conducted and displayed in the plot if byvar is not missing.

The plot.meta function produces basic forest plots. For nicer looking forest plots the forest function can be used.

Review Manager 5 (RevMan 5) is the current software used for preparing and maintaining Cochrane Reviews (http://www.cc-ims.net/revman/). In RevMan 5, subgroup analyses can be defined and data from a Cochrane review can be imported to R using the function read.rm5. If a meta-analysis is then conducted using function metacr, information on subgroups is available in R (components byvar, bylab, and print.byvar, byvar in an object of class "meta"). Accordingly, by using function metacr there is no need to define subgroups in order to redo the statistical analysis conducted in the Cochrane review.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

See Also

forest, metabin, metacont, metagen

Examples

data(Olkin95)
meta1 <- metabin(event.e, n.e, event.c, n.c,
                 data=Olkin95, subset=c(41,47,51,59),
                 sm="RR", meth="I")

oldpar <- par(mfrow=c(2, 2))

plot(meta1)
plot(meta1, byvar=c(1,2,1,2), bylab="label")
plot(meta1, byvar=1:4, xlim=c(0.02, 10))

par(oldpar)

[Package meta version 1.1-8 Index]