print.meta {meta}	R Documentation

Print meta-analysis results

Description

Print method for objects of class meta.

Usage

## S3 method for class 'meta'
print(
  x,
  sortvar,
  comb.fixed = x$comb.fixed,
  comb.random = x$comb.random,
  prediction = x$prediction,
  details = FALSE,
  ma = TRUE,
  overall = x$overall,
  overall.hetstat = x$overall.hetstat,
  backtransf = x$backtransf,
  pscale = x$pscale,
  irscale = x$irscale,
  irunit = x$irunit,
  digits = gs("digits"),
  digits.se = gs("digits.se"),
  digits.pval = max(gs("digits.pval"), 2),
  digits.tau2 = gs("digits.tau2"),
  digits.tau = gs("digits.tau"),
  digits.I2 = gs("digits.I2"),
  digits.prop = gs("digits.prop"),
  digits.weight = gs("digits.weight"),
  scientific.pval = gs("scientific.pval"),
  big.mark = gs("big.mark"),
  zero.pval = gs("zero.pval"),
  JAMA.pval = gs("JAMA.pval"),
  text.tau2 = gs("text.tau2"),
  text.tau = gs("text.tau"),
  text.I2 = gs("text.I2"),
  warn.backtransf = FALSE,
  ...
)

cilayout(bracket = "[", separator = "; ")

Arguments

x
An object of class meta

sortvar
An optional vector used to sort the individual studies (must be of same length as x$TE).

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

prediction
A logical indicating whether a prediction interval should be printed.

details
A logical indicating whether further details of individual studies should be printed.

ma
A logical indicating whether the summary results of the meta-analysis should be printed.

overall
A logical indicating whether overall summaries should be reported. This argument is useful in a meta-analysis with subgroups if overall results should not be reported.

overall.hetstat
A logical value indicating whether to print heterogeneity measures for overall treatment comparisons. This argument is useful in a meta-analysis with subgroups if heterogeneity statistics should only be printed on subgroup level.

backtransf
A logical indicating whether printed results should be back transformed. If backtransf = TRUE, results for sm = "OR" are printed as odds ratios rather than log odds ratios and results for sm = "ZCOR" are printed as correlations rather than Fisher's z transformed correlations, for example.

pscale
A numeric giving scaling factor for printing of single event probabilities or risk differences, i.e. if argument sm is equal to "PLOGIT", "PLN", "PRAW", "PAS", "PFT", or "RD".

irscale
A numeric defining a scaling factor for printing of single incidence rates or incidence rate differences, i.e. if argument sm is equal to "IR", "IRLN", "IRS", "IRFT", or "IRD".

irunit
A character specifying the time unit used to calculate rates, e.g. person-years.

digits
Minimal number of significant digits, see print.default.

digits.se
Minimal number of significant digits for standard deviations and standard errors, see print.default.

digits.pval
Minimal number of significant digits for p-value of test of treatment effect, see print.default.

digits.tau2
Minimal number of significant digits for between-study variance, see print.default.

digits.tau
Minimal number of significant digits for square root of between-study variance, see print.default.

digits.I2
Minimal number of significant digits for I-squared and Rb statistic, see print.default.

digits.prop
Minimal number of significant digits for proportions, see print.default.

digits.weight
Minimal number of significant digits for weights, see print.default.

scientific.pval
A logical specifying whether p-values should be printed in scientific notation, e.g., 1.2345e-01 instead of 0.12345.

big.mark
A character used as thousands separator.

zero.pval
A logical specifying whether p-values should be printed with a leading zero.

JAMA.pval
A logical specifying whether p-values for test of overall effect should be printed according to JAMA reporting standards.

text.tau2
Text printed to identify between-study variance τ^2.

text.tau
Text printed to identify τ, the square root of the between-study variance τ^2.

text.I2
Text printed to identify heterogeneity statistic I^2.

warn.backtransf
A logical indicating whether a warning should be printed if backtransformed proportions and rates are below 0 and backtransformed proportions are above 1.

...
Additional arguments (passed on to print.summary.meta called internally).

bracket
A character with bracket symbol to print lower confidence interval: "[", "(", "{", "".

separator
A character string with information on separator between lower and upper confidence interval.

Details

R function cilayout can be utilised to change the layout to print confidence intervals (both in printout from print.meta and print.summary.meta function as well as in forest plots). The default layout is "[lower; upper]". Another popular layout is "(lower - upper)" which is used throughout an R session by using R command cilayout("(", " - ").

Argument pscale can be used to rescale single proportions or risk differences, e.g. pscale = 1000 means that proportions are expressed as events per 1000 observations. This is useful in situations with (very) low event probabilities.

Argument irscale can be used to rescale single rates or rate differences, e.g. irscale = 1000 means that rates are expressed as events per 1000 time units, e.g. person-years. This is useful in situations with (very) low rates. Argument irunit can be used to specify the time unit used in individual studies (default: "person-years"). This information is printed in summaries and forest plots if argument irscale is not equal to 1.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994), The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation.

Crippa A, Khudyakov P, Wang M, Orsini N, Spiegelman D (2016), A new measure of between-studies heterogeneity in meta-analysis. Statistics in Medicine, 35, 3661–75.

Higgins JPT & Thompson SG (2002), Quantifying heterogeneity in a meta-analysis. Statistics in Medicine, 21, 1539–58.

See Also

summary.meta, update.meta, metabin, metacont, metagen

Examples

data(Fleiss1993cont)
m1 <- metacont(n.psyc, mean.psyc, sd.psyc, n.cont, mean.cont, sd.cont,
               data = Fleiss1993cont, sm = "SMD",
               studlab = paste(study, year))
m1

print(m1, digits = 2)

## Not run: 
# Use unicode characters to print tau^2, tau, and I^2 
print(m1,
      text.tau2 = "\u03c4\u00b2", text.tau = "\u03c4", text.I2 = "I\u00b2")

## End(Not run)

[Package meta version 4.18-1 Index]