print.rm5 {meta}	R Documentation

Print and summary methods for objects of class rm5

Description

Print and summary methods for objects of class rm5.

Usage

## S3 method for class 'rm5'
print(x, ...)

## S3 method for class 'summary.rm5'
print(x, ...)

## S3 method for class 'rm5'
summary(object, comp.no, outcome.no, ...)

## S3 method for class 'rm5'
metabias(x, comp.no, outcome.no,
         method.bias="linreg",
         method.bias.binary=method.bias,
         method.bias.or="score",
         k.min=10, ...)

Arguments

x
An object of class rm5.

object
An object of class rm5.

comp.no
Comparison number.

outcome.no
Outcome number.

method.bias
A character string indicating which test for small-study effects is to be used for all outcomes. Either "rank", "linreg", or "mm", can be abbreviated. See function metabias

method.bias.binary
A character string indicating which test is to be used for binary outcomes. Either "rank", "linreg", "mm", "count", "score", or "peters", can be abbreviated. See function metabias

method.bias.or
A character string indicating which test is to be used for binary outcomes with odds ratio as summary measure. Either "rank", "linreg", "mm", "count", "score", or "peters", can be abbreviated. See function metabias

k.min
Minimum number of studies to perform test for small-study effects.

...
Additional arguments (ignored at the moment)

Details

Review Manager 5 (RevMan 5) is the current software used for preparing and maintaining Cochrane Reviews (http://community.cochrane.org/tools/review-production-tools/revman-5). In RevMan 5, subgroup analyses can be defined and data from a Cochrane review can be imported to R using the function read.rm5.

The R function summary.rm5 can be used to redo all meta-analyses of the imported Cochrane Review.

The R function metabias.rm5 can be used to conduct a test for funnel plot asymmetry for all meta-analyses of the imported Cochrane Review.

The R function metacr is called internally.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Higgins, J.P.T and S. Green (2011), Cochrane Handbook for Systematic Reviews of Interventions Version 5.1.0 [Updated March 2011]. The Cochrane Library: http://www.cochrane-handbook.org

See Also

metabias.meta, summary.meta, read.rm5

Examples

# Locate export data file "Fleiss93_CR.csv"
# in sub-directory of package "meta"
#
filename <- system.file("data/Fleiss93_CR.csv.gz", package = "meta")
#
Fleiss93_CR <- read.rm5(filename)

#
# Print summary results for all meta-analysis:
#
summary(Fleiss93_CR)

#
# Print results for tests of small-study effects:
#
metabias(Fleiss93_CR, k.min=5)

[Package meta version 4.9-4 Index]