radial.meta {meta}	R Documentation

Radial plot

Description

Draw a radial plot (also called Galbraith plot) which can be used to assess bias in meta-analysis.

Usage

## S3 method for class 'meta'
radial(
  x,
  xlim = NULL,
  ylim = NULL,
  xlab = "Inverse of standard error",
  ylab = "Standardised treatment effect (z-score)",
  comb.fixed = TRUE,
  axes = TRUE,
  pch = 1,
  text = NULL,
  cex = 1,
  col = NULL,
  level = NULL,
  ...
)

## Default S3 method:
radial(
  x,
  y,
  xlim = NULL,
  ylim = NULL,
  xlab = "Inverse of standard error",
  ylab = "Standardised treatment effect (z-score)",
  comb.fixed = TRUE,
  axes = TRUE,
  pch = 1,
  text = NULL,
  cex = 1,
  col = NULL,
  level = NULL,
  ...
)

Arguments

x
An object of class meta, or estimated treatment effect in individual studies.

xlim
The x limits (min, max) of the plot.

ylim
The y limits (min, max) of the plot.

xlab
A label for the x-axis.

ylab
A label for the y-axis.

comb.fixed
A logical indicating whether the pooled fixed effect estimate should be plotted.

axes
A logical indicating whether axes should be drawn on the plot.

pch
The plotting symbol used for individual studies.

text
A character vector specifying the text to be used instead of plotting symbol.

cex
The magnification to be used for plotting symbol.

col
A vector with colour of plotting symbols.

level
The confidence level utilised in the plot.

...
Graphical arguments as in par may also be passed as arguments.

y
Standard error of estimated treatment effect.

Details

A radial plot (Galbraith 1988a,b), also called Galbraith plot, is drawn in the active graphics window. If comb.fixed is TRUE, the pooled estimate of the fixed effect model is plotted. If level is not NULL, the corresponding confidence limits are drawn.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Galbraith RF (1988a): Graphical display of estimates having differing standard errors. Technometrics, 30, 271–81

Galbraith RF (1988b): A note on graphical presentation of estimated odds ratios from several clinical trials. Statistics in Medicine, 7, 889–94

See Also

metabias, metabin, metagen, funnel

Examples

data(Olkin1995)
m1 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
              data = Olkin1995, subset = c(41, 47, 51, 59),
              studlab = paste(author, year),
              sm = "RR", method = "I")

# Radial plot
#
radial(m1, level = 0.95)

[Package meta version 4.18-1 Index]