read.rm5 {meta}	R Documentation

Import RevMan 5 data files (.csv)

Description

Reads data file from Cochrane Intervention review created with RevMan 5 and creates a data frame from it.

Usage

read.rm5(file, sep = ",", quote = "\"", title,
  numbers.in.labels = TRUE)

## S3 method for class 'rm5'
print(x, ...)

Arguments

file
The name of a file to read data values from.

sep
The field separator character. Values on each line of the file are separated by this character. The comma is the default field separator character in RevMan 5.

quote
The set of quoting characters. In RevMan 5 a "\"" is the default quoting character.

title
Title of Cochrane review.

numbers.in.labels
A logical indicating whether comparision number and outcome number should be printed at the beginning of the comparison (argument complab) and outcome label (argument outclab); this is the default in RevMan 5.

x
An object of class rm5

...
Additional arguments (passed on to print.data.frame).

Details

Review Manager 5 (RevMan 5) is the current software used for preparing and maintaining Cochrane Reviews (http://community.cochrane.org/tools/review-production-tools/revman-5). RevMan 5 includes the ability to write Systematic reviews of interventions, Diagnostic test accuracy reviews, Methodology reviews and Overviews of reviews.

This function provides the ability to read a data file from a Cochrane Intervention review created with RevMan 5; a data frame is created from it. Cochrane Intervention reviews are based on the comparison of two interventions.

In order to generate a data analysis file in RevMan 5 use the following Menu points: "File" - "Export" - "Data and analyses". It is mandatory to include the following fields in the exported data file by selecting them with the mouse cursor in the Export Analysis Data Wizard: (i) Comparison Number, (ii) Outcome Number, (iii) Subgroup Number. When these fields are not selected a corresponding error message will be printed in R. It is recommended to include all fields in the exported data file except for the last field "Risk of bias tables". For example, in order to redo the meta-analysis in R for the RevMan 5 data type "O-E and Variance" the fields "O-E" and "Variance" have to be selected in the Export Analysis Data Wizard. If the last field "Risk of bias tables" is selected the import in R fails with an error message "line X did not have Y elements".

By default in RevMan 5, the name of the exported data file is the title of the Cochrane Review. Accordingly, information on the title is extracted from the name of the exported data file (argument: file) if argument title is missing (default).

Each respective meta-analysis for arguments event.e.pooled – df.pooled is defined by values for "comp.no" and "outcome.no", and "grp.no".

Value

A data frame containing the following components:

comp.no
Comparison number.

outcome.no
Outcome number.

group.no
Group number.

studlab
Study label.

year
Year of publication.

event.e
Number of events in experimental group.

n.e
Number of observations in experimental group.

event.c
Number of events in control group.

n.c
Number of observations in control group.

mean.e
Estimated mean in experimental group.

sd.e
Standard deviation in experimental group.

mean.c
Estimated mean in control group.

sd.c
Standard deviation in control group.

O.E
Observed minus expected (IPD analysis).

V
Variance of O.E (IPD analysis).

TE, seTE
Estimated treatment effect and standard error of individual studies.

lower, upper
Lower and upper limit of 95% confidence interval for treatment effect in individual studies.

weight
Weight of individual studies (according to meta-analytical method used in respective meta-analysis - see below for details).

order
Ordering of studies.

grplab
Group label.

type
Type of outcome. D = dichotomous, C = continuous, P = IPD.

method
A character string indicating which method has been used for pooling of studies. One of "Inverse", "MH", or "Peto".

sm
A character string indicating which summary measure has been used for pooling of studies.

model
A character string indicating which meta-analytical model has been used (either "Fixed" or "Random").

comb.fixed
A logical indicating whether fixed effect meta-analysis has been used in respective meta-analysis (see below for details).

comb.random
A logical indicating whether random effects meta-analysis has been used in respective meta-analysis (see below for details).

outclab
Outcome label.

k
Total number of studies combined in respective meta-analysis).

event.e.pooled
Number of events in experimental group in respective meta-analysis (see below for details).

n.e.pooled
Number of observations in experimental group in respective meta-analysis (see below for details).

event.c.pooled
Number of events in control group in respective meta-analysis (see below for details).

n.c.pooled
Number of observations in control group in respective meta-analysis (see below for details).

TE.pooled
Estimated treatment effect in respective meta-analysis (see below for details).

lower, upper
Lower and upper limit of 95% confidence interval for treatment effect in respective meta-analysis (see below for details).

weight.pooled
Total weight in respective meta-analysis (see below for details).

Z.pooled
Z-score for test of overall treatment effect in respective meta-analysis (see below for details).

pval.pooled
P-value for test of overall treatment effect in respective meta-analysis (see below for details).

Q
Heterogeneity statistic Q in respective meta-analysis (see below for details).

pval.Q
P-value of heterogeneity statistic Q in respective meta-analysis (see below for details).

I2
Heterogeneity statistic I2 in respective meta-analysis (see below for details).

tau2
Between-study variance (moment estimator of DerSimonian-Laird) in respective meta-analysis.

Q.w
Heterogeneity statistic Q within groups in respective meta-analysis (see below for details).

pval.Q.w
P-value of heterogeneity statistic Q within groups in respective meta-analysis (see below for details).

I2.w
Heterogeneity statistic I2 within groups in respective meta-analysis (see below for details).

label.e
Label for experimental group.

label.c
Label for control group.

label.left
Graph label on left side of forest plot.

label.right
Graph label on right side of forest plot.

RR.cochrane
A logical indicating if 2*incr instead of 1*incr is to be added to n.e and n.c in the calculation of the risk ratio (i.e., sm = "RR") for studies with a zero cell. This is used in RevMan 5.

complab
Comparison label.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Review Manager (RevMan) [Computer program]. Version 5.3. Copenhagen: The Nordic Cochrane Centre, The Cochrane Collaboration, 2014

See Also

summary.rm5, metabias.rm5, metabin, metacont, metagen, metacr

Examples

# Locate export data file "Fleiss93_CR.csv"
# in sub-directory of package "meta"
#
filename <- system.file("extdata", "Fleiss93_CR.csv", package = "meta")
Fleiss93_CR <- read.rm5(filename)

# Same result as R command example(Fleiss93):
#
metacr(Fleiss93_CR)

# Same result as R command example(Fleiss93cont):
#
metacr(Fleiss93_CR, 1, 2)

[Package meta version 4.9-6 Index]