smd2or {meta}	R Documentation

Conversion from standardised mean difference to log odds ratio

Description

Conversion from standardised mean difference to log odds ratio using method by Hasselblad & Hedges (1995) or Cox (1970).

Usage

smd2or(
  smd,
  se.smd,
  studlab,
  data = NULL,
  subset = NULL,
  exclude = NULL,
  method = "HH",
  backtransf = gs("backtransf"),
  ...
)

Arguments

smd
Standardised mean difference(s) (SMD) or meta-analysis object.

se.smd
Standard error(s) of SMD (ignored if argument smd is a meta-analysis object).

studlab
An optional vector with study labels (ignored if argument smd is a meta-analysis object).

data
An optional data frame containing the study information (ignored if argument smd is a meta-analysis object).

subset
An optional vector specifying a subset of studies to be used (ignored if argument smd is a meta-analysis object).

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots (ignored if argument smd is a meta-analysis object).

method
A character string indicating which method is used to convert SMDs to log odds ratios. Either "HH" or "CS", can be abbreviated.

backtransf
A logical indicating whether odds ratios (if TRUE) or log odds ratios (if FALSE) should be shown in printouts and plots.

...
Additional arguments passed on to metagen (ignored if argument smd is a meta-analysis object).

Details

This function implements the following methods for the conversion from standardised mean difference to log odds ratio:

Hasselblad & Hedges (1995) assuming logistic distributions (method == "HH")

Cox (1970) and Cox & Snell (1989) assuming normal distributions (method == "CS")

Internally, metagen is used to conduct a meta-analysis with the odds ratio as summary measure.

Argument smd can be either a vector of standardised mean differences or a meta-analysis object created with metacont or metagen and the standardised mean difference as summary measure.

Argument se.smd is mandatory if argument smd is a vector and ignored otherwise. Additional arguments in ... are only passed on to metagen if argument smd is a vector.

Value

An object of class "meta" and "metagen"; see metagen.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Borenstein M, Hedges LV, Higgins JPT, Rothstein HR (2009): Introduction to Meta-Analysis. Chichester: Wiley

Cox DR (1970): Analysis of Binary Data. London: Chapman and Hall / CRC

Cox DR, Snell EJ (1989): Analysis of Binary Data (2nd edition). London: Chapman and Hall / CRC

Hasselblad V, Hedges LV (1995): Meta-analysis of screening and diagnostic tests. Psychological Bulletin, 117, 167–78

See Also

or2smd, metacont, metagen, metabin

Examples

# Example from Borenstein et al. (2009), Chapter 7
#
mb <- smd2or(0.5, sqrt(0.0205), backtransf = FALSE)
# TE = log odds ratio; seTE = standard error of log odds ratio
data.frame(lnOR = round(mb$TE, 4), varlnOR = round(mb$seTE^2, 4))

# Use dataset from Fleiss (1993)
#
data(Fleiss1993cont)
m1 <- metacont(n.psyc, mean.psyc, sd.psyc, n.cont, mean.cont, sd.cont,
               data = Fleiss1993cont, sm = "SMD",
               studlab = paste(study, year))
smd2or(m1)

[Package meta version 4.18-1 Index]