summary.meta {meta}	R Documentation

Summary of meta-analysis results

Description

Summary method for objects of class meta.

Usage

## S3 method for class 'meta'
summary(
  object,
  comb.fixed = object$comb.fixed,
  comb.random = object$comb.random,
  prediction = object$prediction,
  overall = object$overall,
  overall.hetstat = object$overall.hetstat,
  backtransf = object$backtransf,
  pscale = object$pscale,
  irscale = object$irscale,
  irunit = object$irunit,
  bylab = object$bylab,
  print.byvar = object$print.byvar,
  byseparator = object$byseparator,
  bystud = FALSE,
  print.CMH = object$print.CMH,
  warn = object$warn,
  ...
)

## S3 method for class 'summary.meta'
print(
  x,
  comb.fixed = x$comb.fixed,
  comb.random = x$comb.random,
  prediction = x$prediction,
  overall = x$overall,
  overall.hetstat = x$overall.hetstat,
  print.byvar = x$print.byvar,
  byseparator = x$byseparator,
  print.CMH = x$print.CMH,
  header = TRUE,
  backtransf = x$backtransf,
  pscale = x$pscale,
  irscale = x$irscale,
  irunit = x$irunit,
  bylab.nchar = 35,
  digits = gs("digits"),
  digits.stat = gs("digits.stat"),
  digits.pval = max(gs("digits.pval"), 2),
  digits.pval.Q = max(gs("digits.pval.Q"), 2),
  digits.Q = gs("digits.Q"),
  digits.tau2 = gs("digits.tau2"),
  digits.tau = gs("digits.tau"),
  digits.H = gs("digits.H"),
  digits.I2 = gs("digits.I2"),
  scientific.pval = gs("scientific.pval"),
  big.mark = gs("big.mark"),
  zero.pval = gs("zero.pval"),
  JAMA.pval = gs("JAMA.pval"),
  print.I2 = gs("print.I2"),
  print.H = gs("print.H"),
  print.Rb = gs("print.Rb"),
  text.tau2 = gs("text.tau2"),
  text.tau = gs("text.tau"),
  text.I2 = gs("text.I2"),
  text.Rb = gs("text.Rb"),
  digits.zval = digits.stat,
  warn.backtransf = FALSE,
  ...
)

Arguments

object
An object of class meta.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

prediction
A logical indicating whether a prediction interval should be printed.

overall
A logical indicating whether overall summaries should be reported. This argument is useful in a meta-analysis with subgroups if overall results should not be reported.

overall.hetstat
A logical value indicating whether to print heterogeneity measures for overall treatment comparisons. This argument is useful in a meta-analysis with subgroups if heterogeneity statistics should only be printed on subgroup level.

backtransf
A logical indicating whether printed results should be back transformed. If backtransf=TRUE, results for sm="OR" are printed as odds ratios rather than log odds ratios and results for sm="ZCOR" are printed as correlations rather than Fisher's z transformed correlations, for example.

pscale
A numeric giving scaling factor for printing of single event probabilities or risk differences, i.e. if argument sm is equal to "PLOGIT", "PLN", "PRAW", "PAS", "PFT", or "RD".

irscale
A numeric defining a scaling factor for printing of single incidence rates or incidence rate differences, i.e. if argument sm is equal to "IR", "IRLN", "IRS", "IRFT", or "IRD".

irunit
A character specifying the time unit used to calculate rates, e.g. person-years.

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

bystud
A logical indicating whether results of individual studies should be printed by grouping variable.

print.CMH
A logical indicating whether result of the Cochran-Mantel-Haenszel test for overall effect should be printed.

warn
A logical indicating whether the use of summary.meta in connection with metacum or metainf should result in a warning.

...
Additional arguments (ignored).

x
An object of class summary.meta.

header
A logical indicating whether information on title of meta-analysis, comparison and outcome should be printed at the beginning of the printout.

bylab.nchar
A numeric specifying the number of characters to print from label for the grouping variable.

digits
Minimal number of significant digits, see print.default.

digits.stat
Minimal number of significant digits for z- or t-value of test for overall effect, see print.default.

digits.pval
Minimal number of significant digits for p-value of overall treatment effect, see print.default.

digits.pval.Q
Minimal number of significant digits for p-value of heterogeneity test, see print.default.

digits.Q
Minimal number of significant digits for heterogeneity statistic Q, see print.default.

digits.tau2
Minimal number of significant digits for between-study variance, see print.default.

digits.tau
Minimal number of significant digits for square root of between-study variance, see print.default.

digits.H
Minimal number of significant digits for H statistic, see print.default.

digits.I2
Minimal number of significant digits for I-squared and Rb statistic, see print.default.

scientific.pval
A logical specifying whether p-values should be printed in scientific notation, e.g., 1.2345e-01 instead of 0.12345.

big.mark
A character used as thousands separator.

zero.pval
A logical specifying whether p-values should be printed with a leading zero.

JAMA.pval
A logical specifying whether p-values for test of overall effect should be printed according to JAMA reporting standards.

print.I2
A logical specifying whether heterogeneity statistic I^2 should be printed.

print.H
A logical specifying whether heterogeneity statistic H should be printed.

print.Rb
A logical specifying whether heterogeneity statistic R_b should be printed.

text.tau2
Text printed to identify between-study variance τ^2.

text.tau
Text printed to identify τ, the square root of the between-study variance τ^2.

text.I2
Text printed to identify heterogeneity statistic I^2.

text.Rb
Text printed to identify heterogeneity statistic R_b.

digits.zval
Deprecated argument (replaced by digits.stat).

warn.backtransf
A logical indicating whether a warning should be printed if backtransformed proportions and rates are below 0 and backtransformed proportions are above 1.

Details

Note, in R package meta, version 3.0-0 some arguments have been removed from R functions summary.meta (arguments: byvar, level, level.comb, level.prediction) and print.summary.meta (arguments: level, level.comb, level.prediction). This functionality is now provided by R function update.meta (or directly in meta-analysis functions, e.g., metabin, metacont, metagen, metacor, and metaprop).

Review Manager 5 (RevMan 5) is the current software used for preparing and maintaining Cochrane Reviews (https://training.cochrane.org/online-learning/core-software-cochrane-reviews/revman). In RevMan 5, subgroup analyses can be defined and data from a Cochrane review can be imported to Rusing the function read.rm5. If a meta-analysis is then conducted using function metacr, information on subgroups is available in R (components byvar, bylab, and print.byvar, byvar in an object of class "meta"). Accordingly, by using function metacr there is no need to define subgroups in order to redo the statistical analysis conducted in the Cochrane review.

Note, for an object of type metaprop, starting with version 3.7-0 of meta, list elements TE, lower and upper in element study correspond to transformed proportions and confidence limits (regardless whether exact confidence limits are calculated; argument ciexact=TRUE in metaprop function). Accordingly, the following results are based on the same transformation defined by argument sm: list elements TE, lower and upper in elements study, fixed, random, within.fixed and within.random.

R function cilayout can be utilised to change the layout to print confidence intervals (both in printout from print.meta and print.summary.meta function as well as in forest plots). The default layout is "[lower; upper]". Another popular layout is "(lower - upper)" which is used throughout an R session by using R command cilayout("(", " - ").

Argument pscale can be used to rescale single proportions or risk differences, e.g. pscale=1000 means that proportions are expressed as events per 1000 observations. This is useful in situations with (very) low event probabilities.

Argument irscale can be used to rescale single rates or rate differences, e.g. irscale=1000 means that rates are expressed as events per 1000 time units, e.g. person-years. This is useful in situations with (very) low rates. Argument irunit can be used to specify the time unit used in individual studies (default: "person-years"). This information is printed in summaries and forest plots if argument irscale is not equal to 1.

Value

A list is returned by the function summary.meta with the following elements:

study
Results for individual studies (a list with elements TE, seTE, lower, upper, z, p, level, df).

fixed
Results for fixed effect model (a list with elements TE, seTE, lower, upper, z, p, level, df).

#

random
Results for random effects model (a list with elements TE, seTE, lower, upper, z, p, level, df).

k
Number of studies combined in meta-analysis.

Q
Heterogeneity statistic Q.

tau
Square-root of between-study variance.

se.tau2
Standard error of between-study variance.

H
Heterogeneity statistic H (a list with elements TE, lower, upper).

I2
Heterogeneity statistic I^2 (a list with elements TE, lower, upper), see Higgins & Thompson (2002).

Rb
Heterogeneity statistic R_b (a list with elements TE, lower, upper), see Crippa et al. (2016).

#

k.all
Total number of studies.

Q.CMH
Cochran-Mantel-Haenszel test statistic for overall effect.

sm
A character string indicating underlying summary measure.

method
A character string with the pooling method.

call
Function call.

ci.lab
Label for confidence interval.

hakn
A logical indicating whether method by Hartung and Knapp was used.

adhoc.hakn
A character string indicating whether ad hoc variance correction should be used for Hartung-Knapp method.

method.tau
A character string indicating which method is used to estimate the between-study variance tau-squared.

tau.common
A logical indicating whether tau-squared is assumed to be the same across subgroups.

within.fixed
Result for fixed effect model within groups (a list with elements TE, seTE, lower, upper, z, p, level, df, harmonic.mean) - if byvar is not missing.

within.random
Result for random effects model within groups (a list with elements TE, seTE, lower, upper, z, p, level, df, harmonic.mean) - if byvar is not missing.

k.w
Number of studies combined within groups - if byvar is not missing.

Q.w
Heterogeneity statistic Q within groups - if byvar is not missing.

Q.b.fixed
Heterogeneity statistic Q between groups (based on fixed effect model) - if byvar is not missing.

Q.b.random
Heterogeneity statistic Q between groups (based on random effects model) - if byvar is not missing.

tau.w
Square-root of between-study variance within subgroups - if byvar is not missing.

H.w
Heterogeneity statistic H within subgroups (a list with elements TE, lower, upper) - if byvar is not missing.

I2.w
Heterogeneity statistic I^2 within subgroups (a list with elements TE, lower, upper) - if byvar is not missing.

Rb.w
Heterogeneity statistic R_b within subgroups (a list with elements TE, lower, upper) - if byvar is not missing.

H.resid
Statistic H for residual heterogeneity (a list with elements TE, lower, upper) - if byvar is not missing.

I2.resid
Statistic I^2 for residual heterogeneity (a list with elements TE, lower, upper) - if byvar is not missing.

bylevs
Levels of grouping variable - if byvar is not missing.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

data
Original data (set) used to create meta object.

subset
Information on subset of original data used in meta-analysis.

prediction, level.predict
As defined above.

comb.fixed, comb.random, print.CMH
As defined above.

version
Version of R package meta used to create object.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994): The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation

Crippa A, Khudyakov P, Wang M, Orsini N, Spiegelman D (2016): A new measure of between-studies heterogeneity in meta-analysis. Statistics in Medicine, 35, 3661–75

Higgins JPT & Thompson SG (2002): Quantifying heterogeneity in a meta-analysis. Statistics in Medicine, 21, 1539–58

See Also

update.meta, metabin, metacont, metagen

Examples

data(Fleiss1993cont)
m1 <- metacont(n.psyc, mean.psyc, sd.psyc, n.cont, mean.cont, sd.cont,
               data = Fleiss1993cont, sm = "SMD",
               studlab = paste(study, year))
summary(m1)

summary(update(m1, byvar = c(1, 2, 1, 1, 2), bylab = "group"))
forest(update(m1, byvar = c(1, 2, 1, 1, 2), bylab = "group"))

## Not run: 
# Use unicode characters to print tau^2, tau, and I^2
print(summary(m1),
      text.tau2 = "\u03c4\u00b2", text.tau = "\u03c4", text.I2 = "I\u00b2")

## End(Not run)

[Package meta version 4.18-1 Index]