summary.rm5 {meta}	R Documentation

Cochrane review: Summary of meta-analyses

Description

Calculate and print a summary of all meta-analyses in a Cochrane review.

Usage

## S3 method for class 'rm5'
summary(object, comp.no, outcome.no, ...)

## S3 method for class 'summary.rm5'
print(x, ...)

Arguments

object
An object of class rm5.

comp.no
Comparison number.

outcome.no
Outcome number.

...
Additional arguments (passed on to metacr).

x
An object of class summary.rm5.

Details

This function can be used to redo all or selected meta-analyses of a Cochrane Review.

Review Manager 5 (RevMan 5) is the current software used for preparing and maintaining Cochrane Reviews (https://training.cochrane.org/online-learning/core-software-cochrane-reviews/revman). In RevMan 5, subgroup analyses can be defined and data from a Cochrane review can be imported to R using the function read.rm5.

The R function metacr is called internally.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Higgins, J.P.T and S. Green (2011): Cochrane Handbook for Systematic Reviews of Interventions Version 5.1.0 [Updated March 2011]. The Cochrane Library: http://www.cochrane-handbook.org

See Also

summary.meta, metacr, read.rm5, metabias.rm5

Examples

# Locate export data file "Fleiss1993_CR.csv"
# in sub-directory of package "meta"
#
filename <- system.file("extdata", "Fleiss1993_CR.csv", package = "meta")
Fleiss1993_CR <- read.rm5(filename)

# Print summary results for all meta-analysis
#
summary(Fleiss1993_CR)

# Print summary results only for second outcome of first comparison
#
summary(Fleiss1993_CR, comp.no = 1, outcome.no = 2)

[Package meta version 4.18-1 Index]