trimfill.meta {meta}	R Documentation

Trim and fill method for meta-analysis

Description

Trim and fill method for estimating and adjusting for the number and outcomes of missing studies in a meta-analysis.

Usage

## Default S3 method:
trimfill(x, seTE, left=NULL, ma.fixed=TRUE, type="L", n.iter.max=50,
                        sm=NULL, studlab=NULL, level=0.95, level.comb=0.95,
                        comb.fixed=TRUE, comb.random=TRUE, silent=TRUE, ...)

## S3 method for class 'meta':
trimfill(x, seTE, left=NULL, ma.fixed=TRUE, type="L", n.iter.max=50,
                        sm=NULL, studlab=NULL, level=x$level, level.comb=x$level.comb,
                        comb.fixed=x$comb.fixed, comb.random=x$comb.random, silent=TRUE, ...)

Arguments

x 	 An object of class meta, or estimated treatment effect in individual studies.

seTE 	 Standard error of estimated treatment effect (mandatory if x not of class meta).

left 	 A logical indicating whether studies are supposed to be missing on the left or right side of the funnel plot. If NULL, the linear regression test for funnel plot symmetry (i.e., function metabias(..., meth="linreg")) is used to determine whether trials are missing on the left or right side.

ma.fixed 	 A logical indicating whether a fixed effect or random effects model is used to estimate the number of missing studies.

type 	 A character indicating which method is used to estimate the number of missing studies. Either "L" or "R".

n.iter.max 	 Maximum number of iterations to estimate number of missing studies.

sm 	 An optional character string indicating underlying summary measure, e.g., "RD", "RR", "OR", "AS", "MD", "SMD"; ignored if x is of class meta.

studlab 	 An optional vector with study labels; ignored if x is of class meta.

level 	 The level used to calculate confidence intervals for individual studies. If existing, x$level is used as value for level; otherwise 0.95 is used.

level.comb 	 The level used to calculate confidence interval for the pooled estimate. If existing, x$level.comb is used as value for level.comb; otherwise 0.95 is used.

comb.fixed 	 A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random 	 A logical indicating whether a random effects meta-analysis should be conducted.

silent 	 A logical indicating whether basic information on iterations shown.

... 	 other arguments

Details

The trim and fill method can be used for estimating and adjusting for the number and outcomes of missing studies in a meta-analysis. The method relies on scrutiny of one side of a funnel plot for asymmetry assumed due to publication bias.

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if comb.random=FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random=FALSE.

The function metagen is called internally.

Value

An object of class c("metagen", "meta", "trimfill"). The object is a list containing the following components:

studlab, sm, left, ma.fixed, type

n.iter.max, level, level.comb, 	 As defined above.

comb.fixed, comb.random

TE, seTE 	 Estimated treatment effect and standard error of individual studies.

w.fixed, w.random 	 Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed 	 Estimated overall treatment effect and standard error (fixed effect model).

TE.random, seTE.random 	 Estimated overall treatment effect and standard error (random effects model).

k 	 Number of studies combined in meta-analysis.

Q 	 Heterogeneity statistic Q.

tau 	 Square-root of between-study variance (moment estimator of DerSimonian-Laird).

method 	 Pooling method: "Inverse".

call 	 Function call.

n.iter 	 Actual number of iterations to estimate number of missing studies.

trimfill 	 A logical vector indicating studies that have been added by trim and fill method.

k0 	 Number of studies added by trim and fill.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Duval S & Tweedie R (2000), A nonparametric "Trim and Fill" method of accounting for publication bias in meta-analysis. Journal of the American Statistical Association, 95, 89–98.

Duval S & Tweedie R (2000), Trim and Fill: A simple funnel-plot-based method of testing and adjusting for publication bias in meta-analysis. Biometrics, 56, 455–463.

See Also

metagen, metabias, funnel

Examples

data(Fleiss93)
meta1 <- metabin(event.e, n.e, event.c, n.c,
                 data=Fleiss93, sm="OR")
tf1 <- trimfill(meta1)
summary(tf1)
funnel(tf1, pch=ifelse(tf1$trimfill, 1, 16),
       level=0.95, comb.fixed=TRUE)

trimfill(meta1$TE, meta1$seTE, sm=meta1$sm)

[Package meta version 1.1-8 Index]