update.meta {meta}	R Documentation

Update a meta-analysis object

Description

Update an existing meta-analysis object.

Usage

## S3 method for class 'meta'
update(
  object,
  data = object$data,
  subset,
  studlab,
  exclude,
  id,
  method = object$method,
  sm = object$sm,
  incr,
  allincr = object$allincr,
  addincr = object$addincr,
  allstudies = object$allstudies,
  MH.exact = object$MH.exact,
  RR.Cochrane = object$RR.Cochrane,
  Q.Cochrane = object$Q.Cochrane,
  model.glmm = object$model.glmm,
  level = object$level,
  level.comb = object$level.comb,
  comb.fixed = object$comb.fixed,
  comb.random = object$comb.random,
  overall = object$overall,
  overall.hetstat = object$overall.hetstat,
  hakn = object$hakn,
  adhoc.hakn = object$adhoc.hakn,
  method.tau = object$method.tau,
  method.tau.ci = object$method.tau.ci,
  tau.preset = object$tau.preset,
  TE.tau = object$TE.tau,
  tau.common = object$tau.common,
  prediction = object$prediction,
  level.predict = object$level.predict,
  null.effect = object$null.effect,
  method.bias = object$method.bias,
  backtransf = object$backtransf,
  pscale = object$pscale,
  irscale = object$irscale,
  irunit = object$irunit,
  text.fixed = object$text.fixed,
  text.random = object$text.random,
  text.predict = object$text.predict,
  text.w.fixed = object$text.w.fixed,
  text.w.random = object$text.w.random,
  title = object$title,
  complab = object$complab,
  outclab = object$outclab,
  label.e = object$label.e,
  label.c = object$label.c,
  label.left = object$label.left,
  label.right = object$label.right,
  n.e = object$n.e,
  n.c = object$n.c,
  pooledvar = object$pooledvar,
  method.smd = object$method.smd,
  sd.glass = object$sd.glass,
  exact.smd = object$exact.smd,
  method.ci = object$method.ci,
  byvar,
  bylab = object$bylab,
  print.byvar = object$print.byvar,
  byseparator = object$byseparator,
  print.CMH = object$print.CMH,
  keepdata = TRUE,
  left = object$left,
  ma.fixed = object$ma.fixed,
  type = object$type,
  n.iter.max = object$n.iter.max,
  warn = FALSE,
  control = object$control,
  ...
)

Arguments

object
An object of class meta.

data
Dataset.

subset
Subset.

studlab
Study label.

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots.

id
An optional vector specifying which estimates come from the same study resulting in the use of a three-level meta-analysis model.

method
A character string indicating which method is to be used for pooling of studies; see metabin and metainc function for admissible values.

sm
A character string indicating which summary measure is used for pooling.

incr
Either a numerical value or vector which can be added to each cell frequency for studies with a zero cell count or the character string "TA" which stands for treatment arm continuity correction.

allincr
A logical indicating if incr is added to each cell frequency of all studies if at least one study has a zero cell count. If FALSE (default), incr is added only to each cell frequency of studies with a zero cell count.

addincr
A logical indicating if incr is added to each cell frequency of all studies irrespective of zero cell counts.

allstudies
A logical indicating if studies with zero or all events in both groups are to be included in the meta-analysis (applies only if sm is equal to "RR" or "OR").

MH.exact
A logical indicating if incr is not to be added to all cell frequencies for studies with a zero cell count to calculate the pooled estimate based on the Mantel-Haenszel method.

RR.Cochrane
A logical indicating if 2*incr instead of 1*incr is to be added to n.e and n.c in the calculation of the risk ratio (i.e., sm="RR") for studies with a zero cell. This is used in RevMan 5, the program for preparing and maintaining Cochrane reviews.

Q.Cochrane
A logical indicating if the Mantel-Haenszel estimate is used in the calculation of the heterogeneity statistic Q which is implemented in RevMan 5, the program for preparing and maintaining Cochrane reviews.

model.glmm
A character string indicating which GLMM model should be used.

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

overall
A logical indicating whether overall summaries should be reported. This argument is useful in a meta-analysis with subgroups if overall results should not be reported.

overall.hetstat
A logical value indicating whether to print heterogeneity measures for overall treatment comparisons. This argument is useful in a meta-analysis with subgroups if heterogeneity statistics should only be printed on subgroup level.

hakn
A logical indicating whether the method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

adhoc.hakn
A character string indicating whether an ad hoc variance correction should be applied in the case of an arbitrarily small Hartung-Knapp variance estimate.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2 and its square root τ. Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated. See function metagen.

method.tau.ci
A character string indicating which method is used to estimate the confidence interval of τ^2 and τ. Either "QP", "BJ", or "J", can be abbreviated.

tau.preset
Prespecified value for the square root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance τ^2.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

null.effect
A numeric value specifying the effect under the null hypothesis.

method.bias
A character string indicating which test for funnel plot asymmetry is to be used, can be abbreviated. See function metabias.

backtransf
A logical indicating whether results should be back transformed in printouts and plots. If backtransf = TRUE, results for sm = "OR" are printed as odds ratios rather than log odds ratios and results for sm = "ZCOR" are printed as correlations rather than Fisher's z transformed correlations, for example.

pscale
A numeric giving scaling factor for printing of single event probabilities or risk differences, i.e. if argument sm is equal to "PLOGIT", "PLN", "PRAW", "PAS", "PFT", or "RD".

irscale
A numeric defining a scaling factor for printing of single incidence rates or incidence rate differences, i.e. if argument sm is equal to "IR", "IRLN", "IRS", "IRFT", or "IRD".

irunit
A character specifying the time unit used to calculate rates, e.g. person-years.

text.fixed
A character string used in printouts and forest plot to label the pooled fixed effect estimate.

text.random
A character string used in printouts and forest plot to label the pooled random effects estimate.

text.predict
A character string used in printouts and forest plot to label the prediction interval.

text.w.fixed
A character string used to label weights of fixed effect model.

text.w.random
A character string used to label weights of random effects model.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

label.e
Label for experimental group.

label.c
Label for control group.

label.left
Graph label on left side of forest plot.

label.right
Graph label on right side of forest plot.

n.e
Number of observations in experimental group. (only for metagen object)

n.c
Number of observations in control group. (only for metagen object)

pooledvar
A logical indicating if a pooled variance should be used for the mean difference (only for metacont object with sm = "MD").

method.smd
A character string indicating which method is used to estimate the standardised mean difference (only for metacont object with sm = "SMD"). Either "Hedges" for Hedges' g (default), "Cohen" for Cohen's d, or "Glass" for Glass' delta, can be abbreviated.

sd.glass
A character string indicating which standard deviation is used in the denominator for Glass' method to estimate the standardised mean difference (only for metacont object with sm = "SMD"). Either "control" using the standard deviation in the control group (sd.c) or "experimental" using the standard deviation in the experimental group (sd.e), can be abbreviated.

exact.smd
A logical indicating whether exact formulae should be used in estimation of the standardised mean difference and its standard error.

method.ci
A character string indicating which method is used to calculate confidence intervals for individual studies. Either "z", "t", "WS", "WSCC", "AC", "SA", "SACC", or "NAsm", can be abbreviated. See functions metacont and metaprop.

byvar
An optional vector containing grouping information (must be of same length as event.e).

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

print.CMH
A logical indicating whether result of the Cochran-Mantel-Haenszel test for overall effect should be printed.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

left
A logical indicating whether studies are supposed to be missing on the left or right side of the funnel plot. If NULL, the linear regression test for funnel plot symmetry (i.e., function metabias(..., method = "linreg")) is used to determine whether studies are missing on the left or right side.

ma.fixed
A logical indicating whether a fixed effect or random effects model is used to estimate the number of missing studies.

type
A character indicating which method is used to estimate the number of missing studies. Either "L" or "R".

n.iter.max
Maximum number of iterations to estimate number of missing studies.

warn
A logical indicating whether warnings should be printed (e.g., if incr is added to studies with zero cell frequencies).

control
An optional list to control the iterative process to estimate the between-study variance τ^2. This argument is passed on to rma.uni or rma.glmm, respectively.

...
Additional arguments (ignored at the moment).

Details

Wrapper function to update an existing meta-analysis object which was created with R function metabin, metacont, metacor, metagen, metainc, metamean, metaprop, or metarate. More details on function arguments are available in help files of respective R functions

This function can also be used for objects of class 'trimfill', 'metacum', and 'metainf'.

Value

An object of class "meta" and "metabin", "metacont", "metacor", "metainc", "metagen", "metamean", "metaprop", or "metarate".

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

See Also

metabin, metacont, metacor, metagen, metainc, metamean, metaprop, metarate

Examples

data(Fleiss1993cont)
m1 <- metacont(n.psyc, mean.psyc, sd.psyc, n.cont, mean.cont, sd.cont,
               data = Fleiss1993cont, sm = "SMD",
               studlab = paste(study, year))
m1

# Change summary measure (from 'SMD' to 'MD')
#
update(m1, sm = "MD")

# Restrict analysis to subset of studies
#
update(m1, subset = 1:2)

# Use different levels for confidence intervals
#
m2 <- update(m1, level = 0.66, level.comb = 0.99)
print(m2, digits = 2)
forest(m2)

[Package meta version 4.18-1 Index]