weights.meta {meta}	R Documentation

Calculate absolute and percentage weights for meta-analysis

Description

This function returns a data frame containing information on absolute and percentage weights of individual studies contributing to fixed effect and random effects meta-analysis.

Usage

## S3 method for class 'meta'
weights(
  object,
  comb.fixed = object$comb.fixed,
  comb.random = object$comb.random,
  ...
)

Arguments

object
An object of class meta.

comb.fixed
A logical indicating whether absolute and percentage weights from the fixed effect model should be calculated.

comb.random
A logical indicating whether absolute and percentage weights from the random effects model should be calculated.

...
Additional arguments (ignored at the moment).

Value

A data frame with the following variables is returned:

Variable 	 Definition 	 Condition

w.fixed 	 absolute weights in fixed effect model 	 (if comb.fixed = TRUE)

p.fixed 	 percentage weights in fixed effect model 	 (if comb.fixed = TRUE)

w.random 	 absolute weights in random effects model 	 (if comb.random = TRUE)

p.random 	 percentage weights in random effects model 	 (if comb.random = TRUE)

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

See Also

metabin, metacont, metagen

Examples

data(Fleiss1993cont)
# Do meta-analysis (fixed effect and random effects model)
#
meta1 <- metacont(n.psyc, mean.psyc, sd.psyc, n.cont, mean.cont, sd.cont,
                  data = Fleiss1993cont, sm = "SMD",
                  studlab = paste(study, year))

# Print weights for fixed effect and random effects meta-analysis
#
weights(meta1)

# Do meta-analysis (only random effects model)
#
meta2 <- update(meta1, comb.fixed = FALSE)

# Print weights for random effects meta-analysis
#
weights(meta2)

# Print weights for fixed effect and random effects meta-analysis
#
weights(meta2, comb.fixed = TRUE)

[Package meta version 4.18-1 Index]