woodyplants {meta}	R Documentation

Elevated CO_2 and total biomass of woody plants

Description

Meta-analysis on effects of elevated CO_2 on total biomass of woody plants

Usage

data(woodyplants)

Format

A data frame with the following columns:

obsno

Observation number

papno

Database paper number

treat

Treatment code

level

Treatment level

n.elev

Number of observations in experimental group (elevated CO_2-level)

mean.elev

Estimated mean in experimental group

sd.elev

Standard deviation in experimental group

n.amb

Number of observations in control group (ambient CO_2-level)

mean.amb

Estimated mean in control group

sd.amb

Standard deviation in control group

Details

This dataset has been used as an example in Hedges et al. (1999) to describe methods for the meta-analysis of response ratios. The complete dataset with 102 observations and 26 variables is available online as a supplement. Here only a subset of 10 variables is provided and used in the examples.

Source

Website http://www.esapubs.org/archive/ecol/E080/008/

References

Hedges LV, Gurevitch J, Curtis PS (1999), The meta-analysis of response ratios in experimental ecology. Ecology, 80, 1150–6.

Examples

data(woodyplants)

# Meta-analysis of response ratios (Hedges et al., 1999)
#
meta8 <- metacont(n.elev, mean.elev, sd.elev,
                  n.amb, mean.amb, sd.amb,
                  data=woodyplants, sm="ROM",
                  studlab=paste(obsno, papno, sep = " / "))
summary(meta8, prediction=TRUE)

# Meta-analysis for plants grown with low soil fertility treatment
#
meta9 <- update(meta8, subset=(treat=="fert"&level=="low"))
summary(meta9, prediction=TRUE)

# Meta-analysis for plants grown under low light conditions
#
meta10 <- update(meta8, subset=(treat=="light"&level=="low"))
summary(meta10, prediction=TRUE)

[Package meta version 4.9-4 Index]