amlodipine {meta}	R Documentation

Amlodipine for Work Capacity

Description

Meta-analysis on the effect of amlodipine on work capacity.

This meta-analysis is used as a data example in Hartung and Knapp (2001).

Format

A data frame with the following columns:

study	 study label

n.amlo	 number of observations in amlodipine group

mean.amlo	 estimated mean in amlodipine group

var.amlo	 variance in amlodipine group

n.plac	 number of observations in placebo group

mean.plac	 estimated mean in placebo group

var.plac	 variance in placebo group

Source

Hartung J & Knapp G (2001): On tests of the overall treatment effect in meta-analysis with normally distributed responses. Statistics in Medicine, 20, 1771–82

See Also

metacont

Examples

data(amlodipine)

m <- metacont(n.amlo, mean.amlo, sqrt(var.amlo),
              n.plac, mean.plac, sqrt(var.plac),
              data = amlodipine, studlab = study)
s1 <- summary(m)
s2 <- summary(update(m, hakn = TRUE))

vars <- c("TE", "lower", "upper")

# Same results for mean difference as in Table III in Hartung and
# Knapp (2001)
#
res.md <- rbind(data.frame(s1$fixed)[vars],
                data.frame(s1$random)[vars],
                data.frame(s2$random)[vars])
#
res.md <- round(res.md, 5)
#
row.names(res.md) <- c("FE", "RE", "RE (HaKn)")
names(res.md) <- c("Absolute difference", "CI lower", "CI upper")
#
res.md

[Package meta version 4.18-1 Index]