meta-package {meta}	R Documentation

meta: Brief overview of methods and general hints

Description

R package meta is a user-friendly general package providing standard methods for meta-analysis and supporting Schwarzer et al. (2015), https://www.springer.com/gp/book/9783319214153.

Details

R package meta (Schwarzer, 2007; Balduzzi et al., 2019) provides the following statistical methods for meta-analysis.

Fixed effect and random effects model:

Meta-analysis of continuous outcome data (metacont)

Meta-analysis of binary outcome data (metabin)

Meta-analysis of incidence rates (metainc)

Generic inverse variance meta-analysis (metagen)

Meta-analysis of single correlations (metacor)

Meta-analysis of single means (metamean)

Meta-analysis of single proportions (metaprop)

Meta-analysis of single incidence rates (metarate)

Several plots for meta-analysis:

Forest plot (forest.meta, forest.metabind)

Funnel plot (funnel.meta)

Galbraith plot / radial plot (radial.meta)

L'Abbe plot for meta-analysis with binary outcome data (labbe.metabin, labbe.default)

Baujat plot to explore heterogeneity in meta-analysis (baujat.meta)

Bubble plot to display the result of a meta-regression (bubble.metareg)

Statistical tests for funnel plot asymmetry (metabias.meta, metabias.rm5) and trim-and-fill method (trimfill.meta, trimfill.default) to evaluate bias in meta-analysis

Cumulative meta-analysis (metacum) and leave-one-out meta-analysis (metainf)

Meta-regression (metareg)

Import data from Review Manager 5 (read.rm5); see also metacr to conduct meta-analysis for a single comparison and outcome from a Cochrane review

Prediction interval for the treatment effect of a new study (Higgins et al., 2009); see argument prediction in meta-analysis functions, e.g., metagen

Hartung-Knapp method for random effects meta-analysis (Hartung & Knapp, 2001a,b); see argument hakn in meta-analysis functions, e.g., metagen

Various estimators for the between-study variance τ^2 in a random effects model (Veroniki et al., 2016); see argument method.tau in meta-analysis functions, e.g., metagen

Generalised linear mixed models (metabin, metainc, metaprop, and metarate)

The following more advanced statistical methods are provided by add-on R packages:

Frequentist methods for network meta-analysis (R package netmeta)

Advanced methods to model and adjust for bias in meta-analysis (R package metasens)

Results of several meta-analyses can be combined with metabind. This is, for example, useful to generate a forest plot with results of subgroup analyses.

See settings.meta to learn how to print and specify default meta-analysis methods used during your R session. For example, the function can be used to specify general settings:

settings.meta("revman5")

settings.meta("jama")

settings.meta("iqwig5")

settings.meta("iqwig6")

settings.meta("geneexpr")

The first command can be used to reproduce meta-analyses from Cochrane reviews conducted with Review Manager 5 (RevMan 5, https://training.cochrane.org/online-learning/core-software-cochrane-reviews/revman) and specifies to use a RevMan 5 layout in forest plots. The second command can be used to generate forest plots following instructions for authors of the Journal of the American Medical Association (https://jamanetwork.com/journals/jama/pages/instructions-for-authors/). The next two commands implement the recommendations of the Institute for Quality and Efficiency in Health Care (IQWiG), Germany accordinging to General Methods 5 and 6, respectively (https://www.iqwig.de/en/about-us/methods/methods-paper/). The last setting can be used to print p-values in scientific notation and to suppress the calculation of confidence intervals for the between-study variance.

In addition, settings.meta can be used to change individual settings. For example, the following R command specifies the use of the Hartung-Knapp and Paule-Mandel methods, and the printing of prediction intervals in the current R session for any meta-analysis generated after execution of this command:

settings.meta(hakn=TRUE, method.tau="PM", prediction=TRUE)

Type help(package = "meta") for a listing of R functions and datasets available in meta.

Balduzzi et al. (2019) is the preferred citation in publications for meta. Type citation("meta") for a BibTeX entry of this publication.

To report problems and bugs

type bug.report(package = "meta") if you do not use RStudio,

send an email to Guido Schwarzer sc@imbi.uni-freiburg.de if you use RStudio.

The development version of meta is available on GitHub https://github.com/guido-s/meta/.

Note

R package meta imports R functions from metafor (Viechtbauer, 2010) to

estimate the between-study variance τ^2,

conduct meta-regression,

estimate generalised linear mixed models.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Balduzzi S, Rücker G, Schwarzer G (2019): How to perform a meta-analysis with R: a practical tutorial. Evidence-Based Mental Health, 22, 153–160

Hartung J, Knapp G (2001a): On tests of the overall treatment effect in meta-analysis with normally distributed responses. Statistics in Medicine, 20, 1771–82

Hartung J, Knapp G (2001b): A refined method for the meta-analysis of controlled clinical trials with binary outcome. Statistics in Medicine, 20, 3875–89

Higgins JPT, Thompson SG, Spiegelhalter DJ (2009): A re-evaluation of random-effects meta-analysis. Journal of the Royal Statistical Society: Series A, 172, 137–59

Schwarzer G (2007): meta: An R package for meta-analysis. R News, 7, 40–5

Schwarzer G, Carpenter JR and Rücker G (2015): Meta-Analysis with R (Use-R!). Springer International Publishing, Switzerland

Veroniki AA, Jackson D, Viechtbauer W, Bender R, Bowden J, Knapp G, et al. (2016): Methods to estimate the between-study variance and its uncertainty in meta-analysis. Research Synthesis Methods, 7, 55–79

Viechtbauer W (2010): Conducting Meta-Analyses in R with the metafor Package. Journal of Statistical Software, 36, 1–48

[Package meta version 4.18-1 Index]