metabias {meta}	R Documentation

Test for funnel plot asymmetry

Description

Test for funnel plot asymmetry, based on rank correlation or linear regression method.

Usage

metabias(x, seTE, TE.fixed, seTE.fixed,
         method = "rank",
         plotit = FALSE, correct = FALSE)

Arguments

x 	 An object of class meta or estimated treatment effect in individual studies.

seTE 	 Standard error of estimated treatment effect (mandatory if x not of class meta).

TE.fixed 	 Overall treatment estimate (mandatory if x not of class meta and method = "rank").

seTE.fixed 	 Standard error of overall treatment estimate (mandatory if x not of class meta and method = "rank").

method 	 A character string indicating which test is to be used. Either "rank", "linreg", "mm", "count", "score", or "peters", can be abbreviated.

plotit 	 A logical indicating whether a plot should be produced for method "rank", "linreg", "mm", or "score".

correct 	 A logical indicating whether a continuity corrected statistic is used for rank correlation methods "rank" and "count".

Details

If method is "rank", the test statistic is based on the rank correlation between standardised treatment estimates and variance estimates of estimated treatment effects; Kendall's tau is used as correlation measure (Begg & Mazumdar, 1994). The test statistic follows a standard normal distribution. By default (if correct is FALSE), no continuity correction is utilised (Kendall & Gibbons, 1990).

If method is "linreg", the test statistic is based on a weighted linear regression of the treatment effect on its standard error (Egger et al., 1997). The test statistic follows a t distribution with number of studies - 2 degrees of freedom.

If method is "mm", the test statistic is based on a weighted linear regression of the treatment effect on its standard error using the method of moments estimator for the additive between-study variance component (method 3a in Thompson, Sharp, 1999). The test statistic follows a t distribution with number of studies - 2 degrees of freedom.

If method is "count", the test statistic is based on the rank correlation between a standardised cell frequency and the inverse of the variance of the cell frequency; Kendall's tau is used as correlation measure (Schwarzer et al., 2007). The test statistic follows a standard normal distribution. By default (if correct is FALSE), no continuity correction is utilised (Kendall & Gibbons, 1990).

If method is "score", the test statistic is based on a weighted linear regression utilising efficient score and score variance (Harbord et al., 2006). The test statistic follows a t distribution with number of studies - 2 degrees of freedom.

If method is "peters", the test statistic is based on a weighted linear regression of the treatment effect on the inverse of the total sample size using the variance of the average event rate as weights (Peters et al., 2006). The test statistic follows a t distribution with number of studies - 2 degrees of freedom.

In order to calculate an arcsine test for funnel plot asymmetry (Ruecker et al., 2008), one has to use the metabin function with parameter sm="AS" as input to the metabias command. The three arcsine tests described in Ruecker et al. (2008) can be calculated by setting method to "rank", "linreg" and "mm", respectively.

Value

A list with class "htest" containing the following components:

estimate 	 The estimated degree of funnel plot asymmetry, with name "ks" or "bias" corresponding to the method employed, i.e., rank correlation or regression method.

statistic 	 The value of the test statistic.

parameter 	 The degrees of freedom of the test statistic in the case that it follows a t distribution.

p.value 	 The p-value for the test.

alternative 	 A character string describing the alternative hypothesis.

method 	 A character string indicating what type of test was used.

data.name 	 A character string giving the names of the data.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Begg CB & Mazumdar M (1994), Operating characteristics of a rank correlation test for publication bias. Biometrics, 50, 1088–1101.

Egger M, Smith GD, Schneider M & Minder C (1997), Bias in meta-analysis detected by a simple, graphical test. British Medical Journal, 315, 629–634.

Harbord RM, Egger M & Sterne J (2006), A modified test for small-study effects in meta-analyses of controlled trials with binary endpoints. Statistics in Medicine, 25, 3443–3457.

Kendall M & Gibbons JD (1990), Rank Correlation Methods. London: Edward Arnold.

Peters JL, Sutton AJ, Jones DR, Abrams KR & Rushton L (2006), Comparison of two methods to detect publication bias in meta-analysis. Journal of the American Medical Association, 295, 676–680.

Ruecker G, Schwarzer G, Carpenter JR (2008) Arcsine test for publication bias in meta-analyses with binary outcomes. Statistics in Medicine, 27,746–763.

Schwarzer G, Antes G & Schumacher M (2007), A test for publication bias in meta-analysis with sparse binary data. Statistics in Medicine, 26, 721–733.

Thompson SG & Sharp, SJ (1999), Explaining heterogeneity in meta-analysis: A comparison of methods, Statistics in Medicine, 18, 2693–2708.

See Also

funnel, funnel.meta, metabin, metacont, metagen

Examples

data(Olkin95)
meta1 <- metabin(event.e, n.e, event.c, n.c,
                 data=Olkin95, subset=c(41,47,51,59),
                 sm="RR", meth="I")

metabias(meta1)
metabias(meta1, correct=TRUE)

metabias(meta1, method="linreg")
metabias(meta1, method="linreg", plotit=TRUE)

metabias(meta1, method="count")

##
## Same result:
##
metabias(meta1, method="linreg")$p.value
metabias(meta1$TE, meta1$seTE, method="linreg")$p.value

##
## Arcsine test:
##
meta1.as <- metabin(event.e, n.e, event.c, n.c,
                    data=Olkin95, subset=c(41,47,51,59),
                    sm="AS", meth="I")
metabias(meta1.as, method="linreg")

[Package meta version 1.1-8 Index]