metabin {meta}	R Documentation

Meta-analysis of binary outcome data

Description

Calculation of fixed and random effects estimates (relative risk, odds ratio, risk difference or arcsine difference) for meta-analyses with binary outcome data. Mantel-Haenszel, inverse variance and Peto method are available for pooling.

Usage

metabin(event.e, n.e, event.c, n.c, studlab,
        data = NULL, subset = NULL, method = "MH",
        sm = ifelse(!is.na(charmatch(method, c("Peto", "peto"), nomatch = NA)), "OR", "RR"),
        incr = 0.5, allincr = FALSE, addincr = FALSE, allstudies = FALSE,
        MH.exact = FALSE, RR.cochrane = FALSE,
        level = 0.95, level.comb = level,
        comb.fixed=TRUE, comb.random=TRUE,
        title="", complab="", outclab="",
        label.e="Experimental", label.c="Control",
        byvar, bylab, print.byvar=TRUE,
        warn = TRUE)

Arguments

event.e 	 Number of events in experimental group.

n.e 	 Number of observations in experimental group.

event.c 	 Number of events in control group.

n.c 	 Number of observations in control group.

studlab 	 An optional vector with study labels.

data 	 An optional data frame containing the study information, i.e., event.e, n.e, event.c, and n.c.

subset 	 An optional vector specifying a subset of studies to be used.

method 	 A character string indicating which method is to be used for pooling of studies. One of "Inverse", "MH", or "Peto", can be abbreviated.

sm 	 A character string indicating which summary measure ("RR", "OR", "RD", or "AS") is to be used for pooling of studies, see Details.

incr 	 Could be either a numerical value which is added to each cell frequency for studies with a zero cell count or the character string "TA" which stands for treatment arm continuity correction, see Details.

allincr 	 A logical indicating if incr is added to each cell frequency of all studies if at least one study has a zero cell count. If false, incr is added only to each cell frequency of studies with a zero cell count.

addincr 	 A logical indicating if incr is added to each cell frequency of all studies irrespective of zero cell counts.

allstudies 	 A logical indicating if studies with zero or all events in both groups are to be included in the meta-analysis (applies only if sm = "RR" or "OR").

MH.exact 	 A logical indicating if incr is not to be added to all cell frequencies for studies with a zero cell count to calculate the pooled estimate based on the Mantel-Haenszel method.

RR.cochrane 	 A logical indicating if 2*incr instead of 1*incr is to be added to n.e and n.c in the calculation of the relative risk (i.e., sm="RR") for studies with a zero cell. This is used in RevMan 5, the Cochrane Collaboration's program for preparing and maintaining Cochrane reviews.

level 	 The level used to calculate confidence intervals for individual studies.

level.comb 	 The level used to calculate confidence intervals for pooled estimates.

comb.fixed 	 A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random 	 A logical indicating whether a random effects meta-analysis should be conducted.

title 	 Title of meta-analysis / systematic review.

complab 	 Comparison label.

outclab 	 Outcome label.

label.e 	 Label for experimental group.

label.c 	 Label for control group.

byvar 	 An optional vector containing grouping information (must be of same length as event.e).

bylab 	 A character string with a label for the grouping variable.

print.byvar 	 A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

warn 	 A logical indicating whether the addition of incr to studies with zero cell frequencies should result in a warning.

Details

Treatment estimates and standard errors are calculated for each study. The following measures of treatment effect are available: relative risk (if sm="RR"), odds ratio (sm="OR"), risk difference (sm="RD"), and arcsine difference (sm="AS").

For studies with a zero cell count, by default, 0.5 is added to all cell frequencies of these studies; if incr is "TA" a treatment arm continuity correction is used instead (Sweeting et al., 2004; Diamond et al., 2007). Treatment estimates and standard errors are only calculated for studies with zero or all events in both groups if allstudies is TRUE.

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if comb.random=FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random=FALSE.

By default, both fixed effect and random effects models are considered (arguments comb.fixed=TRUE and comb.random=TRUE). If method is "MH" (default), the Mantel-Haenszel method is used to calculate the fixed effect estimate; if method is "Inverse", inverse variance weighting is used for pooling; finally, if method is "Peto", the Peto method is used for pooling. The DerSimonian-Laird estimate is used in the random effects model. For the Peto method, Peto's log odds ratio, i.e. (O-E)/V and its standard error sqrt(1/V) with O-E and V denoting "Observed minus Expected" and "V", are utilised in the random effects model. Accordingly, results of a random effects model using sm="Peto" can be (slightly) different to results from a random effects model using sm="MH" or sm="Inverse".

For the Mantel-Haenszel method, by default (if MH.exact is FALSE), 0.5 is added to all cell frequencies of a study with a zero cell count in the calculation of the pooled estimate. This approach is also used in other software, e.g. RevMan 5 and the Stata procedure metan. According to Fleiss (in Cooper & Hedges, 1994), there is no need to add 0.5 to a cell frequency of zero to calculate the Mantel-Haenszel estimate and he advocates the exact method (MH.exact=TRUE). Note, the estimate based on the exact method is not defined if the number of events is zero in all studies either in the experimental or control group.

Value

An object of class c("metabin", "meta") with corresponding print, summary, plot function. The object is a list containing the following components:

event.e, n.e, event.c, n.c, studlab,

sm, method, incr, allincr, addincr,

allstudies, MH.exact, RR.cochrane, warn,

level, level.comb, comb.fixed, comb.random,

byvar, bylab, print.byvar 	 As defined above.

TE, seTE 	 Estimated treatment effect and standard error of individual studies.

w.fixed, w.random 	 Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed 	 Estimated overall treatment effect and standard error (fixed effect model).

TE.random, seTE.random 	 Estimated overall treatment effect and standard error (random effects model).

k 	 Number of studies combined in meta-analysis.

Q 	 Heterogeneity statistic Q.

tau 	 Square-root of between-study variance (moment estimator of DerSimonian-Laird).

Q.CMH 	 Cochrane-Mantel-Haenszel heterogeneity statistic.

incr.e, incr.c 	 Increment added to cells in the experimental and control group, respectively

sparse 	 Logical flag indicating if any study included in meta-analysis has any zero cell frequencies.

call 	 Function call.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994), The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation.

Diamond GA, Bax L, Kaul S (2007, Uncertain Effects of Rosiglitazone on the Risk for Myocardial Infarction and Cardiovascular Death. Annals of Internal Medicine, 147, 578–581.

DerSimonian R & Laird N (1986), Meta-analysis in clinical trials. Controlled Clinical Trials, 7, 177–188.

Fleiss JL (1993), The statistical basis of meta-analysis. Statistical Methods in Medical Research, 2, 121–145.

Greenland S & Robins JM (1985), Estimation of a common effect parameter from sparse follow-up data. Biometrics, 41, 55–68.

Review Manager (RevMan) [Computer program]. Version 5.0. The Nordic Cochrane Centre, The Cochrane Collaboration, 2008.

Ruecker G, Schwarzer G, Carpenter JR (2008) Arcsine test for publication bias in meta-analyses with binary outcomes. Statistics in Medicine, 27,746–763.

StataCorp. 2009. Stata Statistical Software: Release 11. College Station, TX: Stata Corporation.

Sweeting MJ, Sutton AJ, Lambert PC (2004), What to add to nothing? Use and avoidance of continuity corrections in meta-analysis of sparse data. Statistics in Medicine, 23, 1351–1375.

See Also

funnel, metabias, metacont, metagen, print.meta

Examples

metabin(10, 20, 15, 20, sm="OR")

##
## Different results:
##
metabin(0, 10, 0, 10, sm="OR")
metabin(0, 10, 0, 10, sm="OR", allstudies=TRUE)

data(Olkin95)

meta1 <- metabin(event.e, n.e, event.c, n.c,
                 data=Olkin95, subset=c(41,47,51,59),
                 sm="RR", meth="I")
summary(meta1)
funnel(meta1)

meta2 <- metabin(event.e, n.e, event.c, n.c,
                 data=Olkin95, subset=Olkin95$year<1970,
                 sm="RR", meth="I")
summary(meta2)

[Package meta version 1.1-8 Index]