metabin {meta}	R Documentation

Meta-analysis of binary outcome data

Description

Calculation of fixed effect and random effects estimates (risk ratio, odds ratio, risk difference, arcsine difference, or diagnostic odds ratio) for meta-analyses with binary outcome data. Mantel-Haenszel, inverse variance, Peto method, generalised linear mixed model (GLMM), and sample size method are available for pooling. For GLMMs, the rma.glmm function from R package metafor (Viechtbauer, 2010) is called internally.

Usage

metabin(
  event.e,
  n.e,
  event.c,
  n.c,
  studlab,
  data = NULL,
  subset = NULL,
  exclude = NULL,
  method = ifelse(tau.common, "Inverse", gs("method")),
  sm = ifelse(!is.na(charmatch(tolower(method), c("peto", "glmm", "ssw"), nomatch =
    NA)), "OR", gs("smbin")),
  incr = gs("incr"),
  allincr = gs("allincr"),
  addincr = gs("addincr"),
  allstudies = gs("allstudies"),
  MH.exact = gs("MH.exact"),
  RR.Cochrane = gs("RR.Cochrane"),
  Q.Cochrane = gs("Q.Cochrane") & method == "MH" & method.tau == "DL",
  model.glmm = "UM.FS",
  level = gs("level"),
  level.comb = gs("level.comb"),
  comb.fixed = gs("comb.fixed"),
  comb.random = gs("comb.random"),
  overall = comb.fixed | comb.random,
  overall.hetstat = comb.fixed | comb.random,
  hakn = gs("hakn"),
  adhoc.hakn = gs("adhoc.hakn"),
  method.tau = ifelse(!is.na(charmatch(tolower(method), "glmm", nomatch = NA)), "ML",
    gs("method.tau")),
  method.tau.ci = gs("method.tau.ci"),
  tau.preset = NULL,
  TE.tau = NULL,
  tau.common = gs("tau.common"),
  prediction = gs("prediction"),
  level.predict = gs("level.predict"),
  method.bias = ifelse(sm == "OR", "Harbord", ifelse(sm == "DOR", "Deeks",
    gs("method.bias"))),
  backtransf = gs("backtransf"),
  pscale = 1,
  text.fixed = gs("text.fixed"),
  text.random = gs("text.random"),
  text.predict = gs("text.predict"),
  text.w.fixed = gs("text.w.fixed"),
  text.w.random = gs("text.w.random"),
  title = gs("title"),
  complab = gs("complab"),
  outclab = "",
  label.e = gs("label.e"),
  label.c = gs("label.c"),
  label.left = gs("label.left"),
  label.right = gs("label.right"),
  byvar,
  bylab,
  print.byvar = gs("print.byvar"),
  byseparator = gs("byseparator"),
  print.CMH = gs("print.CMH"),
  keepdata = gs("keepdata"),
  warn = gs("warn"),
  control = NULL,
  ...
)

Arguments

event.e
Number of events in experimental group or true positives in diagnostic study.

n.e
Number of observations in experimental group or number of ill participants in diagnostic study.

event.c
Number of events in control group or false positives in diagnostic study.

n.c
Number of observations in control group or number of healthy participants in diagnostic study.

studlab
An optional vector with study labels.

data
An optional data frame containing the study information, i.e., event.e, n.e, event.c, and n.c.

subset
An optional vector specifying a subset of studies to be used.

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots.

method
A character string indicating which method is to be used for pooling of studies. One of "Inverse", "MH", "Peto", "GLMM", or "SSW", can be abbreviated.

sm
A character string indicating which summary measure ("RR", "OR", "RD", "ASD", or "DOR") is to be used for pooling of studies, see Details.

incr
Could be either a numerical value which is added to each cell frequency for studies with a zero cell count or the character string "TACC" which stands for treatment arm continuity correction, see Details.

allincr
A logical indicating if incr is added to each cell frequency of all studies if at least one study has a zero cell count. If FALSE (default), incr is added only to each cell frequency of studies with a zero cell count.

addincr
A logical indicating if incr is added to each cell frequency of all studies irrespective of zero cell counts.

allstudies
A logical indicating if studies with zero or all events in both groups are to be included in the meta-analysis (applies only if sm is equal to "RR", "OR", or "DOR").

MH.exact
A logical indicating if incr is not to be added to all cell frequencies for studies with a zero cell count to calculate the pooled estimate based on the Mantel-Haenszel method.

RR.Cochrane
A logical indicating if 2*incr instead of 1*incr is to be added to n.e and n.c in the calculation of the risk ratio (i.e., sm="RR") for studies with a zero cell. This is used in RevMan 5, the program for preparing and maintaining Cochrane reviews.

Q.Cochrane
A logical indicating if the Mantel-Haenszel estimate is used in the calculation of the heterogeneity statistic Q which is implemented in RevMan 5, the program for preparing and maintaining Cochrane reviews.

model.glmm
A character string indicating which GLMM should be used. One of "UM.FS", "UM.RS", "CM.EL", and "CM.AL", see Details.

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

overall
A logical indicating whether overall summaries should be reported. This argument is useful in a meta-analysis with subgroups if overall results should not be reported.

overall.hetstat
A logical value indicating whether to print heterogeneity measures for overall treatment comparisons. This argument is useful in a meta-analysis with subgroups if heterogeneity statistics should only be printed on subgroup level.

hakn
A logical indicating whether the method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

adhoc.hakn
A character string indicating whether an ad hoc variance correction should be applied in the case of an arbitrarily small Hartung-Knapp variance estimate, see Details.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2 and its square root τ. Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.

method.tau.ci
A character string indicating which method is used to estimate the confidence interval of τ^2 and τ. Either "QP", "BJ", or "J", or "", can be abbreviated.

tau.preset
Prespecified value for the square root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance τ^2.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

method.bias
A character string indicating which test for funnel plot asymmetry is to be used. Either "Begg", "Egger", "Thompson", "Schwarzer", "Harbord", "Peters", or "Deeks", can be abbreviated. See function metabias.

backtransf
A logical indicating whether results for odds ratio (sm="OR"), risk ratio (sm="RR"), or diagnostic odds ratio (sm="DOR") should be back transformed in printouts and plots. If TRUE (default), results will be presented as odds ratios and risk ratios; otherwise log odds ratios and log risk ratios will be shown.

pscale
A numeric defining a scaling factor for printing of risk differences.

text.fixed
A character string used in printouts and forest plot to label the pooled fixed effect estimate.

text.random
A character string used in printouts and forest plot to label the pooled random effects estimate.

text.predict
A character string used in printouts and forest plot to label the prediction interval.

text.w.fixed
A character string used to label weights of fixed effect model.

text.w.random
A character string used to label weights of random effects model.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

label.e
Label for experimental group.

label.c
Label for control group.

label.left
Graph label on left side of forest plot.

label.right
Graph label on right side of forest plot.

byvar
An optional vector containing grouping information (must be of same length as event.e).

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

print.CMH
A logical indicating whether result of the Cochran-Mantel-Haenszel test for overall effect should be printed.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

warn
A logical indicating whether warnings should be printed (e.g., if incr is added to studies with zero cell frequencies).

control
An optional list to control the iterative process to estimate the between-study variance τ^2. This argument is passed on to rma.uni or rma.glmm, respectively.

...
Additional arguments passed on to rma.glmm function.

Details

Calculation of fixed and random effects estimates for meta-analyses with binary outcome data.

The following measures of treatment effect are available (Rücker et al., 2009):

Risk ratio (sm = "RR")

Odds ratio (sm = "OR")

Risk difference (sm = "RD")

Arcsine difference (sm = "ASD")

Diagnostic Odds ratio (sm = "DOR")

Note, mathematically, odds ratios and diagnostic odds ratios are identical, however, the labels in printouts and figures differ.

Default settings are utilised for several arguments (assignments using gs function). These defaults can be changed for the current R session using the settings.meta function.

Furthermore, R function update.meta can be used to rerun a meta-analysis with different settings.

Meta-analysis method

By default, both fixed effect and random effects models are considered (see arguments comb.fixed and comb.random). If method is "MH" (default), the Mantel-Haenszel method (Greenland & Robins, 1985; Robins et al., 1986) is used to calculate the fixed effect estimate; if method is "Inverse", inverse variance weighting is used for pooling (Fleiss, 1993); if method is "Peto", the Peto method is used for pooling (Yussuf et al., 1985); if method is "SSW", the sample size method is used for pooling (Bakbergenuly et al., 2020).

While the Mantel-Haenszel and Peto method are defined under the fixed effect model, random effects variants based on these methods are also implemented in metabin. Following RevMan 5, the Mantel-Haenszel estimator is used in the calculation of the between-study heterogeneity statistic Q which is used in the DerSimonian-Laird estimator. Accordlingly, the results for the random effects meta-analysis using the Mantel-Haenszel or inverse variance method are typically very similar. For the Peto method, Peto's log odds ratio, i.e. (O-E) / V and its standard error sqrt(1 / V) with O-E and V denoting "Observed minus Expected" and its variance, are utilised in the random effects model. Accordingly, results of a random effects model using sm = "Peto" can be different to results from a random effects model using sm = "MH" or sm = "Inverse".

A distinctive and frequently overlooked advantage of binary endpoints is that individual patient data (IPD) can be extracted from a two-by-two table. Accordingly, statistical methods for IPD, i.e., logistic regression and generalised linear mixed models, can be utilised in a meta-analysis of binary outcomes (Stijnen et al., 2010; Simmonds et al., 2016). These methods are available (argument method = "GLMM") for the odds ratio as summary measure by calling the rma.glmm function from R package metafor internally.

Four different GLMMs are available for meta-analysis with binary outcomes using argument model.glmm (which corresponds to argument model in the rma.glmm function):

1. 	 Logistic regression model with fixed study effects (default)

(model.glmm = "UM.FS", i.e., Unconditional Model - Fixed Study effects)

2. 	 Mixed-effects logistic regression model with random study effects

(model.glmm = "UM.RS", i.e., Unconditional Model - Random Study effects)

3. 	 Generalised linear mixed model (conditional Hypergeometric-Normal)

(model.glmm = "CM.EL", i.e., Conditional Model - Exact Likelihood)

4. 	 Generalised linear mixed model (conditional Binomial-Normal)

(model.glmm = "CM.AL", i.e., Conditional Model - Approximate Likelihood)

Details on these four GLMMs as well as additional arguments which can be provided using argument '...' in metabin are described in rma.glmm where you can also find information on the iterative algorithms used for estimation. Note, regardless of which value is used for argument model.glmm, results for two different GLMMs are calculated: fixed effect model (with fixed treatment effect) and random effects model (with random treatment effects).

Continuity correction

For studies with a zero cell count, by default, 0.5 is added to all cell frequencies of these studies; if incr is "TACC" a treatment arm continuity correction is used instead (Sweeting et al., 2004; Diamond et al., 2007). For odds ratio and risk ratio, treatment estimates and standard errors are only calculated for studies with zero or all events in both groups if allstudies is TRUE. This continuity correction is used both to calculate individual study results with confidence limits and to conduct meta-analysis based on the inverse variance method. For Peto method and GLMMs no continuity correction is used. For the Mantel-Haenszel method, by default (if MH.exact is FALSE), incr is added to all cell frequencies of a study with a zero cell count in the calculation of the pooled risk ratio or odds ratio as well as the estimation of the variance of the pooled risk difference, risk ratio or odds ratio. This approach is also used in other software, e.g. RevMan 5 and the Stata procedure metan. According to Fleiss (in Cooper & Hedges, 1994), there is no need to add 0.5 to a cell frequency of zero to calculate the Mantel-Haenszel estimate and he advocates the exact method (MH.exact = TRUE). Note, estimates based on exact Mantel-Haenszel method or GLMM are not defined if the number of events is zero in all studies either in the experimental or control group.

Estimation of between-study variance

The following methods to estimate the between-study variance τ^2 are available for the inverse variance method:

DerSimonian-Laird estimator (method.tau = "DL")

Paule-Mandel estimator (method.tau = "PM")

Restricted maximum-likelihood estimator (method.tau = "REML")

Maximum-likelihood estimator (method.tau = "ML")

Hunter-Schmidt estimator (method.tau = "HS")

Sidik-Jonkman estimator (method.tau = "SJ")

Hedges estimator (method.tau = "HE")

Empirical Bayes estimator (method.tau = "EB")

See metagen for more information on these estimators. Note, the maximum-likelihood method is utilized for GLMMs.

Confidence interval for the between-study variance

The following methods to calculate a confidence interval for τ^2 and τ are available.

Argument	 Method

method.tau.ci = "J"	 Method by Jackson

method.tau.ci = "BJ"	 Method by Biggerstaff and Jackson

method.tau.ci = "QP"	 Q-Profile method

See metagen for more information on these methods. For GLMMs, no confidence intervals for τ^2 and τ are calculated. Likewise, no confidence intervals for τ^2 and τ are calculated if method.tau.ci = "".

Hartung-Knapp method

Hartung and Knapp (2001a,b) proposed an alternative method for random effects meta-analysis based on a refined variance estimator for the treatment estimate. Simulation studies (Hartung and Knapp, 2001a,b; IntHout et al., 2014; Langan et al., 2019) show improved coverage probabilities compared to the classic random effects method.

In rare settings with very homogeneous treatment estimates, the Hartung-Knapp variance estimate can be arbitrarily small resulting in a very narrow confidence interval (Knapp and Hartung, 2003; Wiksten et al., 2016). In such cases, an ad hoc variance correction has been proposed by utilising the variance estimate from the classic random effects model with the HK method (Knapp and Hartung, 2003; IQWiQ, 2020). An alternative approach is to use the wider confidence interval of classic fixed or random effects meta-analysis and the HK method (Wiksten et al., 2016; Jackson et al., 2017).

Argument adhoc.hakn can be used to choose the ad hoc method:

Argument	 Ad hoc method

adhoc.hakn = ""	 not used

adhoc.hakn = "se"	 use variance correction if HK standard error is smaller

than standard error from classic random effects

meta-analysis (Knapp and Hartung, 2003)

adhoc.hakn = "iqwig6"	 use variance correction if HK confidence interval

is narrower than CI from classic random effects model

with DerSimonian-Laird estimator (IQWiG, 2020)

adhoc.hakn = "ci"	 use wider confidence interval of classic random effects

and HK meta-analysis

(Hybrid method 2 in Jackson et al., 2017)

For GLMMs, a method similar to Knapp and Hartung (2003) is implemented, see description of argument tdist in rma.glmm, and the ad hoc variance correction is not available.

Prediction interval

A prediction interval for the proportion in a new study (Higgins et al., 2009) is calculated if arguments prediction and comb.random are TRUE. Note, the definition of prediction intervals varies in the literature. This function implements equation (12) of Higgins et al., (2009) which proposed a t distribution with K-2 degrees of freedom where K corresponds to the number of studies in the meta-analysis.

Subgroup analysis

Argument byvar can be used to conduct subgroup analysis for a categorical covariate. The metareg function can be used instead for more than one categorical covariate or continuous covariates.

Exclusion of studies from meta-analysis

Arguments subset and exclude can be used to exclude studies from the meta-analysis. Studies are removed completely from the meta-analysis using argument subset, while excluded studies are shown in printouts and forest plots using argument exclude (see Examples in metagen). Meta-analysis results are the same for both arguments.

Presentation of meta-analysis results

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if argument comb.random = FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random = FALSE.

Value

An object of class c("metabin", "meta") with corresponding print, summary, and forest functions. The object is a list containing the following components:

event.e, n.e, event.c, n.c, studlab, exclude,
As defined above.

sm, method, incr, allincr, addincr,
As defined above.

allstudies, MH.exact, RR.Cochrane, Q.Cochrane, model.glmm,
As defined above.

warn, level, level.comb, comb.fixed, comb.random,
As defined above.

overall, overall.hetstat,
As defined above.

hakn, adhoc.hakn, method.tau, method.tau.ci,
As defined above.

tau.preset, TE.tau, method.bias,
As defined above.

tau.common, title, complab, outclab,
As defined above.

label.e, label.c, label.left, label.right,
As defined above.

byvar, bylab, print.byvar, byseparator
As defined above.

TE, seTE
Estimated treatment effect and standard error of individual studies.

lower, upper
Lower and upper confidence interval limits for individual studies.

zval, pval
z-value and p-value for test of treatment effect for individual studies.

w.fixed, w.random
Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed
Estimated overall treatment effect, e.g., log risk ratio or risk difference, and standard error (fixed effect model).

lower.fixed, upper.fixed
Lower and upper confidence interval limits (fixed effect model).

statistic.fixed, pval.fixed
z-value and p-value for test of overall treatment effect (fixed effect model).

TE.random, seTE.random
Estimated overall treatment effect, e.g., log risk ratio or risk difference, and standard error (random effects model).

lower.random, upper.random
Lower and upper confidence interval limits (random effects model).

statistic.random, pval.random
z-value or t-value and corresponding p-value for test of overall treatment effect (random effects model).

prediction, level.predict
As defined above.

seTE.predict
Standard error utilised for prediction interval.

lower.predict, upper.predict
Lower and upper limits of prediction interval.

k
Number of studies combined in meta-analysis.

Q
Heterogeneity statistic Q.

df.Q
Degrees of freedom for heterogeneity statistic.

pval.Q
P-value of heterogeneity test.

Q.LRT
Heterogeneity statistic for likelihood-ratio test (only if method = "GLMM").

df.Q.LRT
Degrees of freedom for likelihood-ratio test

pval.Q.LRT
P-value of likelihood-ratio test.

tau2
Between-study variance τ^2.

se.tau2
Standard error of τ^2.

lower.tau2, upper.tau2
Lower and upper limit of confidence interval for τ^2.

tau
Square-root of between-study variance τ.

lower.tau, upper.tau
Lower and upper limit of confidence interval for τ.

H
Heterogeneity statistic H.

lower.H, upper.H
Lower and upper confidence limit for heterogeneity statistic H.

I2
Heterogeneity statistic I^2.

lower.I2, upper.I2
Lower and upper confidence limit for heterogeneity statistic I^2.

Rb
Heterogeneity statistic R_b.

lower.Rb, upper.Rb
Lower and upper confidence limit for heterogeneity statistic R_b.

Q.CMH
Cochran-Mantel-Haenszel test statistic for overall effect.

df.Q.CMH
Degrees of freedom for Cochran-Mantel-Haenszel test statistic.

pval.Q.CMH
P-value of Cochran-Mantel-Haenszel test.

incr.e, incr.c
Increment added to cells in the experimental and control group, respectively.

sparse
Logical flag indicating if any study included in meta-analysis has any zero cell frequencies.

doublezeros
Logical flag indicating if any study has zero cell frequencies in both treatment groups.

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn = TRUE).

k.MH
Number of studies combined in meta-analysis using Mantel-Haenszel method.

bylevs
Levels of grouping variable - if byvar is not missing.

TE.fixed.w, seTE.fixed.w
Estimated treatment effect and standard error in subgroups (fixed effect model) - if byvar is not missing.

lower.fixed.w, upper.fixed.w
Lower and upper confidence interval limits in subgroups (fixed effect model) - if byvar is not missing.

statistic.fixed.w, pval.fixed.w
z-value and p-value for test of treatment effect in subgroups (fixed effect model) - if byvar is not missing.

TE.random.w, seTE.random.w
Estimated treatment effect and standard error in subgroups (random effects model) - if byvar is not missing.

lower.random.w, upper.random.w
Lower and upper confidence interval limits in subgroups (random effects model) - if byvar is not missing.

statistic.random.w, pval.random.w
z-value or t-value and corresponding p-value for test of treatment effect in subgroups (random effects model) - if byvar is not missing.

w.fixed.w, w.random.w
Weight of subgroups (in fixed and random effects model) - if byvar is not missing.

df.hakn.w
Degrees of freedom for test of treatment effect for Hartung-Knapp method in subgroups - if byvar is not missing and hakn = TRUE.

event.e.w
Number of events in experimental group in subgroups - if byvar is not missing.

n.e.w
Number of observations in experimental group in subgroups - if byvar is not missing.

event.c.w
Number of events in control group in subgroups - if byvar is not missing.

n.c.w
Number of observations in control group in subgroups - if byvar is not missing.

k.w
Number of studies combined within subgroups - if byvar is not missing.

k.all.w
Number of all studies in subgroups - if byvar is not missing.

Q.w.fixed
Overall within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.w.random
Overall within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing (only calculated if argument tau.common is TRUE).

df.Q.w
Degrees of freedom for test of overall within subgroups heterogeneity - if byvar is not missing.

pval.Q.w.fixed
P-value of within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.w.random
P-value of within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

Q.b.fixed
Overall between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.b.random
Overall between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

df.Q.b
Degrees of freedom for test of overall between subgroups heterogeneity - if byvar is not missing.

pval.Q.b.fixed
P-value of between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.b.random
P-value of between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

tau.w
Square-root of between-study variance within subgroups - if byvar is not missing.

H.w
Heterogeneity statistic H within subgroups - if byvar is not missing.

lower.H.w, upper.H.w
Lower and upper confidence limit for heterogeneity statistic H within subgroups - if byvar is not missing.

I2.w
Heterogeneity statistic I^2 within subgroups - if byvar is not missing.

lower.I2.w, upper.I2.w
Lower and upper confidence limit for heterogeneity statistic I^2 within subgroups - if byvar is not missing.

keepdata
As defined above.

data
Original data (set) used in function call (if keepdata = TRUE).

subset
Information on subset of original data used in meta-analysis (if keepdata = TRUE).

.glmm.fixed
GLMM object generated by call of rma.glmm function (fixed effect model).

.glmm.random
GLMM object generated by call of rma.glmm function (random effects model).

call
Function call.

version
Version of R package meta used to create object.

version.metafor
Version of R package metafor used for GLMMs.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Bakbergenuly I, Hoaglin DC, Kulinskaya E (2020): Methods for estimating between-study variance and overall effect in meta-analysis of odds-ratios. Research Synthesis Methods, DOI: 10.1002/jrsm.1404

Cooper H & Hedges LV (1994): The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation

Diamond GA, Bax L, Kaul S (2007): Uncertain Effects of Rosiglitazone on the Risk for Myocardial Infarction and Cardiovascular Death. Annals of Internal Medicine, 147, 578–81

DerSimonian R & Laird N (1986): Meta-analysis in clinical trials. Controlled Clinical Trials, 7, 177–88

Fleiss JL (1993): The statistical basis of meta-analysis. Statistical Methods in Medical Research, 2, 121–45

Greenland S & Robins JM (1985): Estimation of a common effect parameter from sparse follow-up data. Biometrics, 41, 55–68

Hartung J & Knapp G (2001): A refined method for the meta-analysis of controlled clinical trials with binary outcome. Statistics in Medicine, 20, 3875–89

Higgins JPT, Thompson SG, Spiegelhalter DJ (2009): A re-evaluation of random-effects meta-analysis. Journal of the Royal Statistical Society: Series A, 172, 137–59

IQWiG (2020): General Methods: Version 6.0. https://www.iqwig.de/en/about-us/methods/methods-paper/

Knapp G & Hartung J (2003): Improved tests for a random effects meta-regression with a single covariate. Statistics in Medicine, 22, 2693–710

Review Manager (RevMan) [Computer program]. Version 5.4. The Cochrane Collaboration, 2020

Paule RC & Mandel J (1982): Consensus values and weighting factors. Journal of Research of the National Bureau of Standards, 87, 377–85

Pettigrew HM, Gart JJ, Thomas DG (1986): The bias and higher cumulants of the logarithm of a binomial variate. Biometrika, 73, 425–35

Robins J, Breslow N, Greenland S (1986): Estimators of the Mantel-Haenszel Variance Consistent in Both Sparse Data and Large-Strata Limiting Models. Biometrics, 42, 311–23

Rücker G, Schwarzer G, Carpenter J, Olkin I (2009): Why add anything to nothing? The arcsine difference as a measure of treatment effect in meta-analysis with zero cells. Statistics in Medicine, 28, 721–38

Simmonds MC, Higgins JP (2016): A general framework for the use of logistic regression models in meta-analysis. Statistical Methods in Medical Research, 25, 2858–77

StataCorp. 2011. Stata Statistical Software: Release 12. College Station, TX: StataCorp LP.

Stijnen T, Hamza TH, Ozdemir P (2010): Random effects meta-analysis of event outcome in the framework of the generalized linear mixed model with applications in sparse data. Statistics in Medicine, 29, 3046–67

Sweeting MJ, Sutton AJ, Lambert PC (2004): What to add to nothing? Use and avoidance of continuity corrections in meta-analysis of sparse data. Statistics in Medicine, 23, 1351–75

Viechtbauer W (2010): Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36, 1–48

Wiksten A, Rücker G, Schwarzer G (2016): Hartung-Knapp method is not always conservative compared with fixed-effect meta-analysis. Statistics in Medicine, 35, 2503–15

Yusuf S, Peto R, Lewis J, Collins R, Sleight P (1985): Beta blockade during and after myocardial infarction: An overview of the randomized trials. Progress in Cardiovascular Diseases, 27, 335–71

See Also

update.meta, forest, funnel, metabias, metacont, metagen, metareg, print.meta

Examples

# Calculate odds ratio and confidence interval for a single study
#
metabin(10, 20, 15, 20, sm = "OR")

# Different results (due to handling of studies with double zeros)
#
metabin(0, 10, 0, 10, sm = "OR")
metabin(0, 10, 0, 10, sm = "OR", allstudies = TRUE)

# Use subset of Olkin (1995) to conduct meta-analysis based on
# inverse variance method (with risk ratio as summary measure)
#
data(Olkin1995)
m1 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
              data = Olkin1995, subset = c(41, 47, 51, 59),
              method = "Inverse")
summary(m1)

# Use different subset of Olkin (1995)
#
m2 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
              data = Olkin1995, subset = year < 1970,
              method = "Inverse", studlab = author)
summary(m2)
forest(m2)

# Meta-analysis with odds ratio as summary measure
#
m3 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
              data = Olkin1995, subset = year < 1970,
              sm = "OR", method = "Inverse", studlab = author)
# Same meta-analysis result using 'update.meta' function
m3 <- update(m2, sm = "OR")
summary(m3)

# Meta-analysis based on Mantel-Haenszel method (with odds ratio as
# summary measure)
#
m4 <- update(m3, method = "MH")
summary(m4)

# Meta-analysis based on Peto method (only available for odds ratio
# as summary measure)
#
m5 <- update(m3, method = "Peto")
summary(m5)

## Not run: 
# Meta-analysis using generalised linear mixed models (only if R
# packages 'metafor' and 'lme4' are available)
#
if (suppressMessages(require(metafor, quietly = TRUE, warn = FALSE)) &
    require(lme4, quietly = TRUE)) {

# Logistic regression model with (k = 4) fixed study effects
# (default: model.glmm = "UM.FS")
#
m6 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
              data = Olkin1995, subset = year < 1970,
              method = "GLMM")
# Same results:
m6 <- update(m2, method = "GLMM")
summary(m6)

# Mixed-effects logistic regression model with random study effects
# (warning message printed due to argument 'nAGQ')
#
m7 <- update(m6, model.glmm = "UM.RS")
#
# Use additional argument 'nAGQ' for internal call of 'rma.glmm'
# function
#
m7 <- update(m6, model.glmm = "UM.RS", nAGQ = 1)
summary(m7)

# Generalised linear mixed model (conditional
# Hypergeometric-Normal) (R package 'BiasedUrn' must be available)
#
if (require(BiasedUrn, quietly = TRUE)) {
 m8 <- update(m6, model.glmm = "CM.EL")
 summary(m8)
}

# Generalised linear mixed model (conditional Binomial-Normal)
#
m9 <- update(m6, model.glmm = "CM.AL")
summary(m9)

# Logistic regression model with (k = 70) fixed study effects
# (about 18 seconds with Intel Core i7-3667U, 2.0GHz)
#
m10 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
               data = Olkin1995, method = "GLMM")
summary(m10)

# Mixed-effects logistic regression model with random study effects
# - about 50 seconds with Intel Core i7-3667U, 2.0GHz
# - several warning messages, e.g. "failure to converge, ..."
#
summary(update(m10, model.glmm = "UM.RS"))

# Conditional Hypergeometric-Normal GLMM
# - long computation time (about 12 minutes with Intel Core
#   i7-3667U, 2.0GHz)
# - estimation problems for this very large dataset:
#   * warning that Choleski factorization of Hessian failed
#   * confidence interval for treatment effect smaller in random
#     effects model compared to fixed effect model
#
if (require(BiasedUrn, quietly = TRUE)) {
 system.time(m11 <- update(m10, model.glmm = "CM.EL"))
 summary(m11)
}

# Generalised linear mixed model (conditional Binomial-Normal)
# (less than 1 second with Intel Core i7-3667U, 2.0GHz)
#
summary(update(m10, model.glmm = "CM.AL"))
}

## End(Not run)

[Package meta version 4.18-1 Index]