metabind {meta}	R Documentation

Combine meta-analysis objects

Description

This function can be used to combine meta-analysis objects and is, for example, useful to generate a forest plot with results of subgroup analyses.

Usage

metabind(..., name, pooled, backtransf, outclab)

Arguments

...
Any number of meta-analysis objects (see Details).

name
An optional character vector providing descriptive names for the meta-analysis objects.

pooled
A character string indicating whether results of a fixed effect or random effects model should be considered. Either "fixed" or "random", can be abbreviated.

backtransf
A logical indicating whether results should be back transformed in printouts and plots. If backtransf=TRUE (default), results for sm="OR" are printed as odds ratios rather than log odds ratios, for example.

outclab
Outcome label for all meta-analyis objects.

Details

This function can be used to combine meta-analysis objects and is, for example, useful to generate a forest plot with results of subgroup analyses.

Value

An object of class c("metabind", "meta") with corresponding print, summary, and forest functions. See metagen for more information on list elements.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

See Also

metagen, forest.metabind

Examples

data(Fleiss93cont)

# Add some (fictitious) grouping variables:
Fleiss93cont$age <- c(55, 65, 55, 65, 55)
Fleiss93cont$region <- c("Europe", "Europe", "Asia", "Asia", "Europe")

meta1 <- metacont(n.e, mean.e, sd.e,
                  n.c, mean.c, sd.c,
                  data = Fleiss93cont, sm = "MD")

mu1 <- update(meta1, byvar = age, bylab = "Age group")
mu2 <- update(meta1, byvar = region, bylab = "Region")

mb1 <- metabind(mu1, mu2)
mb1

forest(mb1)

[Package meta version 4.9-4 Index]