metacont {meta}	R Documentation

Meta-analysis of continuous outcome data

Description

Calculation of fixed and random effects estimates for meta-analyses with continuous outcome data; inverse variance weighting is used for pooling.

Usage

metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, studlab,
         data=NULL, subset=NULL, sm="MD",
         level = 0.95, level.comb = level,
         comb.fixed=TRUE, comb.random=TRUE,
         title="", complab="", outclab="",
         label.e="Experimental", label.c="Control",
         byvar, bylab, print.byvar=TRUE)

Arguments

n.e 	 Number of observations in experimental group.

mean.e 	 Estimated mean in experimental group.

sd.e 	 Standard deviation in experimental group.

n.c 	 Number of observations in control group.

mean.c 	 Estimated mean in control group.

sd.c 	 Standard deviation in control group.

studlab 	 An optional vector with study labels.

data 	 An optional data frame containing the study information.

subset 	 An optional vector specifying a subset of studies to be used.

level 	 The level used to calculate confidence intervals for individual studies.

level.comb 	 The level used to calculate confidence intervals for pooled estimates.

comb.fixed 	 A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random 	 A logical indicating whether a random effects meta-analysis should be conducted.

title 	 Title of meta-analysis / systematic review.

complab 	 Comparison label.

outclab 	 Outcome label.

label.e 	 Label for experimental group.

label.c 	 Label for control group.

sm 	 A character string indicating which summary measure ("MD" or "SMD") is to be used for pooling of studies.

byvar 	 An optional vector containing grouping information (must be of same length as n.e).

bylab 	 A character string with a label for the grouping variable.

print.byvar 	 A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

Details

Calculation of fixed and random effects estimates for meta-analyses with continuous outcome data; inverse variance weighting is used for pooling. The DerSimonian-Laird estimate is used in the random effects model. The mean difference is used as measure of treatment effect if sm="MD" – which correspond to sm="WMD" in older versions (<0.9) of the meta package. For the summary measure "SMD", Hedges' adjusted g is utilised for pooling.

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if comb.random=FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random=FALSE.

The function metagen is called internally to calculate individual and overall treatment estimates and standard errors.

Value

An object of class c("metacont", "meta") with corresponding print, summary, plot function. The object is a list containing the following components:

n.e, mean.e, sd.e,

n.c, mean.c, sd.c,

studlab, sm, level, level.comb,

comb.fixed, comb.random,

byvar, bylab, print.byvar 	 As defined above.

TE, seTE 	 Estimated treatment effect and standard error of individual studies.

w.fixed, w.random 	 Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed 	 Estimated overall treatment effect and standard error (fixed effect model).

TE.random, seTE.random 	 Estimated overall treatment effect and standard error (random effects model).

k 	 Number of studies combined in meta-analysis.

Q 	 Heterogeneity statistic.

tau 	 Square-root of between-study variance (moment estimator of DerSimonian-Laird).

method 	 Pooling method: "Inverse".

call 	 Function call.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994), The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation.

See Also

metabin, metagen

Examples

data(Fleiss93cont)
meta1 <- metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, data=Fleiss93cont, sm="SMD")
meta1

meta2 <- metacont(Fleiss93cont$n.e, Fleiss93cont$mean.e,
                  Fleiss93cont$sd.e,
                  Fleiss93cont$n.c, Fleiss93cont$mean.c,
                  Fleiss93cont$sd.c,
                  sm="SMD")
meta2

[Package meta version 1.1-8 Index]