metacont {meta}	R Documentation

Meta-analysis of continuous outcome data

Description

Calculation of fixed and random effects estimates for meta-analyses with continuous outcome data; inverse variance weighting is used for pooling.

Usage

metacont(
  n.e,
  mean.e,
  sd.e,
  n.c,
  mean.c,
  sd.c,
  studlab,
  data = NULL,
  subset = NULL,
  exclude = NULL,
  id = NULL,
  median.e,
  q1.e,
  q3.e,
  min.e,
  max.e,
  median.c,
  q1.c,
  q3.c,
  min.c,
  max.c,
  method.mean = "Luo",
  method.sd = "Shi",
  approx.mean.e,
  approx.mean.c = approx.mean.e,
  approx.sd.e,
  approx.sd.c = approx.sd.e,
  sm = gs("smcont"),
  pooledvar = gs("pooledvar"),
  method.smd = gs("method.smd"),
  sd.glass = gs("sd.glass"),
  exact.smd = gs("exact.smd"),
  method.ci = gs("method.ci.cont"),
  level = gs("level"),
  level.comb = gs("level.comb"),
  comb.fixed = gs("comb.fixed"),
  comb.random = gs("comb.random"),
  overall = comb.fixed | comb.random,
  overall.hetstat = comb.fixed | comb.random,
  hakn = gs("hakn"),
  adhoc.hakn = gs("adhoc.hakn"),
  method.tau = gs("method.tau"),
  method.tau.ci = gs("method.tau.ci"),
  tau.preset = NULL,
  TE.tau = NULL,
  tau.common = gs("tau.common"),
  prediction = gs("prediction"),
  level.predict = gs("level.predict"),
  method.bias = gs("method.bias"),
  backtransf = gs("backtransf"),
  text.fixed = gs("text.fixed"),
  text.random = gs("text.random"),
  text.predict = gs("text.predict"),
  text.w.fixed = gs("text.w.fixed"),
  text.w.random = gs("text.w.random"),
  title = gs("title"),
  complab = gs("complab"),
  outclab = "",
  label.e = gs("label.e"),
  label.c = gs("label.c"),
  label.left = gs("label.left"),
  label.right = gs("label.right"),
  byvar,
  bylab,
  print.byvar = gs("print.byvar"),
  byseparator = gs("byseparator"),
  keepdata = gs("keepdata"),
  warn = gs("warn"),
  control = NULL
)

Arguments

n.e
Number of observations in experimental group.

mean.e
Estimated mean in experimental group.

sd.e
Standard deviation in experimental group.

n.c
Number of observations in control group.

mean.c
Estimated mean in control group.

sd.c
Standard deviation in control group.

studlab
An optional vector with study labels.

data
An optional data frame containing the study information.

subset
An optional vector specifying a subset of studies to be used.

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots.

id
An optional vector specifying which estimates come from the same study resulting in the use of a three-level meta-analysis model.

median.e
Median in experimental group (used to estimate the mean and standard deviation).

q1.e
First quartile in experimental group (used to estimate the mean and standard deviation).

q3.e
Third quartile in experimental group (used to estimate the mean and standard deviation).

min.e
Minimum in experimental group (used to estimate the mean and standard deviation).

max.e
Maximum in experimental group (used to estimate the mean and standard deviation).

median.c
Median in control group (used to estimate the mean and standard deviation).

q1.c
First quartile in control group (used to estimate the mean and standard deviation).

q3.c
Third quartile in control group (used to estimate the mean and standard deviation).

min.c
Minimum in control group (used to estimate the mean and standard deviation).

max.c
Maximum in control group (used to estimate the mean and standard deviation).

method.mean
A character string indicating which method to use to approximate the mean from the median and other statistics (see Details).

method.sd
A character string indicating which method to use to approximate the standard deviation from sample size, median, interquartile range and range (see Details).

approx.mean.e
Approximation method to estimate means in experimental group (see Details).

approx.mean.c
Approximation method to estimate means in control group (see Details).

approx.sd.e
Approximation method to estimate standard deviations in experimental group (see Details).

approx.sd.c
Approximation method to estimate standard deviations in control group (see Details).

sm
A character string indicating which summary measure ("MD", "SMD", or "ROM") is to be used for pooling of studies.

pooledvar
A logical indicating if a pooled variance should be used for the mean difference (sm="MD").

method.smd
A character string indicating which method is used to estimate the standardised mean difference (sm="SMD"). Either "Hedges" for Hedges' g (default), "Cohen" for Cohen's d, or "Glass" for Glass' delta, can be abbreviated.

sd.glass
A character string indicating which standard deviation is used in the denominator for Glass' method to estimate the standardised mean difference. Either "control" using the standard deviation in the control group (sd.c) or "experimental" using the standard deviation in the experimental group (sd.e), can be abbreviated.

exact.smd
A logical indicating whether exact formulae should be used in estimation of the standardised mean difference and its standard error (see Details).

method.ci
A character string indicating which method is used to calculate confidence intervals for individual studies, see Details.

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

overall
A logical indicating whether overall summaries should be reported. This argument is useful in a meta-analysis with subgroups if overall results should not be reported.

overall.hetstat
A logical value indicating whether to print heterogeneity measures for overall treatment comparisons. This argument is useful in a meta-analysis with subgroups if heterogeneity statistics should only be printed on subgroup level.

hakn
A logical indicating whether the method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

adhoc.hakn
A character string indicating whether an ad hoc variance correction should be applied in the case of an arbitrarily small Hartung-Knapp variance estimate, see Details.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2 and its square root τ. Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.

method.tau.ci
A character string indicating which method is used to estimate the confidence interval of τ^2 and τ. Either "QP", "BJ", or "J", or "", can be abbreviated.

tau.preset
Prespecified value for the square root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance tau-squared.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

method.bias
A character string indicating which test is to be used. Either "Begg", "Egger", "Thompson", or "Pustejovsky", can be abbreviated. See function metabias.

backtransf
A logical indicating whether results for ratio of means (sm="ROM") should be back transformed in printouts and plots. If TRUE (default), results will be presented as ratio of means; otherwise log ratio of means will be shown.

text.fixed
A character string used in printouts and forest plot to label the pooled fixed effect estimate.

text.random
A character string used in printouts and forest plot to label the pooled random effects estimate.

text.predict
A character string used in printouts and forest plot to label the prediction interval.

text.w.fixed
A character string used to label weights of fixed effect model.

text.w.random
A character string used to label weights of random effects model.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

label.e
Label for experimental group.

label.c
Label for control group.

label.left
Graph label on left side of forest plot.

label.right
Graph label on right side of forest plot.

byvar
An optional vector containing grouping information (must be of same length as n.e).

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

warn
A logical indicating whether warnings should be printed (e.g., if studies are excluded from meta-analysis due to zero standard deviations).

control
An optional list to control the iterative process to estimate the between-study variance τ^2. This argument is passed on to rma.uni.

Details

Calculation of fixed and random effects estimates for meta-analyses with continuous outcome data; inverse variance weighting is used for pooling.

Three different types of summary measures are available for continuous outcomes:

mean difference (argument sm = "MD")

standardised mean difference (sm = "SMD")

ratio of means (sm = "ROM")

Default settings are utilised for several arguments (assignments using gs function). These defaults can be changed for the current R session using the settings.meta function.

Furthermore, R function update.meta can be used to rerun a meta-analysis with different settings.

Standardised mean difference

For the standardised mean difference three methods are implemented:

Hedges' g (default, method.smd = "Hedges") - see Hedges (1981)

Cohen's d (method.smd = "Cohen") - see Cohen (1988)

Glass' delta (method.smd = "Glass") - see Glass (1976)

Hedges (1981) calculated the exact bias in Cohen's d which is a ratio of gamma distributions with the degrees of freedom, i.e. total sample size minus two, as argument. By default (argument exact.smd = FALSE), an accurate approximation of this bias provided in Hedges (1981) is utilised for Hedges' g as well as its standard error; these approximations are also used in RevMan 5. Following Borenstein et al. (2009) these approximations are not used in the estimation of Cohen's d. White and Thomas (2005) argued that approximations are unnecessary with modern software and accordingly promote to use the exact formulae; this is possible using argument exact.smd = TRUE. For Hedges' g the exact formulae are used to calculate the standardised mean difference as well as the standard error; for Cohen's d the exact formula is only used to calculate the standard error. In typical applications (with sample sizes above 10), the differences between using the exact formulae and the approximation will be minimal.

For Glass' delta, by default (argument sd.glass = "control"), the standard deviation in the control group (sd.c) is used in the denominator of the standard mean difference. The standard deviation in the experimental group (sd.e) can be used by specifying sd.glass = "experimental".

Ratio of means

Meta-analysis of ratio of means – also called response ratios – is described in Hedges et al. (1999) and Friedrich et al. (2008). Calculations are conducted on the log scale and list elements TE, TE.fixed, and TE.random contain the logarithm of the ratio of means. In printouts and plots these values are back transformed if argument backtransf = TRUE.

Approximate means from sample sizes, medians and other statistics

Missing means in the experimental group (analogously for the control group) can be derived from

sample size, median, interquartile range and range (arguments n.e, median.e, q1.e, q3.e, min.e, and max.e),

sample size, median and interquartile range (arguments n.e, median.e, q1.e, and q3.e), or

sample size, median and range (arguments n.e, median.e, min.e, and max.e).

By default, methods described in Luo et al. (2018) are utilized (argument method.mean = "Luo"):

equation (15) if sample size, median, interquartile range and range are available,

equation (11) if sample size, median and interquartile range are available,

equation (7) if sample size, median and range are available.

Instead the methods described in Wan et al. (2014) are used if argument method.mean = "Wan"):

equation (10) if sample size, median, interquartile range and range are available,

equation (14) if sample size, median and interquartile range are available,

equation (2) if sample size, median and range are available.

By default, missing means are replaced successively using interquartile ranges and ranges (if available), interquartile ranges (if available) and finally ranges. Arguments approx.mean.e and approx.mean.c can be used to overwrite this behaviour for each individual study and treatment arm:

use means directly (entry "" in argument approx.mean.e or approx.mean.c);

median, interquartile range and range ("iqr.range");

median and interquartile range ("iqr");

median and range ("range").

Approximate standard deviations from sample sizes, medians and other statistics

Missing standard deviations in the experimental group (analogously for the control group) can be derived from

sample size, median, interquartile range and range (arguments n.e, median.e, q1.e, q3.e, min.e, and max.e),

sample size, median and interquartile range (arguments n.e, median.e, q1.e and q3.e), or

sample size, median and range (arguments n.e, median.e, min.e and max.e).

Wan et al. (2014) describe methods to estimate the standard deviation from the sample size, median and additional statistics. Shi et al. (2020) provide an improved estimate of the standard deviation if the interquartile range and range are available in addition to the sample size and median. Accordingly, equation (11) in Shi et al. (2020) is the default (argument method.sd = "Shi"), if the median, interquartile range and range are provided. The method by Wan et al. (2014) is used if argument method.sd = "Wan" and, depending on the sample size, either equation (12) or (13) is used. If only the interquartile range or range is available, equations (15) / (16) and (7) / (9) in Wan et al. (2014) are used, respectively.

By default, missing standard deviations are replaced successively using these method, i.e., interquartile ranges and ranges are used before interquartile ranges before ranges. Arguments approx.sd.e and approx.sd.c can be used to overwrite this default for each individual study and treatment arms:

sample size, median, interquartile range and range ("iqr.range");

sample size, median and interquartile range ("iqr");

sample size, median and range ("range").

Confidence intervals for individual studies

For the mean difference (argument sm = "MD"), the confidence interval for individual studies can be based on the

standard normal distribution (method.ci = "z", default), or

t-distribution (method.ci = "t").

Note, this choice does not affect the results of the fixed effect and random effects meta-analysis.

Estimation of between-study variance

The following methods to estimate the between-study variance τ^2 are available:

DerSimonian-Laird estimator (method.tau = "DL")

Paule-Mandel estimator (method.tau = "PM")

Restricted maximum-likelihood estimator (method.tau = "REML")

Maximum-likelihood estimator (method.tau = "ML")

Hunter-Schmidt estimator (method.tau = "HS")

Sidik-Jonkman estimator (method.tau = "SJ")

Hedges estimator (method.tau = "HE")

Empirical Bayes estimator (method.tau = "EB")

See metagen for more information on these estimators.

Confidence interval for the between-study variance

The following methods to calculate a confidence interval for τ^2 and τ are available.

Argument	 Method

method.tau.ci = "J"	 Method by Jackson (2013)

method.tau.ci = "BJ"	 Method by Biggerstaff and Jackson (2008)

method.tau.ci = "QP"	 Q-Profile method (Viechtbauer, 2007)

method.tau.ci = "PL"	 Profile-Likelihood method for three-level meta-analysis model

(Van den Noortgate et al., 2013)

See metagen for more information on these methods. No confidence intervals for τ^2 and τ are calculated if method.tau.ci = "".

Hartung-Knapp method

Hartung and Knapp (2001) proposed an alternative method for random effects meta-analysis based on a refined variance estimator for the treatment estimate. Simulation studies (Hartung and Knapp, 2001; IntHout et al., 2014; Langan et al., 2019) show improved coverage probabilities compared to the classic random effects method.

In rare settings with very homogeneous treatment estimates, the Hartung-Knapp variance estimate can be arbitrarily small resulting in a very narrow confidence interval (Knapp and Hartung, 2003; Wiksten et al., 2016). In such cases, an ad hoc variance correction has been proposed by utilising the variance estimate from the classic random effects model with the HK method (Knapp and Hartung, 2003; IQWiQ, 2020). An alternative approach is to use the wider confidence interval of classic fixed or random effects meta-analysis and the HK method (Wiksten et al., 2016; Jackson et al., 2017).

Argument adhoc.hakn can be used to choose the ad hoc method:

Argument	 Ad hoc method

adhoc.hakn = ""	 not used

adhoc.hakn = "se"	 use variance correction if HK standard error is smaller

than standard error from classic random effects

meta-analysis (Knapp and Hartung, 2003)

adhoc.hakn = "iqwig6"	 use variance correction if HK confidence interval

is narrower than CI from classic random effects model

with DerSimonian-Laird estimator (IQWiG, 2020)

adhoc.hakn = "ci"	 use wider confidence interval of classic random effects

and HK meta-analysis

(Hybrid method 2 in Jackson et al., 2017)

Prediction interval

A prediction interval for the proportion in a new study (Higgins et al., 2009) is calculated if arguments prediction and comb.random are TRUE. Note, the definition of prediction intervals varies in the literature. This function implements equation (12) of Higgins et al., (2009) which proposed a t distribution with K-2 degrees of freedom where K corresponds to the number of studies in the meta-analysis.

Subgroup analysis

Argument byvar can be used to conduct subgroup analysis for a categorical covariate. The metareg function can be used instead for more than one categorical covariate or continuous covariates.

Exclusion of studies from meta-analysis

Arguments subset and exclude can be used to exclude studies from the meta-analysis. Studies are removed completely from the meta-analysis using argument subset, while excluded studies are shown in printouts and forest plots using argument exclude (see Examples in metagen). Meta-analysis results are the same for both arguments.

Presentation of meta-analysis results

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if argument comb.random = FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random = FALSE.

Value

An object of class c("metacont", "meta") with corresponding print, summary, and forest functions. The object is a list containing the following components:

n.e, mean.e, sd.e,
As defined above.

n.c, mean.c, sd.c,
As defined above.

studlab, exclude, sm, method.ci,
As defined above.

median.e, q1.e, q3.e, min.e, max.e,
As defined above.

median.c, q1.c, q3.c, min.c, max.c,
As defined above.

method.mean, method.sd,
As defined above.

approx.mean.e, approx.sd.e, approx.mean.c, approx.sd.c,
As defined above.

level, level.comb,
As defined above.

comb.fixed, comb.random,
As defined above.

overall, overall.hetstat,
As defined above.

pooledvar, method.smd, sd.glass,
As defined above.

hakn, adhoc.hakn, method.tau, method.tau.ci,
As defined above.

tau.preset, TE.tau, method.bias,
As defined above.

tau.common, title, complab, outclab,
As defined above.

label.e, label.c, label.left, label.right,
As defined above.

byvar, bylab, print.byvar, byseparator
As defined above.

TE, seTE
Estimated treatment effect and standard error of individual studies.

lower, upper
Lower and upper confidence interval limits for individual studies.

statistic, pval
Statistic and p-value for test of treatment effect for individual studies.

w.fixed, w.random
Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed
Estimated overall treatment effect and standard error (fixed effect model).

lower.fixed, upper.fixed
Lower and upper confidence interval limits (fixed effect model).

statistic.fixed, pval.fixed
Statistic and p-value for test of overall treatment effect (fixed effect model).

TE.random, seTE.random
Estimated overall treatment effect and standard error (random effects model).

lower.random, upper.random
Lower and upper confidence interval limits (random effects model).

statistic.random, pval.random
Statistic and p-value for test of overall treatment effect (random effects model).

prediction, level.predict
As defined above.

seTE.predict
Standard error utilised for prediction interval.

lower.predict, upper.predict
Lower and upper limits of prediction interval.

k
Number of studies combined in meta-analysis.

Q
Heterogeneity statistic Q.

df.Q
Degrees of freedom for heterogeneity statistic.

pval.Q
P-value of heterogeneity test.

tau2
Between-study variance τ^2.

se.tau2
Standard error of τ^2.

lower.tau2, upper.tau2
Lower and upper limit of confidence interval for τ^2.

tau
Square-root of between-study variance τ.

lower.tau, upper.tau
Lower and upper limit of confidence interval for τ.

H
Heterogeneity statistic H.

lower.H, upper.H
Lower and upper confidence limit for heterogeneity statistic H.

I2
Heterogeneity statistic I^2.

lower.I2, upper.I2
Lower and upper confidence limit for heterogeneity statistic I^2.

Rb
Heterogeneity statistic R_b.

lower.Rb, upper.Rb
Lower and upper confidence limit for heterogeneity statistic R_b.

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn = TRUE).

method
Pooling method: "Inverse".

bylevs
Levels of grouping variable - if byvar is not missing.

TE.fixed.w, seTE.fixed.w
Estimated treatment effect and standard error in subgroups (fixed effect model) - if byvar is not missing.

lower.fixed.w, upper.fixed.w
Lower and upper confidence interval limits in subgroups (fixed effect model) - if byvar is not missing.

statistic.fixed.w, pval.fixed.w
Statistics and p-values for test of treatment effect in subgroups (fixed effect model) - if byvar is not missing.

TE.random.w, seTE.random.w
Estimated treatment effect and standard error in subgroups (random effects model) - if byvar is not missing.

lower.random.w, upper.random.w
Lower and upper confidence interval limits in subgroups (random effects model) - if byvar is not missing.

statistic.random.w, pval.random.w
Statistics and p-values for test of treatment effect in subgroups (random effects model) - if byvar is not missing.

w.fixed.w, w.random.w
Weight of subgroups (in fixed and random effects model) - if byvar is not missing.

df.hakn.w
Degrees of freedom for test of treatment effect for Hartung-Knapp method in subgroups - if byvar is not missing and hakn = TRUE.

n.e.w
Number of observations in experimental group in subgroups - if byvar is not missing.

n.c.w
Number of observations in control group in subgroups - if byvar is not missing.

k.w
Number of studies combined within subgroups - if byvar is not missing.

k.all.w
Number of all studies in subgroups - if byvar is not missing.

Q.w.fixed
Overall within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.w.random
Overall within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing (only calculated if argument tau.common is TRUE).

df.Q.w
Degrees of freedom for test of overall within subgroups heterogeneity - if byvar is not missing.

pval.Q.w.fixed
P-value of within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.w.random
P-value of within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

Q.b.fixed
Overall between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.b.random
Overall between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

df.Q.b
Degrees of freedom for test of overall between subgroups heterogeneity - if byvar is not missing.

pval.Q.b.fixed
P-value of between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.b.random
P-value of between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

tau.w
Square-root of between-study variance within subgroups - if byvar is not missing.

H.w
Heterogeneity statistic H within subgroups - if byvar is not missing.

lower.H.w, upper.H.w
Lower and upper confidence limit for heterogeneity statistic H within subgroups - if byvar is not missing.

I2.w
Heterogeneity statistic I^2 within subgroups - if byvar is not missing.

lower.I2.w, upper.I2.w
Lower and upper confidence limit for heterogeneity statistic I^2 within subgroups - if byvar is not missing.

keepdata
As defined above.

data
Original data (set) used in function call (if keepdata = TRUE).

subset
Information on subset of original data used in meta-analysis (if keepdata = TRUE).

call
Function call.

version
Version of R package meta used to create object.

Note

The function metagen is called internally to calculate individual and overall treatment estimates and standard errors.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Borenstein M, Hedges LV, Higgins JPT, Rothstein HR (2009): Introduction to Meta-Analysis. Chichester: Wiley

Cohen J (1988): Statistical Power Analysis for the Behavioral Sciences (second ed.). Lawrence Erlbaum Associates

Cooper H & Hedges LV (1994): The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation

DerSimonian R & Laird N (1986): Meta-analysis in clinical trials. Controlled Clinical Trials, 7, 177–88

Friedrich JO, Adhikari NK, Beyene J (2008): The ratio of means method as an alternative to mean differences for analyzing continuous outcome variables in meta-analysis: A simulation study. BMC Medical Research Methodology, 8, 32

Glass G (1976): Primary, secondary, and meta-analysis of research. Educational Researcher, 5, 3–8

Hartung J & Knapp G (2001): On tests of the overall treatment effect in meta-analysis with normally distributed responses. Statistics in Medicine, 20, 1771–82

Hedges LV (1981): Distribution theory for Glass's estimator of effect size and related estimators. Journal of Educational and Behavioral Statistics, 6, 107–28

Hedges LV, Gurevitch J, Curtis PS (1999): The meta-analysis of response ratios in experimental ecology. Ecology, 80, 1150–6

Higgins JPT, Thompson SG, Spiegelhalter DJ (2009): A re-evaluation of random-effects meta-analysis. Journal of the Royal Statistical Society: Series A, 172, 137–59

IntHout J, Ioannidis JPA, Borm GF (2014): The Hartung-Knapp-Sidik-Jonkman method for random effects meta-analysis is straightforward and considerably outperforms the standard DerSimonian-Laird method. BMC Medical Research Methodology, 14, 25

IQWiG (2020): General Methods: Version 6.0. https://www.iqwig.de/en/about-us/methods/methods-paper/

Jackson D, Law M, Rücker G, Schwarzer G (2017): The Hartung-Knapp modification for random-effects meta-analysis: A useful refinement but are there any residual concerns? Statistics in Medicine, 36, 3923–34

Knapp G & Hartung J (2003): Improved tests for a random effects meta-regression with a single covariate. Statistics in Medicine, 22, 2693–710

Langan D, Higgins JPT, Jackson D, Bowden J, Veroniki AA, Kontopantelis E, et al. (2019): A comparison of heterogeneity variance estimators in simulated random-effects meta-analyses. Research Synthesis Methods, 10, 83–98

Luo D, Wan X, Liu J, Tong T (2018): Optimally estimating the sample mean from the sample size, median, mid-range, and/or mid-quartile range. Statistical Methods in Medical Research, 27, 1785–805

Review Manager (RevMan) [Computer program]. Version 5.4. The Cochrane Collaboration, 2020

Van den Noortgate W, López-López JA, Marín-Martínez F, Sánchez-Meca J (2013): Three-level meta-analysis of dependent effect sizes. Behavior Research Methods, 45, 576–94

Viechtbauer W (2010): Conducting Meta-Analyses in R with the Metafor Package. Journal of Statistical Software, 36, 1–48

Wan X, Wang W, Liu J, Tong T (2014): Estimating the sample mean and standard deviation from the sample size, median, range and/or interquartile range. BMC Medical Research Methodology, 14, 135

White IR, Thomas J (2005): Standardized mean differences in individually-randomized and cluster-randomized trials, with applications to meta-analysis. Clinical Trials, 2, 141–51

Wiksten A, Rücker G, Schwarzer G (2016): Hartung-Knapp method is not always conservative compared with fixed-effect meta-analysis. Statistics in Medicine, 35, 2503–15

See Also

update.meta, metabin, metagen

Examples

data(Fleiss1993cont)

# Meta-analysis with Hedges' g as effect measure
#
m1 <- metacont(n.psyc, mean.psyc, sd.psyc, n.cont, mean.cont, sd.cont,
               data = Fleiss1993cont, sm = "SMD")
m1
forest(m1)

# Use Cohen's d instead of Hedges' g as effect measure
#
update(m1, method.smd = "Cohen")

# Use Glass' delta instead of Hedges' g as effect measure
#
update(m1, method.smd = "Glass")

# Use Glass' delta based on the standard deviation in the experimental group
#
update(m1, method.smd = "Glass", sd.glass = "experimental")

# Calculate Hedges' g based on exact formulae
#
update(m1, exact.smd = TRUE)

data(amlodipine)
m2 <- metacont(n.amlo, mean.amlo, sqrt(var.amlo),
               n.plac, mean.plac, sqrt(var.plac),
               data = amlodipine, studlab = study)
summary(m2)

# Use pooled variance
#
summary(update(m2, pooledvar = TRUE))

# Meta-analysis of response ratios (Hedges et al., 1999)
#
data(woodyplants)
m3 <- metacont(n.elev, mean.elev, sd.elev,
		  n.amb, mean.amb, sd.amb,
               data = woodyplants, sm = "ROM")
summary(m3)
summary(m3, backtransf = FALSE)

[Package meta version 4.18-1 Index]