metacr {meta}	R Documentation

Meta-analysis of outcome data from Cochrane review

Description

Wrapper function to perform meta-analysis for a single outcome of a Cochrane Intervention review.

Usage

metacr(x, comp.no=1, outcome.no=1, smother="", logscale=TRUE)

Arguments

x 	 An object of class rm5 created by R function read.rm5.

comp.no 	 Comparison number.

outcome.no 	 Outcome number.

smother 	 A character specifying the summary measure to use instead of the value "OTHER"

logscale 	 A logical indicating whether treatment effects have been entered on the log scale in RevMan5. Only relevant for an outcome with summary measure equal to "OTHER".

Details

Cochrane Intervention reviews are based on the comparison of two interventions. Each Cochrane Intervention review can have a variable number of comparisons. For each comparison, a variable number of outcomes can be define. For each outcome, a seperate meta-analysis is conducted. Review Manager 5 (RevMan 5) is the current software used for preparing and maintaining Cochrane Reviews (http://www.cc-ims.net/revman/).

This wrapper function can be used to perform meta-analysis for a single outcome of a Cochrane Intervention review. Internally, R functions metabin, metacont, and metagen are called - depending on the definition of the outcome in RevMan 5.

Value

An object of class "meta" and "metabin", "metacont", or "metagen" depending on outcome type utilised in Cochrane Intervention review for selected outcome.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Review Manager (RevMan) [Computer program]. Version 5.0, The Nordic Cochrane Centre, The Cochrane Collaboration, 2008.

See Also

metabin, metacont, metagen, read.rm5

Examples

## Locate export data file "Fleiss93_CR.csv" or "Fleiss93_CR_Windows.csv"
## in sub-directory of package "meta"
##
filename <- paste(searchpaths()[seq(along=search())[search()==
                  "package:meta"]], "/data/Fleiss93_CR",
                  if (Sys.info()[["sysname"]]=="Windows") "_Windows" else "",
                  ".csv", sep="")
Fleiss93_CR <- read.rm5(filename)

## Same result: example(Fleiss93)
##
metacr(Fleiss93_CR)

## Same result: example(Fleiss93cont)
##
metacr(Fleiss93_CR, 1, 2)

[Package meta version 1.1-8 Index]