metagen {meta}	R Documentation

Generic inverse variance meta-analysis

Description

Fixed and random effects meta-analysis based on estimates (e.g. log hazard ratios) and their standard errors; inverse variance weighting is used for pooling.

Usage

metagen(TE, seTE, studlab, data=NULL, subset=NULL, sm="",
        level = 0.95, level.comb = level,
        comb.fixed=TRUE, comb.random=TRUE,
        title="", complab="", outclab="",
        label.e="Experimental", label.c="Control",
        byvar, bylab, print.byvar=TRUE)

Arguments

TE 	 Estimate of treatment effect.

seTE 	 Standard error of treatment estimate.

studlab 	 An optional vector with study labels.

data 	 An optional data frame containing the study information.

subset 	 An optional vector specifying a subset of studies to be used.

sm 	 A character string indicating underlying summary measure, e.g., "RD", "RR", "OR", "AS", "MD", "SMD".

level 	 The level used to calculate confidence intervals for individual studies.

level.comb 	 The level used to calculate confidence intervals for pooled estimates.

comb.fixed 	 A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random 	 A logical indicating whether a random effects meta-analysis should be conducted.

title 	 Title of meta-analysis / systematic review.

complab 	 Comparison label.

outclab 	 Outcome label.

label.e 	 Label for experimental group.

label.c 	 Label for control group.

byvar 	 An optional vector containing grouping information (must be of same length as TE).

bylab 	 A character string with a label for the grouping variable.

print.byvar 	 A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

Details

Generic method for meta-analysis, only treatment estimates and their standard error are needed. The method is useful, e.g., for pooling of survival data (using log hazard ratio and standard errors as input). The inverse variance method is used for pooling. Random effects estimate is based on the DerSimonian-Laird method.

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if comb.random=FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random=FALSE.

Value

An object of class c("metagen", "meta") with corresponding print, summary, plot function. The object is a list containing the following components:

TE, seTE, studlab,

sm, level, level.comb,

comb.fixed, comb.random,

byvar, bylab, print.byvar 	 As defined above.

w.fixed, w.random 	 Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed 	 Estimated overall treatment effect and standard error (fixed effect model).

TE.random, seTE.random 	 Estimated overall treatment effect and standard error (random effects model).

k 	 Number of studies combined in meta-analysis.

Q 	 Heterogeneity statistic.

tau 	 Square-root of between-study variance (moment estimator of DerSimonian-Laird).

method 	 Pooling method: "Inverse".

call 	 Function call.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994), The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation.

See Also

metabin, metacont, print.meta

Examples

data(Fleiss93)
meta1 <- metabin(event.e, n.e, event.c, n.c, data=Fleiss93, sm="RR", meth="I")
meta1

##
## Identical results by using the following commands:
##
meta1
metagen(meta1$TE, meta1$seTE, sm="RR")

##
## Meta-analysis of survival data:
##
logHR <- log(c(0.95, 1.5))
selogHR <- c(0.25, 0.35)

metagen(logHR, selogHR, sm="HR")

[Package meta version 1.1-8 Index]