metagen {meta}	R Documentation

Generic inverse variance meta-analysis

Description

Fixed effect and random effects meta-analysis based on estimates (e.g. log hazard ratios) and their standard errors. The inverse variance method is used for pooling.

Usage

metagen(
  TE,
  seTE,
  studlab,
  data = NULL,
  subset = NULL,
  exclude = NULL,
  id = NULL,
  sm = "",
  method.ci = if (missing(df)) "z" else "t",
  level = gs("level"),
  level.comb = gs("level.comb"),
  comb.fixed = gs("comb.fixed"),
  comb.random = gs("comb.random"),
  overall = comb.fixed | comb.random,
  overall.hetstat = comb.fixed | comb.random,
  hakn = gs("hakn"),
  adhoc.hakn = gs("adhoc.hakn"),
  method.tau = gs("method.tau"),
  method.tau.ci = gs("method.tau.ci"),
  tau.preset = NULL,
  TE.tau = NULL,
  tau.common = gs("tau.common"),
  detail.tau = "",
  prediction = gs("prediction"),
  level.predict = gs("level.predict"),
  null.effect = 0,
  method.bias = gs("method.bias"),
  n.e = NULL,
  n.c = NULL,
  pval,
  df,
  lower,
  upper,
  level.ci = 0.95,
  median,
  q1,
  q3,
  min,
  max,
  method.mean = "Luo",
  method.sd = "Shi",
  approx.TE,
  approx.seTE,
  backtransf = gs("backtransf"),
  pscale = 1,
  irscale = 1,
  irunit = "person-years",
  text.fixed = gs("text.fixed"),
  text.random = gs("text.random"),
  text.predict = gs("text.predict"),
  text.w.fixed = gs("text.w.fixed"),
  text.w.random = gs("text.w.random"),
  title = gs("title"),
  complab = gs("complab"),
  outclab = "",
  label.e = gs("label.e"),
  label.c = gs("label.c"),
  label.left = gs("label.left"),
  label.right = gs("label.right"),
  byvar,
  bylab,
  print.byvar = gs("print.byvar"),
  byseparator = gs("byseparator"),
  keepdata = gs("keepdata"),
  warn = gs("warn"),
  control = NULL
)

Arguments

TE
Estimate of treatment effect, e.g., log hazard ratio or risk difference.

seTE
Standard error of treatment estimate.

studlab
An optional vector with study labels.

data
An optional data frame containing the study information.

subset
An optional vector specifying a subset of studies to be used (see Details).

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots (see Details).

id
An optional vector specifying which estimates come from the same study resulting in the use of a three-level meta-analysis model.

sm
A character string indicating underlying summary measure, e.g., "RD", "RR", "OR", "ASD", "HR", "MD", "SMD", or "ROM".

method.ci
A character string indicating which method is used to calculate confidence intervals for individual studies, see Details.

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

overall
A logical indicating whether overall summaries should be reported. This argument is useful in a meta-analysis with subgroups if overall results should not be reported.

overall.hetstat
A logical value indicating whether to print heterogeneity measures for overall treatment comparisons. This argument is useful in a meta-analysis with subgroups if heterogeneity statistics should only be printed on subgroup level.

hakn
A logical indicating whether method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

adhoc.hakn
A character string indicating whether an ad hoc variance correction should be applied in the case of an arbitrarily small Hartung-Knapp variance estimate. Either "", "se", "ci", or "iqwig6" (see Details), can be abbreviated.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2 and its square root τ. Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.

method.tau.ci
A character string indicating which method is used to estimate the confidence interval of τ^2 and τ. Either "QP", "BJ", "J", "PL", or "", can be abbreviated.

tau.preset
Prespecified value for the square root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance tau-squared.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

detail.tau
Detail on between-study variance estimate.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

null.effect
A numeric value specifying the effect under the null hypothesis.

method.bias
A character string indicating which test is to be used. Either "Begg", "Egger", or "Thompson", can be abbreviated. See function metabias.

n.e
Number of observations in experimental group (or total sample size in study).

n.c
Number of observations in control group.

pval
P-value (used to estimate the standard error).

df
Degrees of freedom (used in test or to construct confidence interval).

lower
Lower limit of confidence interval (used to estimate the standard error).

upper
Upper limit of confidence interval (used to estimate the standard error).

level.ci
Level of confidence interval.

median
Median (used to estimate the treatment effect and standard error).

q1
First quartile (used to estimate the treatment effect and standard error).

q3
Third quartile (used to estimate the treatment effect and standard error).

min
Minimum (used to estimate the treatment effect and standard error).

max
Maximum (used to estimate the treatment effect and standard error).

method.mean
A character string indicating which method to use to approximate the mean from the median and other statistics (see Details).

method.sd
A character string indicating which method to use to approximate the standard deviation from sample size, median, interquartile range and range (see Details).

approx.TE
Approximation method to estimate treatment estimate (see Details).

approx.seTE
Approximation method to estimate standard error (see Details).

backtransf
A logical indicating whether results should be back transformed in printouts and plots. If backtransf = TRUE (default), results for sm = "OR" are printed as odds ratios rather than log odds ratios and results for sm = "ZCOR" are printed as correlations rather than Fisher's z transformed correlations, for example.

pscale
A numeric giving scaling factor for printing of single event probabilities or risk differences, i.e. if argument sm is equal to "PLOGIT", "PLN", "PRAW", "PAS", "PFT", or "RD".

irscale
A numeric defining a scaling factor for printing of single incidence rates or incidence rate differences, i.e. if argument sm is equal to "IR", "IRLN", "IRS", "IRFT", or "IRD".

irunit
A character specifying the time unit used to calculate rates, e.g. person-years.

text.fixed
A character string used in printouts and forest plot to label the pooled fixed effect estimate.

text.random
A character string used in printouts and forest plot to label the pooled random effects estimate.

text.predict
A character string used in printouts and forest plot to label the prediction interval.

text.w.fixed
A character string used to label weights of fixed effect model.

text.w.random
A character string used to label weights of random effects model.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

label.e
Label for experimental group.

label.c
Label for control group.

label.left
Graph label on left side of forest plot.

label.right
Graph label on right side of forest plot.

byvar
An optional vector containing grouping information (must be of same length as TE).

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

warn
A logical indicating whether warnings should be printed (e.g., if studies are excluded from meta-analysis due to zero standard errors).

control
An optional list to control the iterative process to estimate the between-study variance τ^2. This argument is passed on to rma.uni or rma.mv.

Details

This function provides the generic inverse variance method for meta-analysis which requires treatment estimates and their standard errors (Borenstein et al., 2010). The method is useful, e.g., for pooling of survival data (using log hazard ratio and standard errors as input). Arguments TE and seTE can be used to provide treatment estimates and standard errors directly. However, it is possible to derive these quantities from other information.

Default settings are utilised for several arguments (assignments using gs function). These defaults can be changed for the current R session using the settings.meta function.

Furthermore, R function update.meta can be used to rerun a meta-analysis with different settings.

Approximate treatment estimates

Missing treatment estimates can be derived from

confidence limits provided by arguments lower and upper;

median, interquartile range and range (arguments median, q1, q3, min, and max);

median and interquartile range (arguments median, q1 and q3);

median and range (arguments median, min and max).

For confidence limits, the treatment estimate is defined as the center of the confidence interval (on the log scale for relative effect measures like the odds ratio or hazard ratio).

If the treatment effect is a mean it can be approximated from sample size, median, interquartile range and range. By default, methods described in Luo et al. (2018) are utilized (argument method.mean = "Luo"):

equation (7) if sample size, median and range are available,

equation (11) if sample size, median and interquartile range are available,

equation (15) if sample size, median, range and interquartile range are available.

Instead the methods described in Wan et al. (2014) are used if argument method.mean = "Wan"):

equation (2) if sample size, median and range are available,

equation (14) if sample size, median and interquartile range are available,

equation (10) if sample size, median, range and interquartile range are available.

By default, missing treatment estimates are replaced successively using these method, i.e., confidence limits are utilised before interquartile ranges. Argument approx.TE can be used to overwrite this default for each individual study:

Use treatment estimate directly (entry "" in argument approx.TE);

confidence limits ("ci" in argument approx.TE);

median, interquartile range and range ("iqr.range");

median and interquartile range ("iqr");

median and range ("range").

Approximate standard errors

Missing standard errors can be derived from

p-value provided by arguments pval and (optional) df;

confidence limits (arguments lower, upper, and (optional) df);

sample size, median, interquartile range and range (arguments n.e and / or n.c, median, q1, q3, min, and max);

sample size, median and interquartile range (arguments n.e and / or n.c, median, q1 and q3);

sample size, median and range (arguments n.e and / or n.c, median, min and max).

For p-values and confidence limits, calculations are either based on the standard normal or t distribution if argument df is provided. Furthermore, argument level.ci can be used to provide the level of the confidence interval.

Wan et al. (2014) describe methods to estimate the standard deviation (and thus the standard error by deviding the standard deviation with the square root of the sample size) from the sample size, median and additional statistics. Shi et al. (2020) provide an improved estimate of the standard deviation if the interquartile range and range are available in addition to the sample size and median. Accordingly, equation (11) in Shi et al. (2020) is the default (argument method.sd = "Shi"), if the median, interquartile range and range are provided (arguments median, q1, q3, min and max). The method by Wan et al. (2014) is used if argument method.sd = "Wan" and, depending on the sample size, either equation (12) or (13) is used. If only the interquartile range or range is available, equations (15) / (16) and (7) / (9) in Wan et al. (2014) are used, respectively. The sample size of individual studies must be provided with arguments n.e and / or n.c. The total sample size is calculated as n.e + n.c if both arguments are provided.

By default, missing standard errors are replaced successively using these method, e.g., p-value before confidence limits before interquartile range and range. Argument approx.seTE can be used to overwrite this default for each individual study:

Use standard error directly (entry "" in argument approx.seTE);

p-value ("pval" in argument approx.seTE);

confidence limits ("ci");

median, interquartile range and range ("iqr.range");

median and interquartile range ("iqr");

median and range ("range").

Confidence intervals for individual studies

For the mean difference (argument sm = "MD"), the confidence interval for individual studies can be based on the

standard normal distribution (method.ci = "z"), or

t-distribution (method.ci = "t").

By default, the first method is used if argument df is missing and the second method otherwise.

Note, this choice does not affect the results of the fixed effect and random effects meta-analysis.

Estimation of between-study variance

The following methods are available to estimate the between-study variance τ^2.

Argument	 Method

method.tau = "DL" 	 DerSimonian-Laird estimator (DerSimonian and Laird, 1986)

method.tau = "PM" 	 Paule-Mandel estimator (Paule and Mandel, 1982)

method.tau = "REML" 	 Restricted maximum-likelihood estimator (Viechtbauer, 2005)

method.tau = "ML" 	 Maximum-likelihood estimator (Viechtbauer, 2005)

method.tau = "HS" 	 Hunter-Schmidt estimator (Hunter and Schmidt, 2015)

method.tau = "SJ" 	 Sidik-Jonkman estimator (Sidik and Jonkman, 2005)

method.tau = "HE" 	 Hedges estimator (Hedges and Olkin, 1985)

method.tau = "EB" 	 Empirical Bayes estimator (Morris, 1983)

Historically, the DerSimonian-Laird method was the de facto standard to estimate the between-study variance τ^2 and is still the default in many software packages including Review Manager 5 (RevMan 5) and R package meta. However, its role has been challenged and especially the Paule-Mandel and REML estimators have been recommended (Veroniki et al., 2016). Accordingly, the following R command can be used to use the Paule-Mandel estimator in all meta-analyses of the R session: settings.meta(method.tau = "PM")

Confidence interval for the between-study variance

The following methods to calculate a confidence interval for τ^2 and τ are available.

Argument	 Method

method.tau.ci = "J"	 Method by Jackson (2013)

method.tau.ci = "BJ"	 Method by Biggerstaff and Jackson (2008)

method.tau.ci = "QP"	 Q-Profile method (Viechtbauer, 2007)

method.tau.ci = "PL"	 Profile-Likelihood method for three-level meta-analysis model

(Van den Noortgate et al., 2013)

The first three methods have been recommended by Veroniki et al. (2016). By default, the Jackson method is used for the DerSimonian-Laird estimator of τ^2 and the Q-profile method for all other estimators of τ^2. The Profile-Likelihood method is the only method available for the three-level meta-analysis model. No confidence intervals for τ^2 and τ are calculated if method.tau.ci = "".

Hartung-Knapp method

Hartung and Knapp (2001a,b) proposed an alternative method for random effects meta-analysis based on a refined variance estimator for the treatment estimate. Simulation studies (Hartung and Knapp, 2001a,b; IntHout et al., 2014; Langan et al., 2019) show improved coverage probabilities compared to the classic random effects method. However, in rare settings with very homogeneous treatment estimates, the Hartung-Knapp (HK) variance estimate can be arbitrarily small resulting in a very narrow confidence interval (Knapp and Hartung, 2003; Wiksten et al., 2016). In such cases, an ad hoc variance correction has been proposed by utilising the variance estimate from the classic random effects model with the HK method (Knapp and Hartung, 2003; IQWiQ, 2020). An alternative approach is to use the wider confidence interval of classic fixed or random effects meta-analysis and the HK method (Wiksten et al., 2016; Jackson et al., 2017).

Argument adhoc.hakn can be used to choose the ad hoc method:

Argument	 Ad hoc method

adhoc.hakn = ""	 not used

adhoc.hakn = "se"	 use variance correction if HK standard error is smaller

than standard error from classic random effects

meta-analysis (Knapp and Hartung, 2003)

adhoc.hakn = "iqwig6"	 use variance correction if HK confidence interval

is narrower than CI from classic random effects model

with DerSimonian-Laird estimator (IQWiG, 2020)

adhoc.hakn = "ci"	 use wider confidence interval of classic random effects

and HK meta-analysis

(Hybrid method 2 in Jackson et al., 2017)

Prediction interval

A prediction interval for the treatment effect of a new study (Higgins et al., 2009) is calculated if arguments prediction and comb.random are TRUE. Note, the definition of prediction intervals varies in the literature. This function implements equation (12) of Higgins et al., (2009) which proposed a t distribution with K-2 degrees of freedom where K corresponds to the number of studies in the meta-analysis.

Subgroup analysis

Argument byvar can be used to conduct subgroup analysis for a categorical covariate. The metareg function can be used instead for more than one categorical covariate or continuous covariates.

Specify the null hypothesis of test for an overall effect

Argument null.effect can be used to specify the (treatment) effect under the null hypothesis in a test for an overall effect.

By default (null.effect = 0), the null hypothesis corresponds to "no difference" (which is obvious for absolute effect measures like the mean difference (sm = "MD") or standardised mean difference (sm = "SMD")). For relative effect measures, e.g., risk ratio (sm = "RR") or odds ratio (sm = "OR"), the null effect is defined on the log scale, i.e., ln(RR) = 0 or ln(OR) = 0 which is equivalent to testing RR = 1 or OR = 1.

Use of argument null.effect is especially useful for summary measures without a "natural" null effect, i.e., in situations without a second (treatment) group. For example, an overall proportion of 50% could be tested in the meta-analysis of single proportions with argument null.effect = 0.5.

Note, all tests for an overall effect are two-sided with the alternative hypothesis that the effect is unequal to null.effect.

Exclusion of studies from meta-analysis

Arguments subset and exclude can be used to exclude studies from the meta-analysis. Studies are removed completely from the meta-analysis using argument subset, while excluded studies are shown in printouts and forest plots using argument exclude (see Examples). Meta-analysis results are the same for both arguments.

Presentation of meta-analysis results

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if argument comb.random = FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. For example, functions print.meta and forest.meta will not show results for the random effects model if comb.random = FALSE.

Argument pscale can be used to rescale single proportions or risk differences, e.g. pscale = 1000 means that proportions are expressed as events per 1000 observations. This is useful in situations with (very) low event probabilities.

Argument irscale can be used to rescale single rates or rate differences, e.g. irscale = 1000 means that rates are expressed as events per 1000 time units, e.g. person-years. This is useful in situations with (very) low rates. Argument irunit can be used to specify the time unit used in individual studies (default: "person-years"). This information is printed in summaries and forest plots if argument irscale is not equal to 1.

Default settings for comb.fixed, comb.random, pscale, irscale, irunit and several other arguments can be set for the whole R session using settings.meta.

Value

An object of class c("metagen", "meta") with corresponding print, summary, and forest functions. The object is a list containing the following components:

TE, seTE, studlab, exclude, n.e, n.c
As defined above.

id, sm, method.ci, level, level.comb,
As defined above.

comb.fixed, comb.random,
As defined above.

overall, overall.hetstat,
As defined above.

hakn, adhoc.hakn, method.tau, method.tau.ci,
As defined above.

tau.preset, TE.tau, method.bias,
As defined above.

tau.common, title, complab, outclab,
As defined above.

label.e, label.c, label.left, label.right,
As defined above.

byvar, bylab, print.byvar, byseparator, warn
As defined above.

lower, upper
Lower and upper confidence interval limits for individual studies.

statistic, pval
Statistic and p-value for test of treatment effect for individual studies.

w.fixed, w.random
Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed
Estimated overall treatment effect and standard error (fixed effect model).

lower.fixed, upper.fixed
Lower and upper confidence interval limits (fixed effect model).

statistic.fixed, pval.fixed
Statistic and p-value for test of overall treatment effect (fixed effect model).

TE.random, seTE.random
Estimated overall treatment effect and standard error (random effects model).

lower.random, upper.random
Lower and upper confidence interval limits (random effects model).

statistic.random, pval.random
Statistic and p-value for test of overall treatment effect (random effects model).

prediction, level.predict
As defined above.

seTE.predict
Standard error utilised for prediction interval.

lower.predict, upper.predict
Lower and upper limits of prediction interval.

null.effect
As defined above.

k
Number of studies combined in meta-analysis.

Q
Heterogeneity statistic.

df.Q
Degrees of freedom for heterogeneity statistic.

pval.Q
P-value of heterogeneity test.

tau2
Between-study variance τ^2.

se.tau2
Standard error of τ^2.

lower.tau2, upper.tau2
Lower and upper limit of confidence interval for τ^2.

tau
Square-root of between-study variance τ.

lower.tau, upper.tau
Lower and upper limit of confidence interval for τ.

H
Heterogeneity statistic H.

lower.H, upper.H
Lower and upper confidence limit for heterogeneity statistic H.

I2
Heterogeneity statistic I^2.

lower.I2, upper.I2
Lower and upper confidence limit for heterogeneity statistic I^2.

Rb
Heterogeneity statistic R_b.

lower.Rb, upper.Rb
Lower and upper confidence limit for heterogeneity statistic R_b.

approx.TE, approx.seTE
As defined above.

method
Pooling method: "Inverse".

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn = TRUE).

bylevs
Levels of grouping variable - if byvar is not missing.

TE.fixed.w, seTE.fixed.w
Estimated treatment effect and standard error in subgroups (fixed effect model) - if byvar is not missing.

lower.fixed.w, upper.fixed.w
Lower and upper confidence interval limits in subgroups (fixed effect model) - if byvar is not missing.

statistic.fixed.w, pval.fixed.w
Statistics and p-values for test of treatment effect in subgroups (fixed effect model) - if byvar is not missing.

TE.random.w, seTE.random.w
Estimated treatment effect and standard error in subgroups (random effects model) - if byvar is not missing.

lower.random.w, upper.random.w
Lower and upper confidence interval limits in subgroups (random effects model) - if byvar is not missing.

statistic.random.w, pval.random.w
Statistics and p-values for test of treatment effect in subgroups (random effects model) - if byvar is not missing.

w.fixed.w, w.random.w
Weight of subgroups (in fixed and random effects model) - if byvar is not missing.

df.hakn.w
Degrees of freedom for test of treatment effect for Hartung-Knapp method in subgroups - if byvar is not missing and hakn = TRUE.

n.harmonic.mean.w
Harmonic mean of number of observations in subgroups (for back transformation of Freeman-Tukey Double arcsine transformation) - if byvar is not missing.

n.e.w
Number of observations in experimental group in subgroups - if byvar is not missing.

n.c.w
Number of observations in control group in subgroups - if byvar is not missing.

k.w
Number of studies combined within subgroups - if byvar is not missing.

k.all.w
Number of all studies in subgroups - if byvar is not missing.

Q.w.fixed
Overall within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.w.random
Overall within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing (only calculated if argument tau.common is TRUE).

df.Q.w
Degrees of freedom for test of overall within subgroups heterogeneity - if byvar is not missing.

pval.Q.w.fixed
P-value of within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.w.random
P-value of within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

Q.b.fixed
Overall between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.b.random
Overall between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

df.Q.b
Degrees of freedom for test of overall between subgroups heterogeneity - if byvar is not missing.

pval.Q.b.fixed
P-value of between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.b.random
P-value of between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

tau.w
Square-root of between-study variance within subgroups - if byvar is not missing.

H.w
Heterogeneity statistic H within subgroups - if byvar is not missing.

lower.H.w, upper.H.w
Lower and upper confidence limit for heterogeneity statistic H within subgroups - if byvar is not missing.

I2.w
Heterogeneity statistic I^2 within subgroups - if byvar is not missing.

lower.I2.w, upper.I2.w
Lower and upper confidence limit for heterogeneity statistic I^2 within subgroups - if byvar is not missing.

keepdata
As defined above.

data
Original data (set) used in function call (if keepdata = TRUE).

subset
Information on subset of original data used in meta-analysis (if keepdata = TRUE).

call
Function call.

version
Version of R package meta used to create object.

Note

R function rma.uni from R package metafor (Viechtbauer 2010) is called internally to estimate the between-study variance τ^2.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Biggerstaff BJ, Jackson D (2008): The exact distribution of Cochran’s heterogeneity statistic in one-way random effects meta-analysis. Statistics in Medicine, 27, 6093–110

Borenstein M, Hedges LV, Higgins JP, Rothstein HR (2010): A basic introduction to fixed-effect and random-effects models for meta-analysis. Research Synthesis Methods, 1, 97–111

Cooper H & Hedges LV (1994): The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation

DerSimonian R & Laird N (1986): Meta-analysis in clinical trials. Controlled Clinical Trials, 7, 177–88

Hedges LV & Olkin I (1985): Statistical methods for meta-analysis. San Diego, CA: Academic Press

Higgins JPT, Thompson SG, Spiegelhalter DJ (2009): A re-evaluation of random-effects meta-analysis. Journal of the Royal Statistical Society: Series A, 172, 137–59

Hunter JE & Schmidt FL (2015): Methods of Meta-Analysis: Correcting Error and Bias in Research Findings (Third edition). Thousand Oaks, CA: Sage

Hartung J, Knapp G (2001a): On tests of the overall treatment effect in meta-analysis with normally distributed responses. Statistics in Medicine, 20, 1771–82

Hartung J, Knapp G (2001b): A refined method for the meta-analysis of controlled clinical trials with binary outcome. Statistics in Medicine, 20, 3875–89

IntHout J, Ioannidis JPA, Borm GF (2014): The Hartung-Knapp-Sidik-Jonkman method for random effects meta-analysis is straightforward and considerably outperforms the standard DerSimonian-Laird method. BMC Medical Research Methodology, 14, 25

IQWiG (2020): General Methods: Version 6.0. https://www.iqwig.de/en/about-us/methods/methods-paper/

Jackson D (2013): Confidence intervals for the between-study variance in random effects meta-analysis using generalised Cochran heterogeneity statistics. Research Synthesis Methods, 4, 220–229

Jackson D, Law M, Rücker G, Schwarzer G (2017): The Hartung-Knapp modification for random-effects meta-analysis: A useful refinement but are there any residual concerns? Statistics in Medicine, 36, 3923–34

Knapp G & Hartung J (2003): Improved tests for a random effects meta-regression with a single covariate. Statistics in Medicine, 22, 2693–710

Langan D, Higgins JPT, Jackson D, Bowden J, Veroniki AA, Kontopantelis E, et al. (2019): A comparison of heterogeneity variance estimators in simulated random-effects meta-analyses. Research Synthesis Methods, 10, 83–98

Luo D, Wan X, Liu J, Tong T (2018): Optimally estimating the sample mean from the sample size, median, mid-range, and/or mid-quartile range. Statistical Methods in Medical Research, 27, 1785–805

Morris CN (1983): Parametric empirical Bayes inference: Theory and applications (with discussion). Journal of the American Statistical Association 78, 47–65

Paule RC & Mandel J (1982): Consensus values and weighting factors. Journal of Research of the National Bureau of Standards, 87, 377–85

Review Manager (RevMan) [Computer program]. Version 5.4. The Cochrane Collaboration, 2020

Shi J, Luo D, Weng H, Zeng X-T, Lin L, Chu H, et al. (2020): Optimally estimating the sample standard deviation from the five-number summary. Research Synthesis Methods.

Sidik K & Jonkman JN (2005): Simple heterogeneity variance estimation for meta-analysis. Journal of the Royal Statistical Society: Series C (Applied Statistics), 54, 367–84

Veroniki AA, Jackson D, Viechtbauer W, Bender R, Bowden J, Knapp G, et al. (2016): Methods to estimate the between-study variance and its uncertainty in meta-analysis. Research Synthesis Methods, 7, 55–79

Viechtbauer W (2005): Bias and efficiency of meta-analytic variance estimators in the random-effects model. Journal of Educational and Behavioral Statistics, 30, 261–93

Viechtbauer W (2007): Confidence intervals for the amount of heterogeneity in meta-analysis. Statistics in Medicine, 26, 37–52

Viechtbauer W (2010): Conducting Meta-Analyses in R with the metafor Package. Journal of Statistical Software, 36, 1–48

Van den Noortgate W, López-López JA, Marín-Martínez F, Sánchez-Meca J (2013): Three-level meta-analysis of dependent effect sizes. Behavior Research Methods, 45, 576–94

Wan X, Wang W, Liu J, Tong T (2014): Estimating the sample mean and standard deviation from the sample size, median, range and/or interquartile range. BMC Medical Research Methodology, 14, 135

Wiksten A, Rücker G, Schwarzer G (2016): Hartung-Knapp method is not always conservative compared with fixed-effect meta-analysis. Statistics in Medicine, 35, 2503–15

See Also

update.meta, metabin, metacont, print.meta, settings.meta

Examples

data(Fleiss1993bin)
m1 <- metabin(d.asp, n.asp, d.plac, n.plac, study,
              data = Fleiss1993bin, sm = "RR", method = "I")
m1

# Identical results using the generic inverse variance method with
# log risk ratio and its standard error:
# Note, argument 'n.e' in metagen() is used to provide the total
# sample size which is calculated from the group sample sizes n.e
# and n.c in meta-analysis m1.
m1.gen <- metagen(TE, seTE, studlab, n.e = n.e + n.c, data = m1, sm = "RR")
m1.gen
forest(m1.gen, leftcols = c("studlab", "n.e", "TE", "seTE"))

# Meta-analysis with prespecified between-study variance
#
summary(metagen(m1$TE, m1$seTE, sm = "RR", tau.preset = sqrt(0.1)))

# Meta-analysis of survival data:
#
logHR <- log(c(0.95, 1.5))
selogHR <- c(0.25, 0.35)
metagen(logHR, selogHR, sm = "HR")

# Paule-Mandel method to estimate between-study variance for data
# from Paule & Mandel (1982)
#
average <- c(27.044, 26.022, 26.340, 26.787, 26.796)
variance <- c(0.003, 0.076, 0.464, 0.003, 0.014)
#
summary(metagen(average, sqrt(variance), sm = "MD", method.tau = "PM"))

# Conduct meta-analysis using hazard ratios and 95% confidence intervals
#
# Data from Steurer et al. (2006), Analysis 1.1 Overall survival
# https://www.cochranelibrary.com/cdsr/doi/10.1002/14651858.CD004270.pub2/abstract
#
study <- c("FCG on CLL 1996", "Leporrier 2001", "Rai 2000", "Robak 2000")
HR <- c(0.55, 0.92, 0.79, 1.18)
lower.HR <- c(0.28, 0.79, 0.59, 0.64)
upper.HR <- c(1.09, 1.08, 1.05, 2.17)
#
# Input must be log hazard ratios, not hazard ratios
#
metagen(log(HR), lower = log(lower.HR), upper = log(upper.HR),
        studlab = study, sm = "HR")

# Exclude MRC-1 and MRC-2 studies from meta-analysis, however,
# show them in printouts and forest plots
#
metabin(d.asp, n.asp, d.plac, n.plac, study,
        data = Fleiss1993bin, sm = "RR", method = "I",
        exclude = study %in% c("MRC-1", "MRC-2"))
#
# Exclude MRC-1 and MRC-2 studies completely from meta-analysis
#
metabin(d.asp, n.asp, d.plac, n.plac, study,
        data = Fleiss1993bin, sm = "RR", method = "I",
        subset = !(study %in% c("MRC-1", "MRC-2")))

# Exclude studies with total sample size above 1500
#
metabin(d.asp, n.asp, d.plac, n.plac, study,
        data = Fleiss1993bin, sm = "RR", method = "I",
        exclude = (n.asp + n.plac) > 1500)

# Exclude studies containing "MRC" in study name
#
metabin(d.asp, n.asp, d.plac, n.plac, study,
        data = Fleiss1993bin, sm = "RR", method = "I",
        exclude = grep("MRC", study))

# Use both arguments 'subset' and 'exclude'
#
metabin(d.asp, n.asp, d.plac, n.plac, study,
        data = Fleiss1993bin, sm = "RR", method = "I",
        subset = (n.asp + n.plac) > 1500,
        exclude = grep("MRC", study))

[Package meta version 4.18-1 Index]