metainf {meta}	R Documentation

Influence analysis in meta-analysis

Description

Performs a influence analysis. Pooled estimates are calculated omitting one study at a time.

Usage

metainf(x, pooled, sortvar, level=x$level, level.comb=x$level.comb)

Arguments

x 	 An object of class meta.

pooled 	 A character string indicating whether a fixed effect or random effects model is used for pooling. Either missing (see Details), "fixed" or "random", can be abbreviated.

sortvar 	 An optional vector used to sort the individual studies (must be of same length as x$TE).

level 	 The level used to calculate confidence intervals for individual studies.

level.comb 	 The level used to calculate confidence intervals for pooled estimates.

Details

Performs a influence analysis; pooled estimates are calculated omitting one study at a time. Studies are sorted according to sortvar.

Information from object x is utilised if argument pooled is missing. A fixed effect model is assumed (pooled="fixed") if parameter x$comb.fixed is TRUE; a random effects model is assumed (pooled="random") if parameter x$comb.random is TRUE and x$comb.fixed is FALSE.

Value

An object of class c("metainf", "meta") with corresponding print, plot function. The object is a list containing the following components:

TE, seTE 	 Estimated treatment effect and standard error of pooled estimate in influence analysis.

studlab 	 Study label describing omission of studies.

p.value 	 P-value for test of overall effect.

w 	 Sum of weights from fixed effect or random effects model.

I2 	 Heterogeneity statistic I2.

tau 	 Square-root of between-study variance.

sm 	 Summary measure.

method 	 Method used for pooling.

k 	 Number of studies combined in meta-analysis.

pooled 	 As defined above.

TE.fixed, seTE.fixed 	 Value is NA.

TE.random, seTE.random 	 Value is NA.

Q 	 Value is NA.

tau 	 Value is NA.

level 	 The level used to calculate confidence intervals for individual studies.

level.comb 	 The level used to calculate confidence intervals for pooled estimates.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994), The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation.

See Also

metabin, metacont, print.meta

Examples

data(Fleiss93)
meta1 <- metabin(event.e, n.e, event.c, n.c,
                 data=Fleiss93, studlab=study,
                 sm="RR", meth="I")
meta1

metainf(meta1)
metainf(meta1, pooled="random")

grid.newpage()
forest(metainf(meta1, pooled="random"), comb.random=TRUE)

[Package meta version 1.1-8 Index]