metainf {meta}	R Documentation

Influence analysis in meta-analysis using leave-one-out method

Description

Performs an influence analysis. Pooled estimates are calculated omitting one study at a time.

Usage

metainf(x, pooled, sortvar)

Arguments

x
An object of class meta.

pooled
A character string indicating whether a fixed effect or random effects model is used for pooling. Either missing (see Details), "fixed" or "random", can be abbreviated.

sortvar
An optional vector used to sort the individual studies (must be of same length as x$TE).

Details

Performs a influence analysis; pooled estimates are calculated omitting one study at a time. Studies are sorted according to sortvar.

Information from object x is utilised if argument pooled is missing. A fixed effect model is assumed (pooled="fixed") if argument x$comb.fixed is TRUE; a random effects model is assumed (pooled="random") if argument x$comb.random is TRUE and x$comb.fixed is FALSE.

Value

An object of class c("metainf", "meta") with corresponding print, and forest functions. The object is a list containing the following components:

TE, seTE
Estimated treatment effect and standard error of pooled estimate in influence analysis.

lower, upper
Lower and upper confidence interval limits.

statistic
Statistic for test of overall effect.

pval
P-value for test of overall effect.

studlab
Study label describing omission of studies.

w
Sum of weights from fixed effect or random effects model.

I2
Heterogeneity statistic I^2.

Rb
Heterogeneity statistic R_b.

tau
Square-root of between-study variance.

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn = TRUE).

sm
Summary measure.

method
Method used for pooling.

k
Number of studies combined in meta-analysis.

pooled
As defined above.

comb.fixed
A logical indicating whether analysis is based on fixed effect model.

comb.random
A logical indicating whether analysis is based on random effects model.

TE.fixed, seTE.fixed
Value is NA.

TE.random, seTE.random
Value is NA.

Q
Value is NA.

level.comb
The level used to calculate confidence intervals for pooled estimates.

hakn
A logical indicating whether the method by Hartung and Knapp is used to adjust test statistics and confidence intervals.

adhoc.hakn
A character string indicating whether ad hoc variance correction should be used for Hartung-Knapp method.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2.

tau.preset
Prespecified value for the square root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance τ^2.

n.harmonic.mean
Harmonic mean of number of observations (for back transformation of Freeman-Tukey Double arcsine transformation).

version
Version of R package meta used to create object.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994): The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation

See Also

metabin, metacont, print.meta

Examples

data(Fleiss1993bin)
m1 <- metabin(d.asp, n.asp, d.plac, n.plac,
              data = Fleiss1993bin, studlab = study,
              sm = "RR", method = "I")
m1
metainf(m1)
metainf(m1, pooled = "random")

forest(metainf(m1))
forest(metainf(m1), layout = "revman5")
forest(metainf(m1, pooled = "random"))

metainf(m1, sortvar = study)
metainf(m1, sortvar = 7:1)

m2 <- update(m1, title = "Fleiss1993bin meta-analysis",
             backtransf = FALSE)
metainf(m2)

data(Fleiss1993cont)
m3 <- metacont(n.psyc, mean.psyc, sd.psyc, n.cont, mean.cont, sd.cont,
               data = Fleiss1993cont, sm = "SMD")
metainf(m3)

[Package meta version 4.18-1 Index]