metaprop {meta}	R Documentation

Meta-analysis of single proportions

Description

Calculation of an overall proportion from studies reporting a single proportion.

Usage

metaprop(event, n, studlab,
         data = NULL, subset = NULL,
         freeman.tukey=TRUE,
         level = 0.95, level.comb = level,
         comb.fixed=TRUE, comb.random=TRUE,
         title="", complab="", outclab="",
         byvar, bylab, print.byvar=TRUE)

Arguments

event 	 Number of events.

n 	 Number of observations.

studlab 	 An optional vector with study labels.

data 	 An optional data frame containing the study information, i.e., event and n.

subset 	 An optional vector specifying a subset of studies to be used.

freeman.tukey 	 A logical indicating if the Freeman-Tukey Double arcsine transformation should be used; otherwise the arcsine transformation is used.

level 	 The level used to calculate confidence intervals for individual studies.

level.comb 	 The level used to calculate confidence intervals for pooled estimates.

comb.fixed 	 A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random 	 A logical indicating whether a random effects meta-analysis should be conducted.

title 	 Title of meta-analysis / systematic review.

complab 	 Comparison label.

outclab 	 Outcome label.

byvar 	 An optional vector containing grouping information (must be of same length as event.e).

bylab 	 A character string with a label for the grouping variable.

print.byvar 	 A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

Details

Fixed effect and random effects meta-analysis of single proportions using either the Freeman-Tukey Double arcsine transformation or the arcsine transformation of proportions to calculate an overall proportion.

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if comb.random=FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random=FALSE.

Value

An object of class c("metaprop", "meta") with corresponding print, summary, plot function. The object is a list containing the following components:

event, n, studlab,

freeman.tukey, level, level.comb,

comb.fixed, comb.random,

byvar, bylab, print.byvar 	 As defined above.

TE, seTE 	 Arcsine transformation of proportion and its standard error for individual studies.

w.fixed, w.random 	 Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed 	 Estimated overall arcsine transformed proportion and standard error (fixed effect model).

TE.random, seTE.random 	 Estimated overall arcsine transformed proportion and standard error (random effects model).

k 	 Number of studies combined in meta-analysis.

Q 	 Heterogeneity statistic Q.

tau 	 Square-root of between-study variance (moment estimator of DerSimonian-Laird).

sm 	 A character string: "proportion"

method 	 A character string indicating method used for pooling: "Inverse"

call 	 Function call.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Freeman MF & Tukey JW (1950), Transformations related to the angular and the square root. Annals of Mathematical Statistics, 21, 607–611.

Miller JJ (1978), The inverse of the Freeman-Tukey double arcsine transformation. The American Statistician, 32, 138.

See Also

metacont, metagen, print.meta

Examples

metaprop(0, 10)
metaprop(0, 10, freeman.tukey=FALSE)

[Package meta version 1.1-8 Index]