metaprop {meta}	R Documentation

Meta-analysis of single proportions

Description

Calculation of an overall proportion from studies reporting a single proportion. Inverse variance method and generalised linear mixed model (GLMM) are available for pooling. For GLMMs, the rma.glmm function from R package metafor (Viechtbauer 2010) is called internally.

Usage

metaprop(
  event,
  n,
  studlab,
  data = NULL,
  subset = NULL,
  exclude = NULL,
  method,
  sm = gs("smprop"),
  incr = gs("incr"),
  allincr = gs("allincr"),
  addincr = gs("addincr"),
  method.ci = gs("method.ci.prop"),
  level = gs("level"),
  level.comb = gs("level.comb"),
  comb.fixed = gs("comb.fixed"),
  comb.random = gs("comb.random"),
  overall = comb.fixed | comb.random,
  overall.hetstat = comb.fixed | comb.random,
  hakn = gs("hakn"),
  adhoc.hakn = gs("adhoc.hakn"),
  method.tau,
  method.tau.ci = gs("method.tau.ci"),
  tau.preset = NULL,
  TE.tau = NULL,
  tau.common = gs("tau.common"),
  prediction = gs("prediction"),
  level.predict = gs("level.predict"),
  null.effect = NA,
  method.bias = gs("method.bias"),
  backtransf = gs("backtransf"),
  pscale = 1,
  text.fixed = gs("text.fixed"),
  text.random = gs("text.random"),
  text.predict = gs("text.predict"),
  text.w.fixed = gs("text.w.fixed"),
  text.w.random = gs("text.w.random"),
  title = gs("title"),
  complab = gs("complab"),
  outclab = "",
  byvar,
  bylab,
  print.byvar = gs("print.byvar"),
  byseparator = gs("byseparator"),
  keepdata = gs("keepdata"),
  warn = gs("warn"),
  control = NULL,
  ...
)

Arguments

event
Number of events.

n
Number of observations.

studlab
An optional vector with study labels.

data
An optional data frame containing the study information, i.e., event and n.

subset
An optional vector specifying a subset of studies to be used.

exclude
An optional vector specifying studies to exclude from meta-analysis, however, to include in printouts and forest plots.

method
A character string indicating which method is to be used for pooling of studies. One of "Inverse" and "GLMM", can be abbreviated.

sm
A character string indicating which summary measure ("PLOGIT", "PAS", "PFT", "PLN", or "PRAW") is to be used for pooling of studies, see Details.

incr
A numeric which is added to event number and sample size of studies with zero or all events, i.e., studies with an event probability of either 0 or 1.

allincr
A logical indicating if incr is considered for all studies if at least one study has either zero or all events. If FALSE (default), incr is considered only in studies with zero or all events.

addincr
A logical indicating if incr is used for all studies irrespective of number of events.

method.ci
A character string indicating which method is used to calculate confidence intervals for individual studies, see Details.

level
The level used to calculate confidence intervals for individual studies.

level.comb
The level used to calculate confidence intervals for pooled estimates.

comb.fixed
A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random
A logical indicating whether a random effects meta-analysis should be conducted.

overall
A logical indicating whether overall summaries should be reported. This argument is useful in a meta-analysis with subgroups if overall results should not be reported.

overall.hetstat
A logical value indicating whether to print heterogeneity measures for overall treatment comparisons. This argument is useful in a meta-analysis with subgroups if heterogeneity statistics should only be printed on subgroup level.

hakn
A logical indicating whether the method by Hartung and Knapp should be used to adjust test statistics and confidence intervals.

adhoc.hakn
A character string indicating whether an ad hoc variance correction should be applied in the case of an arbitrarily small Hartung-Knapp variance estimate, see Details.

method.tau
A character string indicating which method is used to estimate the between-study variance τ^2 and its square root τ. Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.

method.tau.ci
A character string indicating which method is used to estimate the confidence interval of τ^2 and τ. Either "QP", "BJ", or "J", or "", can be abbreviated.

tau.preset
Prespecified value for the square root of the between-study variance τ^2.

TE.tau
Overall treatment effect used to estimate the between-study variance tau-squared.

tau.common
A logical indicating whether tau-squared should be the same across subgroups.

prediction
A logical indicating whether a prediction interval should be printed.

level.predict
The level used to calculate prediction interval for a new study.

null.effect
A numeric value specifying the effect under the null hypothesis.

method.bias
A character string indicating which test is to be used. Either "Begg", "Egger", or "Thompson", can be abbreviated. See function metabias.

backtransf
A logical indicating whether results for transformed proportions (argument sm != "PRAW") should be back transformed in printouts and plots. If TRUE (default), results will be presented as proportions; otherwise transformed proportions will be shown. See Details for presentation of confidence intervals.

pscale
A numeric defining a scaling factor for printing of single event probabilities.

text.fixed
A character string used in printouts and forest plot to label the pooled fixed effect estimate.

text.random
A character string used in printouts and forest plot to label the pooled random effects estimate.

text.predict
A character string used in printouts and forest plot to label the prediction interval.

text.w.fixed
A character string used to label weights of fixed effect model.

text.w.random
A character string used to label weights of random effects model.

title
Title of meta-analysis / systematic review.

complab
Comparison label.

outclab
Outcome label.

byvar
An optional vector containing grouping information (must be of same length as event).

bylab
A character string with a label for the grouping variable.

print.byvar
A logical indicating whether the name of the grouping variable should be printed in front of the group labels.

byseparator
A character string defining the separator between label and levels of grouping variable.

keepdata
A logical indicating whether original data (set) should be kept in meta object.

warn
A logical indicating whether the addition of incr to studies with zero or all events should result in a warning.

control
An optional list to control the iterative process to estimate the between-study variance τ^2. This argument is passed on to rma.uni or rma.glmm, respectively.

...
Additional arguments passed on to rma.glmm function.

Details

This function provides methods for fixed effect and random effects meta-analysis of single proportions to calculate an overall proportion. Note, you should use R function metabin to compare proportions of pairwise comparisons instead of using metaprop for each treatment arm separately which will break randomisation in randomised controlled trials.

The following transformations of proportions are implemented to calculate an overall proportion:

Logit transformation (sm = "PLOGIT", default)

Arcsine transformation (sm = "PAS")

Freeman-Tukey Double arcsine transformation (sm = "PFT")

Log transformation (sm = "PLN")

Raw, i.e. untransformed, proportions (sm = "PRAW")

A generalised linear mixed model (GLMM) - more specific, a random intercept logistic regression model - can be utilised for the meta-analysis of proportions (Stijnen et al., 2010). This is the default method for the logit transformation (argument sm = "PLOGIT"). Internally, the rma.glmm function from R package metafor is called to fit a GLMM.

Classic meta-analysis (Borenstein et al., 2010) utilising the (un)transformed proportions and corresponding standard errors in the inverse variance method is conducted by calling the metagen function internally. This is the only available method for all transformations but the logit transformation. The classic meta-analysis model with logit transformed proportions is used by setting argument method = "Inverse".

Default settings are utilised for several arguments (assignments using gs function). These defaults can be changed for the current R session using the settings.meta function.

Furthermore, R function update.meta can be used to rerun a meta-analysis with different settings.

Choice of transformation / meta-analysis method

Contradictory recommendations on the use of transformations of proportions have been published in the literature. For example, Barendregt et al. (2013) recommend the use of the Freeman-Tukey double arcsine transformation instead of the logit transformation whereas Warton & Hui (2011) strongly advise to use generalised linear mixed models with the logit transformation instead of the arcsine transformation.

Schwarzer et al. (2019) describe seriously misleading results in a meta-analysis with very different sample sizes due to problems with the back-transformation of the Freeman-Tukey transformation which requires a single sample size (Miller, 1978). Accordingly, Schwarzer et al. (2019) also recommend to use GLMMs for the meta-analysis of single proportions, however, admit that individual study weights are not available with this method. Meta-analysts which require individual study weights should consider the inverse variance method with the arcsine or logit transformation.

In order to prevent misleading conclusions for the Freeman-Tukey double arcsine transformation, sensitivity analyses using other transformations or using a range of sample sizes should be conducted (Schwarzer et al., 2019).

Continuity correction

If the summary measure is equal to "PLOGIT", "PLN", or "PRAW", a continuity correction is applied if any study has either zero or all events, i.e., an event probability of either 0 or 1.

By default, 0.5 is used as continuity correction (argument incr). This continuity correction is used both to calculate individual study results with confidence limits and to conduct meta-analysis based on the inverse variance method. For GLMMs no continuity correction is used.

Confidence intervals for individual studies

Various methods are available to calculate confidence intervals for individual study results (see Agresti & Coull 1998 and Newcombe 1988):

Clopper-Pearson interval also called 'exact' binomial interval (method.ci = "CP", default)

Wilson Score interval (method.ci = "WS")

Wilson Score interval with continuity correction (method.ci = "WSCC")

Agresti-Coull interval (method.ci = "AC")

Simple approximation interval (method.ci = "SA")

Simple approximation interval with continuity correction (method.ci = "SACC")

Normal approximation interval based on summary measure, i.e. defined by argument sm (method.ci = "NAsm")

Note, with exception of the normal approximation based on the summary measure, i.e. method.ci = "NAsm", the same confidence interval is calculated for individual studies for any summary measure (argument sm) as only number of events and observations are used in the calculation disregarding the chosen transformation.

Results will be presented for transformed proportions if argument backtransf = FALSE in the print.meta, print.summary.meta, or forest.meta function. In this case, argument method.ci = "NAsm" is used, i.e. confidence intervals based on the normal approximation based on the summary measure.

Estimation of between-study variance

The following methods to estimate the between-study variance τ^2 are available for the inverse variance method:

DerSimonian-Laird estimator (method.tau = "DL")

Paule-Mandel estimator (method.tau = "PM")

Restricted maximum-likelihood estimator (method.tau = "REML")

Maximum-likelihood estimator (method.tau = "ML")

Hunter-Schmidt estimator (method.tau = "HS")

Sidik-Jonkman estimator (method.tau = "SJ")

Hedges estimator (method.tau = "HE")

Empirical Bayes estimator (method.tau = "EB")

See metagen for more information on these estimators. Note, the maximum-likelihood method is utilized for GLMMs.

Confidence interval for the between-study variance

The following methods to calculate a confidence interval for τ^2 and τ are available.

Argument	 Method

method.tau.ci = "J"	 Method by Jackson

method.tau.ci = "BJ"	 Method by Biggerstaff and Jackson

method.tau.ci = "QP"	 Q-Profile method

See metagen for more information on these methods. For GLMMs, no confidence intervals for τ^2 and τ are calculated. Likewise, no confidence intervals for τ^2 and τ are calculated if method.tau.ci = "".

Hartung-Knapp method

Hartung and Knapp (2001a,b) proposed an alternative method for random effects meta-analysis based on a refined variance estimator for the treatment estimate. Simulation studies (Hartung and Knapp, 2001a,b; IntHout et al., 2014; Langan et al., 2019) show improved coverage probabilities compared to the classic random effects method.

In rare settings with very homogeneous treatment estimates, the Hartung-Knapp variance estimate can be arbitrarily small resulting in a very narrow confidence interval (Knapp and Hartung, 2003; Wiksten et al., 2016). In such cases, an ad hoc variance correction has been proposed by utilising the variance estimate from the classic random effects model with the HK method (Knapp and Hartung, 2003; IQWiQ, 2020). An alternative approach is to use the wider confidence interval of classic fixed or random effects meta-analysis and the HK method (Wiksten et al., 2016; Jackson et al., 2017).

Argument adhoc.hakn can be used to choose the ad hoc method:

Argument	 Ad hoc method

adhoc.hakn = ""	 not used

adhoc.hakn = "se"	 use variance correction if HK standard error is smaller

than standard error from classic random effects

meta-analysis (Knapp and Hartung, 2003)

adhoc.hakn = "iqwig6"	 use variance correction if HK confidence interval

is narrower than CI from classic random effects model

with DerSimonian-Laird estimator (IQWiG, 2020)

adhoc.hakn = "ci"	 use wider confidence interval of classic random effects

and HK meta-analysis

(Hybrid method 2 in Jackson et al., 2017)

For GLMMs, a method similar to Knapp and Hartung (2003) is implemented, see description of argument tdist in rma.glmm, and the ad hoc variance correction is not available.

Prediction interval

A prediction interval for the proportion in a new study (Higgins et al., 2009) is calculated if arguments prediction and comb.random are TRUE. Note, the definition of prediction intervals varies in the literature. This function implements equation (12) of Higgins et al., (2009) which proposed a t distribution with K-2 degrees of freedom where K corresponds to the number of studies in the meta-analysis.

Subgroup analysis

Argument byvar can be used to conduct subgroup analysis for a categorical covariate. The metareg function can be used instead for more than one categorical covariate or continuous covariates.

Specify the null hypothesis of test for an overall proportion

Argument null.effect can be used to specify the proportion used under the null hypothesis in a test for an overall effect.

By default (null.effect = NA), no hypothesis test is conducted as it is unclear which value is a sensible choice for the data at hand. An overall proportion of 50%, for example, could be tested by setting argument null.effect = 0.5.

Note, all tests for an overall effect are two-sided with the alternative hypothesis that the effect is unequal to null.effect.

Exclusion of studies from meta-analysis

Arguments subset and exclude can be used to exclude studies from the meta-analysis. Studies are removed completely from the meta-analysis using argument subset, while excluded studies are shown in printouts and forest plots using argument exclude (see Examples in metagen). Meta-analysis results are the same for both arguments.

Presentation of meta-analysis results

Internally, both fixed effect and random effects models are calculated regardless of values choosen for arguments comb.fixed and comb.random. Accordingly, the estimate for the random effects model can be extracted from component TE.random of an object of class "meta" even if argument comb.random = FALSE. However, all functions in R package meta will adequately consider the values for comb.fixed and comb.random. E.g. function print.meta will not print results for the random effects model if comb.random = FALSE.

Argument pscale can be used to rescale proportions, e.g. pscale = 1000 means that proportions are expressed as events per 1000 observations. This is useful in situations with (very) low event probabilities.

Value

An object of class c("metaprop", "meta") with corresponding print, summary, and forest functions. The object is a list containing the following components:

event, n, studlab, exclude,
As defined above.

sm, incr, allincr, addincr, method.ci,
As defined above.

level, level.comb,
As defined above.

comb.fixed, comb.random,
As defined above.

overall, overall.hetstat,
As defined above.

hakn, adhoc.hakn, method.tau, method.tau.ci,
As defined above.

tau.preset, TE.tau, null.hypothesis,
As defined above.

method.bias, tau.common, title, complab, outclab,
As defined above.

byvar, bylab, print.byvar, byseparator, warn
As defined above.

TE, seTE
Estimated (un)transformed proportion and its standard error for individual studies.

lower, upper
Lower and upper confidence interval limits for individual studies.

zval, pval
z-value and p-value for test of treatment effect for individual studies.

w.fixed, w.random
Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed
Estimated overall (un)transformed proportion and standard error (fixed effect model).

lower.fixed, upper.fixed
Lower and upper confidence interval limits (fixed effect model).

statistic.fixed, pval.fixed
z-value and p-value for test of overall effect (fixed effect model).

TE.random, seTE.random
Estimated overall (un)transformed proportion and standard error (random effects model).

lower.random, upper.random
Lower and upper confidence interval limits (random effects model).

statistic.random, pval.random
z-value or t-value and corresponding p-value for test of overall effect (random effects model).

prediction, level.predict
As defined above.

seTE.predict
Standard error utilised for prediction interval.

lower.predict, upper.predict
Lower and upper limits of prediction interval.

k
Number of studies combined in meta-analysis.

Q
Heterogeneity statistic Q.

df.Q
Degrees of freedom for heterogeneity statistic.

pval.Q
P-value of heterogeneity test.

Q.LRT
Heterogeneity statistic for likelihood-ratio test (only if method = "GLMM").

df.Q.LRT
Degrees of freedom for likelihood-ratio test

pval.Q.LRT
P-value of likelihood-ratio test.

tau2
Between-study variance τ^2.

se.tau2
Standard error of τ^2.

lower.tau2, upper.tau2
Lower and upper limit of confidence interval for τ^2.

tau
Square-root of between-study variance τ.

lower.tau, upper.tau
Lower and upper limit of confidence interval for τ.

H
Heterogeneity statistic H.

lower.H, upper.H
Lower and upper confidence limit for heterogeneity statistic H.

I2
Heterogeneity statistic I^2.

lower.I2, upper.I2
Lower and upper confidence limit for heterogeneity statistic I^2.

Rb
Heterogeneity statistic R_b.

lower.Rb, upper.Rb
Lower and upper confidence limit for heterogeneity statistic R_b.

method
A character string indicating method used for pooling: "Inverse"

df.hakn
Degrees of freedom for test of treatment effect for Hartung-Knapp method (only if hakn=TRUE).

bylevs
Levels of grouping variable - if byvar is not missing.

TE.fixed.w, seTE.fixed.w
Estimated treatment effect and standard error in subgroups (fixed effect model) - if byvar is not missing.

lower.fixed.w, upper.fixed.w
Lower and upper confidence interval limits in subgroups (fixed effect model) - if byvar is not missing.

statistic.fixed.w, pval.fixed.w
z-value and p-value for test of treatment effect in subgroups (fixed effect model) - if byvar is not missing.

TE.random.w, seTE.random.w
Estimated treatment effect and standard error in subgroups (random effects model) - if byvar is not missing.

lower.random.w, upper.random.w
Lower and upper confidence interval limits in subgroups (random effects model) - if byvar is not missing.

statistic.random.w, pval.random.w
z-value or t-value and corresponding p-value for test of treatment effect in subgroups (random effects model) - if byvar is not missing.

w.fixed.w, w.random.w
Weight of subgroups (in fixed and random effects model) - if byvar is not missing.

df.hakn.w
Degrees of freedom for test of treatment effect for Hartung-Knapp method in subgroups - if byvar is not missing and hakn=TRUE.

n.harmonic.mean.w
Harmonic mean of number of observations in subgroups (for back transformation of Freeman-Tukey Double arcsine transformation) - if byvar is not missing.

event.w
Number of events in subgroups - if byvar is not missing.

n.w
Number of observations in subgroups - if byvar is not missing.

k.w
Number of studies combined within subgroups - if byvar is not missing.

k.all.w
Number of all studies in subgroups - if byvar is not missing.

Q.w.fixed
Overall within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.w.random
Overall within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing (only calculated if argument tau.common is TRUE).

df.Q.w
Degrees of freedom for test of overall within subgroups heterogeneity - if byvar is not missing.

pval.Q.w.fixed
P-value of within subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.w.random
P-value of within subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

Q.b.fixed
Overall between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

Q.b.random
Overall between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

df.Q.b
Degrees of freedom for test of overall between subgroups heterogeneity - if byvar is not missing.

pval.Q.b.fixed
P-value of between subgroups heterogeneity statistic Q (based on fixed effect model) - if byvar is not missing.

pval.Q.b.random
P-value of between subgroups heterogeneity statistic Q (based on random effects model) - if byvar is not missing.

tau.w
Square-root of between-study variance within subgroups - if byvar is not missing.

H.w
Heterogeneity statistic H within subgroups - if byvar is not missing.

lower.H.w, upper.H.w
Lower and upper confidence limit for heterogeneity statistic H within subgroups - if byvar is not missing.

I2.w
Heterogeneity statistic I^2 within subgroups - if byvar is not missing.

lower.I2.w, upper.I2.w
Lower and upper confidence limit for heterogeneity statistic I^2 within subgroups - if byvar is not missing.

incr.event
Increment added to number of events.

keepdata
As defined above.

data
Original data (set) used in function call (if keepdata=TRUE).

subset
Information on subset of original data used in meta-analysis (if keepdata=TRUE).

.glmm.fixed
GLMM object generated by call of rma.glmm function (fixed effect model).

.glmm.random
GLMM object generated by call of rma.glmm function (random effects model).

call
Function call.

version
Version of R package meta used to create object.

version.metafor
Version of R package metafor used for GLMMs.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Agresti A & Coull BA (1998): Approximate is better than "exact" for interval estimation of binomial proportions. The American Statistician, 52, 119–26

Barendregt JJ, Doi SA, Lee YY, Norman RE, Vos T (2013): Meta-analysis of prevalence. Journal of Epidemiology and Community Health, 67, 974–8

Borenstein M, Hedges LV, Higgins JP, Rothstein HR (2010): A basic introduction to fixed-effect and random-effects models for meta-analysis. Research Synthesis Methods, 1, 97–111

DerSimonian R & Laird N (1986): Meta-analysis in clinical trials. Controlled Clinical Trials, 7, 177–88

Edward JM et al. (2006): Adherence to antiretroviral therapy in sub-saharan Africa and North America - a meta-analysis. Journal of the American Medical Association, 296, 679–90

Freeman MF & Tukey JW (1950): Transformations related to the angular and the square root. Annals of Mathematical Statistics, 21, 607–11

Higgins JPT, Thompson SG, Spiegelhalter DJ (2009): A re-evaluation of random-effects meta-analysis. Journal of the Royal Statistical Society: Series A, 172, 137–59

Hartung J, Knapp G (2001a): On tests of the overall treatment effect in meta-analysis with normally distributed responses. Statistics in Medicine, 20, 1771–82

Hartung J, Knapp G (2001b): A refined method for the meta-analysis of controlled clinical trials with binary outcome. Statistics in Medicine, 20, 3875–89

IntHout J, Ioannidis JPA, Borm GF (2014): The Hartung-Knapp-Sidik-Jonkman method for random effects meta-analysis is straightforward and considerably outperforms the standard DerSimonian-Laird method. BMC Medical Research Methodology, 14, 25

IQWiG (2020): General Methods: Version 6.0. https://www.iqwig.de/en/about-us/methods/methods-paper/

Jackson D, Law M, Rücker G, Schwarzer G (2017): The Hartung-Knapp modification for random-effects meta-analysis: A useful refinement but are there any residual concerns? Statistics in Medicine, 36, 3923–34

Langan D, Higgins JPT, Jackson D, Bowden J, Veroniki AA, Kontopantelis E, et al. (2019): A comparison of heterogeneity variance estimators in simulated random-effects meta-analyses. Research Synthesis Methods, 10, 83–98

Miller JJ (1978): The inverse of the Freeman-Tukey double arcsine transformation. The American Statistician, 32, 138

Newcombe RG (1998): Two-sided confidence intervals for the single proportion: comparison of seven methods. Statistics in Medicine, 17, 857–72

Pettigrew HM, Gart JJ, Thomas DG (1986): The bias and higher cumulants of the logarithm of a binomial variate. Biometrika, 73, 425–35

Schwarzer G, Chemaitelly H, Abu-Raddad LJ, Rücker G (2019): Seriously misleading results using inverse of Freeman-Tukey double arcsine transformation in meta-analysis of single proportions. Research Synthesis Methods, 10, 476–83

Stijnen T, Hamza TH, Ozdemir P (2010): Random effects meta-analysis of event outcome in the framework of the generalized linear mixed model with applications in sparse data. Statistics in Medicine, 29, 3046–67

Viechtbauer W (2010): Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36, 1–48

Warton DI, Hui FKC (2011): The arcsine is asinine: the analysis of proportions in ecology. Ecology, 92, 3–10

Wiksten A, Rücker G, Schwarzer G (2016): Hartung-Knapp method is not always conservative compared with fixed-effect meta-analysis. Statistics in Medicine, 35, 2503–15

See Also

update.meta, metacont, metagen, print.meta

Examples

# Meta-analysis using generalised linear mixed model
#
metaprop(4:1, 10 * 1:4)

# Apply various classic meta-analysis methods to estimate
# proportions
#
m1 <- metaprop(4:1, 10 * 1:4, method = "Inverse")
m2 <- update(m1, sm = "PAS")
m3 <- update(m1, sm = "PRAW")
m4 <- update(m1, sm = "PLN")
m5 <- update(m1, sm = "PFT")
#
m1
m2
m3
m4
m5
#
forest(m1)
## Not run: 
forest(m2)
forest(m3)
forest(m3, pscale = 100)
forest(m4)
forest(m5)

## End(Not run)

# Do not back transform results, e.g. print logit transformed
# proportions if sm = "PLOGIT" and store old settings
#
oldset <- settings.meta(backtransf = FALSE)
#
m6  <- metaprop(4:1, c(10, 20, 30, 40), method = "Inverse")
m7  <- update(m6, sm = "PAS")
m8  <- update(m6, sm = "PRAW")
m9  <- update(m6, sm = "PLN")
m10 <- update(m6, sm = "PFT")
#
forest(m6)
## Not run: 
forest(m7)
forest(m8)
forest(m8, pscale = 100)
forest(m9)
forest(m10)

## End(Not run)

# Use old settings
#
settings.meta(oldset)

# Examples with zero events
#
m1 <- metaprop(c(0, 0, 10, 10), rep(100, 4), method = "Inverse")
m2 <- metaprop(c(0, 0, 10, 10), rep(100, 4), incr = 0.1, method = "Inverse")
#
summary(m1)
summary(m2)
#
## Not run: 
forest(m1)
forest(m2)

## End(Not run)

# Example from Miller (1978):
#
death <- c(3, 6, 10, 1)
animals <- c(11, 17, 21, 6)
#
m3 <- metaprop(death, animals, sm = "PFT")
forest(m3)

# Data examples from Newcombe (1998)
# - apply various methods to estimate confidence intervals for
#   individual studies
#
event <- c(81, 15, 0, 1)
n <- c(263, 148, 20, 29)
#
m1 <- metaprop(event, n, method.ci = "SA", method = "Inverse")
m2 <- update(m1, method.ci = "SACC")
m3 <- update(m1, method.ci = "WS")
m4 <- update(m1, method.ci = "WSCC")
m5 <- update(m1, method.ci = "CP")
#
lower <- round(rbind(NA, m1$lower, m2$lower, NA, m3$lower,
                     m4$lower, NA, m5$lower), 4)
upper <- round(rbind(NA, m1$upper, m2$upper, NA, m3$upper,
                     m4$upper, NA, m5$upper), 4)
#
tab1 <- data.frame(
  scen1 = meta:::formatCI(lower[, 1], upper[, 1]),
  scen2 = meta:::formatCI(lower[, 2], upper[, 2]),
  scen3 = meta:::formatCI(lower[, 3], upper[, 3]),
  scen4 = meta:::formatCI(lower[, 4], upper[, 4]),
  stringsAsFactors = FALSE
  )
names(tab1) <- c("r=81, n=263", "r=15, n=148",
                 "r=0, n=20", "r=1, n=29")
row.names(tab1) <- c("Simple", "- SA", "- SACC",
                     "Score", "- WS", "- WSCC",
                     "Binomial", "- CP")
tab1[is.na(tab1)] <- ""
# Newcombe (1998), Table I, methods 1-5:
tab1

# Same confidence interval, i.e. unaffected by choice of summary
# measure
#
print(metaprop(event, n, method.ci = "WS", method = "Inverse"), ma = FALSE)
print(metaprop(event, n, sm = "PLN", method.ci = "WS"), ma = FALSE)
print(metaprop(event, n, sm = "PFT", method.ci = "WS"), ma = FALSE)
print(metaprop(event, n, sm = "PAS", method.ci = "WS"), ma = FALSE)
print(metaprop(event, n, sm = "PRAW", method.ci = "WS"), ma = FALSE)

# Different confidence intervals as argument sm = "NAsm"
#
print(metaprop(event, n, method.ci = "NAsm", method = "Inverse"), ma = FALSE)
print(metaprop(event, n, sm = "PLN", method.ci = "NAsm"), ma = FALSE)
print(metaprop(event, n, sm = "PFT", method.ci = "NAsm"), ma = FALSE)
print(metaprop(event, n, sm = "PAS", method.ci = "NAsm"), ma = FALSE)
print(metaprop(event, n, sm = "PRAW", method.ci = "NAsm"), ma = FALSE)

# Different confidence intervals as argument backtransf = FALSE.
# Accordingly, method.ci = "NAsm" used internally.
#
print(metaprop(event, n, method.ci = "WS", method = "Inverse"),
      ma = FALSE, backtransf = FALSE)
print(metaprop(event, n, sm = "PLN", method.ci = "WS"),
      ma = FALSE, backtransf = FALSE)
print(metaprop(event, n, sm = "PFT", method.ci = "WS"),
      ma = FALSE, backtransf = FALSE)
print(metaprop(event, n, sm = "PAS", method.ci = "WS"),
      ma = FALSE, backtransf = FALSE)
print(metaprop(event, n, sm = "PRAW", method.ci = "WS"),
      ma = FALSE, backtransf = FALSE)

# Same results (printed on original and log scale, respectively)
#
print(metaprop(event, n, sm = "PLN", method.ci = "NAsm"), ma = FALSE)
print(metaprop(event, n, sm = "PLN"), ma = FALSE, backtransf = FALSE)
# Results for first study (on log scale)
round(log(c(0.3079848, 0.2569522, 0.3691529)), 4)

# Print results as events per 1000 observations
#
print(metaprop(6:8, c(100, 1200, 1000), method = "Inverse"),
      pscale = 1000, digits = 1)

[Package meta version 4.18-1 Index]