print.meta {meta}	R Documentation

Print and summary method for objects of class meta

Description

Print and summary method for objects of class meta.

Usage

## S3 method for class 'meta':
print(x, sortvar, level=x$level, level.comb=x$level.comb,
        comb.fixed=x$comb.fixed, comb.random=x$comb.random,
        details=FALSE, ma=TRUE, digits=max(4, .Options$digits - 3), ...)

## S3 method for class 'metabias':
print(x, ...)

## S3 method for class 'meta':
summary(object, byvar=object$byvar,
        bylab=object$bylab, print.byvar=object$print.byvar,
        bystud=FALSE,
        level=object$level, level.comb=object$level.comb,
        comb.fixed=object$comb.fixed, comb.random=object$comb.random,
        warn=TRUE, ...)

## S3 method for class 'summary.meta':
print(x, digits = max(3, .Options$digits - 3),
        print.byvar,
        comb.fixed=x$comb.fixed, comb.random=x$comb.random,
        header=TRUE, ...)

Arguments

x 	 An object of class meta, metabias, or summary.meta.

object 	 An object of class meta.

sortvar 	 An optional vector used to sort the individual studies (must be of same length as x$TE).

level 	 The level used to calculate confidence intervals for individual studies.

level.comb 	 The level used to calculate confidence intervals for pooled estimates.

comb.fixed 	 A logical indicating whether a fixed effect meta-analysis should be conducted.

comb.random 	 A logical indicating whether a random effects meta-analysis should be conducted.

header 	 A logical indicating whether information on title of meta-analysis, comparison and outcome should be printed at the beginning of the printout.

details 	 A logical indicating whether further details of individual studies should be printed.

ma 	 A logical indicating whether the summary results of the meta-analysis should be printed.

byvar 	 An optional vector containing grouping information (must be of same length as x$TE).

bylab 	 A character string with a label for the grouping variable.

bystud 	 A logical indicating whether results of individual studies should be printed by grouping variable.

digits 	 Minimal number of significant digits, see print.default.

print.byvar 	 A logical indicating whether the name of the grouping variable should be printed in front of the group labels. By default, the value of print.byvar is set to TRUE.

warn 	 A logical indicating whether the use of summary.meta in connection with metacum or metainf should result in a warning.

... 	 other arguments

Details

Review Manager 5 (RevMan 5) is the current software used for preparing and maintaining Cochrane Reviews (http://www.cc-ims.net/revman/). In RevMan 5, subgroup analyses can be defined and data from a Cochrane review can be imported to R using the function read.rm5. If a meta-analysis is then conducted using function metacr, information on subgroups is available in R (components byvar, bylab, and print.byvar, byvar in an object of class "meta"). Accordingly, by using function metacr there is no need to define subgroups in order to redo the statistical analysis conducted in the Cochrane review.

For subgroups (argument byvar not NULL), results for the fixed effect model will be printed if both arguments comb.fixed and comb.random are TRUE. In order to get results for the random effects model within subgroups, use comb.fixed==FALSE and comb.random==TRUE.

Value

A list is returned by the function summary.meta with the following elements:

study 	 Results for individual studies (a list with elements TE, seTE, lower, upper, z, p, level).

fixed 	 Results for fixed effect model (a list with elements TE, seTE, lower, upper, z, p, level).

random 	 Results for random effects model (a list with elements TE, seTE, lower, upper, z, p, level).

k 	 Number of studies combined in meta-analysis.

Q 	 Heterogeneity statistic Q.

tau 	 Square-root of between-study variance (moment estimator of DerSimonian-Laird).

H 	 Heterogeneity statistic H (a list with elements TE, lower, upper).

I2 	 Heterogeneity statistic I2 (a list with elements TE, lower, upper), see Higgins & Thompson (2002).

k.all 	 Total number of trials.

Q.CMH 	 Cochrane-Mantel-Haenszel heterogeneity statistic.

sm 	 A character string indicating underlying summary measure.

method 	 A character string with the pooling method.

call 	 Function call.

ci.lab 	 Label for confidence interval.

within.fixed 	 Result for fixed effect model within groups (a list with elements TE, seTE, lower, upper, z, p, level) - if byvar is not missing.

within.random 	 Result for random effects model within groups (a list with elements TE, seTE, lower, upper, z, p, level) - if byvar is not missing.

k.w 	 Number of studies combined within groups - if byvar is not missing.

Q.w 	 Heterogeneity statistic Q within groups - if byvar is not missing.

bylab 	 Label for grouping variable - if byvar is not missing.

by.levs 	 Levels of grouping variable - if byvar is not missing.

comb.fixed, comb.random 	 As defined above.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Cooper H & Hedges LV (1994), The Handbook of Research Synthesis. Newbury Park, CA: Russell Sage Foundation.

Higgins JPT & Thompson SG (2002), Quantifying heterogeneity in a meta-analysis. Statistics in Medicine, 21, 1539–1558.

See Also

metabin, metacont, metagen

Examples

data(Fleiss93cont)
meta1 <- metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, data=Fleiss93cont, sm="SMD")
summary(meta1)
summary(meta1, byvar=c(1,2,1,1,2), bylab="group")

[Package meta version 1.1-8 Index]