trimfill {meta}	R Documentation

Generic function for trim-and-fill method

Description

Trim and fill method for estimating and adjusting for the number and outcomes of missing studies in a meta-analysis.

Usage

trimfill(x, seTE, ...)

Arguments

x 	 An object of class meta, or estimated treatment effect in individual studies.

seTE 	 Standard error of estimated treatment effect (mandatory if x not of class meta).

... 	 Additional arguments.

Details

The trim and fill method can be used for estimating and adjusting for the number and outcomes of missing studies in a meta-analysis. The method relies on scrutiny of one side of a funnel plot for asymmetry assumed due to publication bias.

Value

An object of class c("metagen", "meta", "trimfill"). The object is a list containing the following components:

studlab, sm, left, ma.fixed, type

n.iter.max, level, level.comb, 	 As defined above.

comb.fixed, comb.random

TE, seTE 	 Estimated treatment effect and standard error of individual studies.

w.fixed, w.random 	 Weight of individual studies (in fixed and random effects model).

TE.fixed, seTE.fixed 	 Estimated overall treatment effect and standard error (fixed effect model).

TE.random, seTE.random 	 Estimated overall treatment effect and standard error (random effects model).

k 	 Number of studies combined in meta-analysis.

Q 	 Heterogeneity statistic Q.

tau 	 Square-root of between-study variance (moment estimator of DerSimonian-Laird).

method 	 Pooling method: "Inverse".

call 	 Function call.

n.iter 	 Actual number of iterations to estimate number of missing studies.

trimfill 	 A logical vector indicating studies that have been added by trim and fill method.

k0 	 Number of studies added by trim and fill.

Author(s)

Guido Schwarzer sc@imbi.uni-freiburg.de

References

Duval S & Tweedie R (2000), A nonparametric "Trim and Fill" method of accounting for publication bias in meta-analysis. Journal of the American Statistical Association, 95, 89–98.

Duval S & Tweedie R (2000), Trim and Fill: A simple funnel-plot-based method of testing and adjusting for publication bias in meta-analysis. Biometrics, 56, 455–463.

See Also

metagen, metabias, trimfill.meta, funnel

Examples

data(Fleiss93)
meta1 <- metabin(event.e, n.e, event.c, n.c,
                 data=Fleiss93, sm="OR")
tf1 <- trimfill(meta1)
summary(tf1)
funnel(tf1, pch=ifelse(tf1$trimfill, 1, 16),
       level=0.95, comb.fixed=TRUE)

trimfill(meta1$TE, meta1$seTE, sm=meta1$sm)

[Package meta version 1.1-8 Index]