woodyplants {meta}	R Documentation

Elevated CO_2 and total biomass of woody plants

Description

Meta-analysis on effects of elevated CO_2 on total biomass of woody plants

This dataset has been used as an example in Hedges et al. (1999) to describe methods for the meta-analysis of response ratios. The complete dataset with 102 observations and 26 variables is available online as a supplement. Here only a subset of 10 variables is provided and used in the examples.

Format

A data frame with the following columns:

obsno	 observation number

papno	 database paper number

treat	 treatment code

level	 treatment level

n.elev	 number of observations in experimental group (elevated CO_2-level)

mean.elev	 estimated mean in experimental group

sd.elev	 standard deviation in experimental group

n.amb	 number of observations in control group (ambient CO_2-level)

mean.amb	 estimated mean in control group

sd.amb	 standard deviation in control group

Source

Website http://www.esapubs.org/archive/ecol/E080/008/

References

Hedges LV, Gurevitch J, Curtis PS (1999): The meta-analysis of response ratios in experimental ecology. Ecology, 80, 1150–6

Examples

data(woodyplants)

# Meta-analysis of response ratios (Hedges et al., 1999)
#
m1 <- metacont(n.elev, mean.elev, sd.elev,
               n.amb, mean.amb, sd.amb,
               data = woodyplants, sm = "ROM",
               studlab = paste(obsno, papno, sep = " / "))
summary(m1, prediction = TRUE)

# Meta-analysis for plants grown with low soil fertility treatment
#
m2 <- update(m1, subset = (treat == "fert" & level == "low"))
summary(m2, prediction = TRUE)

# Meta-analysis for plants grown under low light conditions
#
m3 <- update(m1, subset = (treat == "light" & level == "low"))
summary(m3, prediction = TRUE)

[Package meta version 4.18-1 Index]